@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: info/auth.proto
 
 package info
@@ -742,7 +742,7 @@ func file_info_auth_proto_rawDescGZIP() []byte {
 }
 
 var file_info_auth_proto_msgTypes = make([]protoimpl.MessageInfo, 11)
-var file_info_auth_proto_goTypes = []interface{}{
+var file_info_auth_proto_goTypes = []any{
 	(*Account)(nil),                          // 0: smartcore.info.Account
 	(*Token)(nil),                            // 1: smartcore.info.Token
 	(*Permission)(nil),                       // 2: smartcore.info.Permission
@@ -787,7 +787,7 @@ func file_info_auth_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_info_auth_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Account); i {
 			case 0:
 				return &v.state
@@ -799,7 +799,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Token); i {
 			case 0:
 				return &v.state
@@ -811,7 +811,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*Permission); i {
 			case 0:
 				return &v.state
@@ -823,7 +823,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*AddAccountRequest); i {
 			case 0:
 				return &v.state
@@ -835,7 +835,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*AddAccountResponse); i {
 			case 0:
 				return &v.state
@@ -847,7 +847,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*RemoveAccountRequest); i {
 			case 0:
 				return &v.state
@@ -859,7 +859,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*RemoveAccountResponse); i {
 			case 0:
 				return &v.state
@@ -871,7 +871,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateAccountPermissionsRequest); i {
 			case 0:
 				return &v.state
@@ -883,7 +883,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateAccountPermissionsResponse); i {
 			case 0:
 				return &v.state
@@ -895,7 +895,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*GenerateTokenRequest); i {
 			case 0:
 				return &v.state
@@ -907,7 +907,7 @@ func file_info_auth_proto_init() {
 				return nil
 			}
 		}
-		file_info_auth_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_info_auth_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*GenerateTokenResponse); i {
 			case 0:
 				return &v.state
