@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: info/health.proto
 
 package info
@@ -520,7 +520,7 @@ func file_info_health_proto_rawDescGZIP() []byte {
 }
 
 var file_info_health_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_info_health_proto_goTypes = []interface{}{
+var file_info_health_proto_goTypes = []any{
 	(*HealthState)(nil),              // 0: smartcore.info.HealthState
 	(*ConnectionHealth)(nil),         // 1: smartcore.info.ConnectionHealth
 	(*CommHealth)(nil),               // 2: smartcore.info.CommHealth
@@ -561,7 +561,7 @@ func file_info_health_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_info_health_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_info_health_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*HealthState); i {
 			case 0:
 				return &v.state
@@ -573,7 +573,7 @@ func file_info_health_proto_init() {
 				return nil
 			}
 		}
-		file_info_health_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_info_health_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ConnectionHealth); i {
 			case 0:
 				return &v.state
@@ -585,7 +585,7 @@ func file_info_health_proto_init() {
 				return nil
 			}
 		}
-		file_info_health_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_info_health_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*CommHealth); i {
 			case 0:
 				return &v.state
@@ -597,7 +597,7 @@ func file_info_health_proto_init() {
 				return nil
 			}
 		}
-		file_info_health_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_info_health_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetHealthStateRequest); i {
 			case 0:
 				return &v.state
@@ -609,7 +609,7 @@ func file_info_health_proto_init() {
 				return nil
 			}
 		}
-		file_info_health_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_info_health_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHealthStatesRequest); i {
 			case 0:
 				return &v.state
@@ -621,7 +621,7 @@ func file_info_health_proto_init() {
 				return nil
 			}
 		}
-		file_info_health_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_info_health_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHealthStatesResponse); i {
 			case 0:
 				return &v.state
@@ -633,7 +633,7 @@ func file_info_health_proto_init() {
 				return nil
 			}
 		}
-		file_info_health_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_info_health_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*HealthStateChange); i {
 			case 0:
 				return &v.state
