@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: info/info.proto
 
 package info
@@ -674,7 +674,7 @@ func file_info_info_proto_rawDescGZIP() []byte {
 }
 
 var file_info_info_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
-var file_info_info_proto_goTypes = []interface{}{
+var file_info_info_proto_goTypes = []any{
 	(*ListDevicesRequest)(nil),         // 0: smartcore.info.ListDevicesRequest
 	(*ListDevicesResponse)(nil),        // 1: smartcore.info.ListDevicesResponse
 	(*PullDevicesRequest)(nil),         // 2: smartcore.info.PullDevicesRequest
@@ -713,7 +713,7 @@ func file_info_info_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_info_info_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*ListDevicesRequest); i {
 			case 0:
 				return &v.state
@@ -725,7 +725,7 @@ func file_info_info_proto_init() {
 				return nil
 			}
 		}
-		file_info_info_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ListDevicesResponse); i {
 			case 0:
 				return &v.state
@@ -737,7 +737,7 @@ func file_info_info_proto_init() {
 				return nil
 			}
 		}
-		file_info_info_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PullDevicesRequest); i {
 			case 0:
 				return &v.state
@@ -749,7 +749,7 @@ func file_info_info_proto_init() {
 				return nil
 			}
 		}
-		file_info_info_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullDevicesResponse); i {
 			case 0:
 				return &v.state
@@ -761,7 +761,7 @@ func file_info_info_proto_init() {
 				return nil
 			}
 		}
-		file_info_info_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*Device); i {
 			case 0:
 				return &v.state
@@ -773,7 +773,7 @@ func file_info_info_proto_init() {
 				return nil
 			}
 		}
-		file_info_info_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*Trait); i {
 			case 0:
 				return &v.state
@@ -785,7 +785,7 @@ func file_info_info_proto_init() {
 				return nil
 			}
 		}
-		file_info_info_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*GrpcClientOptions); i {
 			case 0:
 				return &v.state
@@ -797,7 +797,7 @@ func file_info_info_proto_init() {
 				return nil
 			}
 		}
-		file_info_info_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_info_info_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullDevicesResponse_Change); i {
 			case 0:
 				return &v.state
