@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/air_quality_sensor.proto
 
 package traits
@@ -841,7 +841,7 @@ func file_traits_air_quality_sensor_proto_rawDescGZIP() []byte {
 
 var file_traits_air_quality_sensor_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_air_quality_sensor_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_traits_air_quality_sensor_proto_goTypes = []interface{}{
+var file_traits_air_quality_sensor_proto_goTypes = []any{
 	(AirQuality_Comfort)(0),               // 0: smartcore.traits.AirQuality.Comfort
 	(*AirQuality)(nil),                    // 1: smartcore.traits.AirQuality
 	(*AirQualitySupport)(nil),             // 2: smartcore.traits.AirQualitySupport
@@ -893,7 +893,7 @@ func file_traits_air_quality_sensor_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_air_quality_sensor_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_quality_sensor_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*AirQuality); i {
 			case 0:
 				return &v.state
@@ -905,7 +905,7 @@ func file_traits_air_quality_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_quality_sensor_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_quality_sensor_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*AirQualitySupport); i {
 			case 0:
 				return &v.state
@@ -917,7 +917,7 @@ func file_traits_air_quality_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_quality_sensor_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_quality_sensor_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetAirQualityRequest); i {
 			case 0:
 				return &v.state
@@ -929,7 +929,7 @@ func file_traits_air_quality_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_quality_sensor_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_quality_sensor_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAirQualityRequest); i {
 			case 0:
 				return &v.state
@@ -941,7 +941,7 @@ func file_traits_air_quality_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_quality_sensor_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_quality_sensor_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAirQualityResponse); i {
 			case 0:
 				return &v.state
@@ -953,7 +953,7 @@ func file_traits_air_quality_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_quality_sensor_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_quality_sensor_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeAirQualityRequest); i {
 			case 0:
 				return &v.state
@@ -965,7 +965,7 @@ func file_traits_air_quality_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_quality_sensor_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_quality_sensor_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAirQualityResponse_Change); i {
 			case 0:
 				return &v.state
@@ -978,7 +978,7 @@ func file_traits_air_quality_sensor_proto_init() {
 			}
 		}
 	}
-	file_traits_air_quality_sensor_proto_msgTypes[0].OneofWrappers = []interface{}{}
+	file_traits_air_quality_sensor_proto_msgTypes[0].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
