@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/air_temperature.proto
 
 package traits
@@ -131,7 +131,7 @@ type AirTemperature struct {
 	TemperatureGoal isAirTemperature_TemperatureGoal `protobuf_oneof:"temperature_goal"`
 	// Optional, read-only. The ambient temperature as read by the device
 	AmbientTemperature *types.Temperature `protobuf:"bytes,5,opt,name=ambient_temperature,json=ambientTemperature,proto3" json:"ambient_temperature,omitempty"`
-	// Optional, read-only. The ambient humidity as read by the device
+	// Optional, read-only. The ambient relative humidity percentage, as read by the device
 	AmbientHumidity *float32 `protobuf:"fixed32,6,opt,name=ambient_humidity,json=ambientHumidity,proto3,oneof" json:"ambient_humidity,omitempty"`
 	// Optional, read-only. The dew-point as read by the device
 	DewPoint *types.Temperature `protobuf:"bytes,7,opt,name=dew_point,json=dewPoint,proto3" json:"dew_point,omitempty"`
@@ -933,7 +933,7 @@ func file_traits_air_temperature_proto_rawDescGZIP() []byte {
 
 var file_traits_air_temperature_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_air_temperature_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
-var file_traits_air_temperature_proto_goTypes = []interface{}{
+var file_traits_air_temperature_proto_goTypes = []any{
 	(AirTemperature_Mode)(0),                  // 0: smartcore.traits.AirTemperature.Mode
 	(*AirTemperature)(nil),                    // 1: smartcore.traits.AirTemperature
 	(*TemperatureRange)(nil),                  // 2: smartcore.traits.TemperatureRange
@@ -991,7 +991,7 @@ func file_traits_air_temperature_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_air_temperature_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*AirTemperature); i {
 			case 0:
 				return &v.state
@@ -1003,7 +1003,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*TemperatureRange); i {
 			case 0:
 				return &v.state
@@ -1015,7 +1015,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*AirTemperatureSupport); i {
 			case 0:
 				return &v.state
@@ -1027,7 +1027,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetAirTemperatureRequest); i {
 			case 0:
 				return &v.state
@@ -1039,7 +1039,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateAirTemperatureRequest); i {
 			case 0:
 				return &v.state
@@ -1051,7 +1051,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAirTemperatureRequest); i {
 			case 0:
 				return &v.state
@@ -1063,7 +1063,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAirTemperatureResponse); i {
 			case 0:
 				return &v.state
@@ -1075,7 +1075,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeAirTemperatureRequest); i {
 			case 0:
 				return &v.state
@@ -1087,7 +1087,7 @@ func file_traits_air_temperature_proto_init() {
 				return nil
 			}
 		}
-		file_traits_air_temperature_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_air_temperature_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAirTemperatureResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1100,7 +1100,7 @@ func file_traits_air_temperature_proto_init() {
 			}
 		}
 	}
-	file_traits_air_temperature_proto_msgTypes[0].OneofWrappers = []interface{}{
+	file_traits_air_temperature_proto_msgTypes[0].OneofWrappers = []any{
 		(*AirTemperature_TemperatureSetPoint)(nil),
 		(*AirTemperature_TemperatureSetPointDelta)(nil),
 		(*AirTemperature_TemperatureRange)(nil),
