@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/booking.proto
 
 package traits
@@ -1209,7 +1209,7 @@ func file_traits_booking_proto_rawDescGZIP() []byte {
 
 var file_traits_booking_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_booking_proto_msgTypes = make([]protoimpl.MessageInfo, 15)
-var file_traits_booking_proto_goTypes = []interface{}{
+var file_traits_booking_proto_goTypes = []any{
 	(BookingSupport_CheckInSupport)(0),  // 0: smartcore.traits.BookingSupport.CheckInSupport
 	(*Booking)(nil),                     // 1: smartcore.traits.Booking
 	(*BookingSupport)(nil),              // 2: smartcore.traits.BookingSupport
@@ -1281,7 +1281,7 @@ func file_traits_booking_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_booking_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Booking); i {
 			case 0:
 				return &v.state
@@ -1293,7 +1293,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*BookingSupport); i {
 			case 0:
 				return &v.state
@@ -1305,7 +1305,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ListBookingsRequest); i {
 			case 0:
 				return &v.state
@@ -1317,7 +1317,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ListBookingsResponse); i {
 			case 0:
 				return &v.state
@@ -1329,7 +1329,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*CheckInBookingRequest); i {
 			case 0:
 				return &v.state
@@ -1341,7 +1341,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*CheckInBookingResponse); i {
 			case 0:
 				return &v.state
@@ -1353,7 +1353,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*CheckOutBookingRequest); i {
 			case 0:
 				return &v.state
@@ -1365,7 +1365,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*CheckOutBookingResponse); i {
 			case 0:
 				return &v.state
@@ -1377,7 +1377,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateBookingRequest); i {
 			case 0:
 				return &v.state
@@ -1389,7 +1389,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateBookingResponse); i {
 			case 0:
 				return &v.state
@@ -1401,7 +1401,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateBookingRequest); i {
 			case 0:
 				return &v.state
@@ -1413,7 +1413,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateBookingResponse); i {
 			case 0:
 				return &v.state
@@ -1425,7 +1425,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*PullBookingsResponse); i {
 			case 0:
 				return &v.state
@@ -1437,7 +1437,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeBookingRequest); i {
 			case 0:
 				return &v.state
@@ -1449,7 +1449,7 @@ func file_traits_booking_proto_init() {
 				return nil
 			}
 		}
-		file_traits_booking_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_booking_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*PullBookingsResponse_Change); i {
 			case 0:
 				return &v.state
