@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/brightness_sensor.proto
 
 package traits
@@ -528,7 +528,7 @@ func file_traits_brightness_sensor_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_brightness_sensor_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_traits_brightness_sensor_proto_goTypes = []interface{}{
+var file_traits_brightness_sensor_proto_goTypes = []any{
 	(*AmbientBrightness)(nil),                    // 0: smartcore.traits.AmbientBrightness
 	(*AmbientBrightnessSupport)(nil),             // 1: smartcore.traits.AmbientBrightnessSupport
 	(*GetAmbientBrightnessRequest)(nil),          // 2: smartcore.traits.GetAmbientBrightnessRequest
@@ -568,7 +568,7 @@ func file_traits_brightness_sensor_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_brightness_sensor_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_brightness_sensor_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*AmbientBrightness); i {
 			case 0:
 				return &v.state
@@ -580,7 +580,7 @@ func file_traits_brightness_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_brightness_sensor_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_brightness_sensor_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*AmbientBrightnessSupport); i {
 			case 0:
 				return &v.state
@@ -592,7 +592,7 @@ func file_traits_brightness_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_brightness_sensor_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_brightness_sensor_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetAmbientBrightnessRequest); i {
 			case 0:
 				return &v.state
@@ -604,7 +604,7 @@ func file_traits_brightness_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_brightness_sensor_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_brightness_sensor_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAmbientBrightnessRequest); i {
 			case 0:
 				return &v.state
@@ -616,7 +616,7 @@ func file_traits_brightness_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_brightness_sensor_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_brightness_sensor_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAmbientBrightnessResponse); i {
 			case 0:
 				return &v.state
@@ -628,7 +628,7 @@ func file_traits_brightness_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_brightness_sensor_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_brightness_sensor_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeAmbientBrightnessRequest); i {
 			case 0:
 				return &v.state
@@ -640,7 +640,7 @@ func file_traits_brightness_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_brightness_sensor_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_brightness_sensor_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullAmbientBrightnessResponse_Change); i {
 			case 0:
 				return &v.state
