@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/channel.proto
 
 package traits
@@ -735,7 +735,7 @@ func file_traits_channel_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_channel_proto_msgTypes = make([]protoimpl.MessageInfo, 10)
-var file_traits_channel_proto_goTypes = []interface{}{
+var file_traits_channel_proto_goTypes = []any{
 	(*Channel)(nil),                          // 0: smartcore.traits.Channel
 	(*ChosenChannelSupport)(nil),             // 1: smartcore.traits.ChosenChannelSupport
 	(*GetChosenChannelRequest)(nil),          // 2: smartcore.traits.GetChosenChannelRequest
@@ -783,7 +783,7 @@ func file_traits_channel_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_channel_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Channel); i {
 			case 0:
 				return &v.state
@@ -795,7 +795,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ChosenChannelSupport); i {
 			case 0:
 				return &v.state
@@ -807,7 +807,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetChosenChannelRequest); i {
 			case 0:
 				return &v.state
@@ -819,7 +819,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ChooseChannelRequest); i {
 			case 0:
 				return &v.state
@@ -831,7 +831,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*AdjustChannelRequest); i {
 			case 0:
 				return &v.state
@@ -843,7 +843,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ReturnChannelRequest); i {
 			case 0:
 				return &v.state
@@ -855,7 +855,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullChosenChannelRequest); i {
 			case 0:
 				return &v.state
@@ -867,7 +867,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullChosenChannelResponse); i {
 			case 0:
 				return &v.state
@@ -879,7 +879,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeChosenChannelRequest); i {
 			case 0:
 				return &v.state
@@ -891,7 +891,7 @@ func file_traits_channel_proto_init() {
 				return nil
 			}
 		}
-		file_traits_channel_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_channel_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*PullChosenChannelResponse_Change); i {
 			case 0:
 				return &v.state
