@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/count.proto
 
 package traits
@@ -706,7 +706,7 @@ func file_traits_count_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_count_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
-var file_traits_count_proto_goTypes = []interface{}{
+var file_traits_count_proto_goTypes = []any{
 	(*Count)(nil),                     // 0: smartcore.traits.Count
 	(*CountSupport)(nil),              // 1: smartcore.traits.CountSupport
 	(*GetCountRequest)(nil),           // 2: smartcore.traits.GetCountRequest
@@ -754,7 +754,7 @@ func file_traits_count_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_count_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Count); i {
 			case 0:
 				return &v.state
@@ -766,7 +766,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CountSupport); i {
 			case 0:
 				return &v.state
@@ -778,7 +778,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetCountRequest); i {
 			case 0:
 				return &v.state
@@ -790,7 +790,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ResetCountRequest); i {
 			case 0:
 				return &v.state
@@ -802,7 +802,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateCountRequest); i {
 			case 0:
 				return &v.state
@@ -814,7 +814,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullCountsRequest); i {
 			case 0:
 				return &v.state
@@ -826,7 +826,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullCountsResponse); i {
 			case 0:
 				return &v.state
@@ -838,7 +838,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeCountRequest); i {
 			case 0:
 				return &v.state
@@ -850,7 +850,7 @@ func file_traits_count_proto_init() {
 				return nil
 			}
 		}
-		file_traits_count_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_count_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullCountsResponse_Change); i {
 			case 0:
 				return &v.state
