@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/electric.proto
 
 package traits
@@ -1536,7 +1536,7 @@ func file_traits_electric_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_electric_proto_msgTypes = make([]protoimpl.MessageInfo, 18)
-var file_traits_electric_proto_goTypes = []interface{}{
+var file_traits_electric_proto_goTypes = []any{
 	(*ElectricDemand)(nil),                // 0: smartcore.traits.ElectricDemand
 	(*ElectricMode)(nil),                  // 1: smartcore.traits.ElectricMode
 	(*GetDemandRequest)(nil),              // 2: smartcore.traits.GetDemandRequest
@@ -1613,7 +1613,7 @@ func file_traits_electric_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_electric_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*ElectricDemand); i {
 			case 0:
 				return &v.state
@@ -1625,7 +1625,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ElectricMode); i {
 			case 0:
 				return &v.state
@@ -1637,7 +1637,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetDemandRequest); i {
 			case 0:
 				return &v.state
@@ -1649,7 +1649,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullDemandRequest); i {
 			case 0:
 				return &v.state
@@ -1661,7 +1661,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullDemandResponse); i {
 			case 0:
 				return &v.state
@@ -1673,7 +1673,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*GetActiveModeRequest); i {
 			case 0:
 				return &v.state
@@ -1685,7 +1685,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateActiveModeRequest); i {
 			case 0:
 				return &v.state
@@ -1697,7 +1697,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*ClearActiveModeRequest); i {
 			case 0:
 				return &v.state
@@ -1709,7 +1709,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullActiveModeRequest); i {
 			case 0:
 				return &v.state
@@ -1721,7 +1721,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*PullActiveModeResponse); i {
 			case 0:
 				return &v.state
@@ -1733,7 +1733,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*ListModesRequest); i {
 			case 0:
 				return &v.state
@@ -1745,7 +1745,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*ListModesResponse); i {
 			case 0:
 				return &v.state
@@ -1757,7 +1757,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*PullModesRequest); i {
 			case 0:
 				return &v.state
@@ -1769,7 +1769,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*PullModesResponse); i {
 			case 0:
 				return &v.state
@@ -1781,7 +1781,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*ElectricMode_Segment); i {
 			case 0:
 				return &v.state
@@ -1793,7 +1793,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*PullDemandResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1805,7 +1805,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[16].Exporter = func(v any, i int) any {
 			switch v := v.(*PullActiveModeResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1817,7 +1817,7 @@ func file_traits_electric_proto_init() {
 				return nil
 			}
 		}
-		file_traits_electric_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_electric_proto_msgTypes[17].Exporter = func(v any, i int) any {
 			switch v := v.(*PullModesResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1830,8 +1830,8 @@ func file_traits_electric_proto_init() {
 			}
 		}
 	}
-	file_traits_electric_proto_msgTypes[0].OneofWrappers = []interface{}{}
-	file_traits_electric_proto_msgTypes[14].OneofWrappers = []interface{}{
+	file_traits_electric_proto_msgTypes[0].OneofWrappers = []any{}
+	file_traits_electric_proto_msgTypes[14].OneofWrappers = []any{
 		(*ElectricMode_Segment_Fixed)(nil),
 	}
 	type x struct{}
