@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/emergency.proto
 
 package traits
@@ -691,7 +691,7 @@ func file_traits_emergency_proto_rawDescGZIP() []byte {
 
 var file_traits_emergency_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_emergency_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_traits_emergency_proto_goTypes = []interface{}{
+var file_traits_emergency_proto_goTypes = []any{
 	(Emergency_Level)(0),                 // 0: smartcore.traits.Emergency.Level
 	(*Emergency)(nil),                    // 1: smartcore.traits.Emergency
 	(*EmergencySupport)(nil),             // 2: smartcore.traits.EmergencySupport
@@ -737,7 +737,7 @@ func file_traits_emergency_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_emergency_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Emergency); i {
 			case 0:
 				return &v.state
@@ -749,7 +749,7 @@ func file_traits_emergency_proto_init() {
 				return nil
 			}
 		}
-		file_traits_emergency_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*EmergencySupport); i {
 			case 0:
 				return &v.state
@@ -761,7 +761,7 @@ func file_traits_emergency_proto_init() {
 				return nil
 			}
 		}
-		file_traits_emergency_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetEmergencyRequest); i {
 			case 0:
 				return &v.state
@@ -773,7 +773,7 @@ func file_traits_emergency_proto_init() {
 				return nil
 			}
 		}
-		file_traits_emergency_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateEmergencyRequest); i {
 			case 0:
 				return &v.state
@@ -785,7 +785,7 @@ func file_traits_emergency_proto_init() {
 				return nil
 			}
 		}
-		file_traits_emergency_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEmergencyRequest); i {
 			case 0:
 				return &v.state
@@ -797,7 +797,7 @@ func file_traits_emergency_proto_init() {
 				return nil
 			}
 		}
-		file_traits_emergency_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEmergencyResponse); i {
 			case 0:
 				return &v.state
@@ -809,7 +809,7 @@ func file_traits_emergency_proto_init() {
 				return nil
 			}
 		}
-		file_traits_emergency_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeEmergencyRequest); i {
 			case 0:
 				return &v.state
@@ -821,7 +821,7 @@ func file_traits_emergency_proto_init() {
 				return nil
 			}
 		}
-		file_traits_emergency_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_emergency_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEmergencyResponse_Change); i {
 			case 0:
 				return &v.state
