@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/energy_storage.proto
 
 package traits
@@ -1190,7 +1190,7 @@ func file_traits_energy_storage_proto_rawDescGZIP() []byte {
 
 var file_traits_energy_storage_proto_enumTypes = make([]protoimpl.EnumInfo, 3)
 var file_traits_energy_storage_proto_msgTypes = make([]protoimpl.MessageInfo, 12)
-var file_traits_energy_storage_proto_goTypes = []interface{}{
+var file_traits_energy_storage_proto_goTypes = []any{
 	(EnergyLevelSupport_ChargeControl)(0),  // 0: smartcore.traits.EnergyLevelSupport.ChargeControl
 	(EnergyLevel_Transfer_Speed)(0),        // 1: smartcore.traits.EnergyLevel.Transfer.Speed
 	(EnergyLevel_Quantity_Threshold)(0),    // 2: smartcore.traits.EnergyLevel.Quantity.Threshold
@@ -1250,7 +1250,7 @@ func file_traits_energy_storage_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_energy_storage_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*EnergyLevelSupport); i {
 			case 0:
 				return &v.state
@@ -1262,7 +1262,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*EnergyLevel); i {
 			case 0:
 				return &v.state
@@ -1274,7 +1274,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetEnergyLevelRequest); i {
 			case 0:
 				return &v.state
@@ -1286,7 +1286,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEnergyLevelRequest); i {
 			case 0:
 				return &v.state
@@ -1298,7 +1298,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEnergyLevelResponse); i {
 			case 0:
 				return &v.state
@@ -1310,7 +1310,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ChargeRequest); i {
 			case 0:
 				return &v.state
@@ -1322,7 +1322,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*ChargeResponse); i {
 			case 0:
 				return &v.state
@@ -1334,7 +1334,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeEnergyLevelRequest); i {
 			case 0:
 				return &v.state
@@ -1346,7 +1346,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*EnergyLevel_Transfer); i {
 			case 0:
 				return &v.state
@@ -1358,7 +1358,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*EnergyLevel_Steady); i {
 			case 0:
 				return &v.state
@@ -1370,7 +1370,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*EnergyLevel_Quantity); i {
 			case 0:
 				return &v.state
@@ -1382,7 +1382,7 @@ func file_traits_energy_storage_proto_init() {
 				return nil
 			}
 		}
-		file_traits_energy_storage_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_energy_storage_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEnergyLevelResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1395,7 +1395,7 @@ func file_traits_energy_storage_proto_init() {
 			}
 		}
 	}
-	file_traits_energy_storage_proto_msgTypes[1].OneofWrappers = []interface{}{
+	file_traits_energy_storage_proto_msgTypes[1].OneofWrappers = []any{
 		(*EnergyLevel_Discharge)(nil),
 		(*EnergyLevel_Charge)(nil),
 		(*EnergyLevel_Idle)(nil),
