@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/enter_leave_sensor.proto
 
 package traits
@@ -743,7 +743,7 @@ func file_traits_enter_leave_sensor_proto_rawDescGZIP() []byte {
 
 var file_traits_enter_leave_sensor_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_enter_leave_sensor_proto_msgTypes = make([]protoimpl.MessageInfo, 10)
-var file_traits_enter_leave_sensor_proto_goTypes = []interface{}{
+var file_traits_enter_leave_sensor_proto_goTypes = []any{
 	(EnterLeaveEvent_Direction)(0),              // 0: smartcore.traits.EnterLeaveEvent.Direction
 	(*EnterLeaveEvent)(nil),                     // 1: smartcore.traits.EnterLeaveEvent
 	(*PullEnterLeaveEventsRequest)(nil),         // 2: smartcore.traits.PullEnterLeaveEventsRequest
@@ -789,7 +789,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_enter_leave_sensor_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*EnterLeaveEvent); i {
 			case 0:
 				return &v.state
@@ -801,7 +801,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_enter_leave_sensor_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEnterLeaveEventsRequest); i {
 			case 0:
 				return &v.state
@@ -813,7 +813,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_enter_leave_sensor_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEnterLeaveEventsResponse); i {
 			case 0:
 				return &v.state
@@ -825,7 +825,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_enter_leave_sensor_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetEnterLeaveEventRequest); i {
 			case 0:
 				return &v.state
@@ -837,7 +837,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_enter_leave_sensor_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ResetEnterLeaveTotalsRequest); i {
 			case 0:
 				return &v.state
@@ -849,7 +849,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_enter_leave_sensor_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ResetEnterLeaveTotalsResponse); i {
 			case 0:
 				return &v.state
@@ -861,7 +861,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_enter_leave_sensor_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*EnterLeaveEvent_Occupant); i {
 			case 0:
 				return &v.state
@@ -873,7 +873,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_enter_leave_sensor_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_enter_leave_sensor_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*PullEnterLeaveEventsResponse_Change); i {
 			case 0:
 				return &v.state
@@ -886,7 +886,7 @@ func file_traits_enter_leave_sensor_proto_init() {
 			}
 		}
 	}
-	file_traits_enter_leave_sensor_proto_msgTypes[0].OneofWrappers = []interface{}{}
+	file_traits_enter_leave_sensor_proto_msgTypes[0].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
