@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/extend_retract.proto
 
 package traits
@@ -892,7 +892,7 @@ func file_traits_extend_retract_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_extend_retract_proto_msgTypes = make([]protoimpl.MessageInfo, 11)
-var file_traits_extend_retract_proto_goTypes = []interface{}{
+var file_traits_extend_retract_proto_goTypes = []any{
 	(*Extension)(nil),                     // 0: smartcore.traits.Extension
 	(*ExtensionPreset)(nil),               // 1: smartcore.traits.ExtensionPreset
 	(*ExtensionSupport)(nil),              // 2: smartcore.traits.ExtensionSupport
@@ -948,7 +948,7 @@ func file_traits_extend_retract_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_extend_retract_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Extension); i {
 			case 0:
 				return &v.state
@@ -960,7 +960,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ExtensionPreset); i {
 			case 0:
 				return &v.state
@@ -972,7 +972,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ExtensionSupport); i {
 			case 0:
 				return &v.state
@@ -984,7 +984,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetExtensionRequest); i {
 			case 0:
 				return &v.state
@@ -996,7 +996,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateExtensionRequest); i {
 			case 0:
 				return &v.state
@@ -1008,7 +1008,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ExtendRetractStopRequest); i {
 			case 0:
 				return &v.state
@@ -1020,7 +1020,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateExtensionPresetRequest); i {
 			case 0:
 				return &v.state
@@ -1032,7 +1032,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullExtensionsRequest); i {
 			case 0:
 				return &v.state
@@ -1044,7 +1044,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullExtensionsResponse); i {
 			case 0:
 				return &v.state
@@ -1056,7 +1056,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeExtensionRequest); i {
 			case 0:
 				return &v.state
@@ -1068,7 +1068,7 @@ func file_traits_extend_retract_proto_init() {
 				return nil
 			}
 		}
-		file_traits_extend_retract_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_extend_retract_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*PullExtensionsResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1081,7 +1081,7 @@ func file_traits_extend_retract_proto_init() {
 			}
 		}
 	}
-	file_traits_extend_retract_proto_msgTypes[1].OneofWrappers = []interface{}{}
+	file_traits_extend_retract_proto_msgTypes[1].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
