@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/fan_speed.proto
 
 package traits
@@ -760,7 +760,7 @@ func file_traits_fan_speed_proto_rawDescGZIP() []byte {
 
 var file_traits_fan_speed_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_fan_speed_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
-var file_traits_fan_speed_proto_goTypes = []interface{}{
+var file_traits_fan_speed_proto_goTypes = []any{
 	(FanSpeed_Direction)(0),                 // 0: smartcore.traits.FanSpeed.Direction
 	(*FanSpeed)(nil),                        // 1: smartcore.traits.FanSpeed
 	(*FanSpeedSupport)(nil),                 // 2: smartcore.traits.FanSpeedSupport
@@ -808,7 +808,7 @@ func file_traits_fan_speed_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_fan_speed_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*FanSpeed); i {
 			case 0:
 				return &v.state
@@ -820,7 +820,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*FanSpeedSupport); i {
 			case 0:
 				return &v.state
@@ -832,7 +832,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetFanSpeedRequest); i {
 			case 0:
 				return &v.state
@@ -844,7 +844,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateFanSpeedRequest); i {
 			case 0:
 				return &v.state
@@ -856,7 +856,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ReverseFanSpeedDirectionRequest); i {
 			case 0:
 				return &v.state
@@ -868,7 +868,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullFanSpeedRequest); i {
 			case 0:
 				return &v.state
@@ -880,7 +880,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullFanSpeedResponse); i {
 			case 0:
 				return &v.state
@@ -892,7 +892,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeFanSpeedRequest); i {
 			case 0:
 				return &v.state
@@ -904,7 +904,7 @@ func file_traits_fan_speed_proto_init() {
 				return nil
 			}
 		}
-		file_traits_fan_speed_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_fan_speed_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullFanSpeedResponse_Change); i {
 			case 0:
 				return &v.state
