@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/hail.proto
 
 package traits
@@ -1325,7 +1325,7 @@ func file_traits_hail_proto_rawDescGZIP() []byte {
 
 var file_traits_hail_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_hail_proto_msgTypes = make([]protoimpl.MessageInfo, 15)
-var file_traits_hail_proto_goTypes = []interface{}{
+var file_traits_hail_proto_goTypes = []any{
 	(Hail_State)(0),                  // 0: smartcore.traits.Hail.State
 	(*Hail)(nil),                     // 1: smartcore.traits.Hail
 	(*CreateHailRequest)(nil),        // 2: smartcore.traits.CreateHailRequest
@@ -1397,7 +1397,7 @@ func file_traits_hail_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_hail_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Hail); i {
 			case 0:
 				return &v.state
@@ -1409,7 +1409,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateHailRequest); i {
 			case 0:
 				return &v.state
@@ -1421,7 +1421,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetHailRequest); i {
 			case 0:
 				return &v.state
@@ -1433,7 +1433,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateHailRequest); i {
 			case 0:
 				return &v.state
@@ -1445,7 +1445,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteHailRequest); i {
 			case 0:
 				return &v.state
@@ -1457,7 +1457,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteHailResponse); i {
 			case 0:
 				return &v.state
@@ -1469,7 +1469,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHailRequest); i {
 			case 0:
 				return &v.state
@@ -1481,7 +1481,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHailResponse); i {
 			case 0:
 				return &v.state
@@ -1493,7 +1493,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*ListHailsRequest); i {
 			case 0:
 				return &v.state
@@ -1505,7 +1505,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*ListHailsResponse); i {
 			case 0:
 				return &v.state
@@ -1517,7 +1517,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHailsRequest); i {
 			case 0:
 				return &v.state
@@ -1529,7 +1529,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHailsResponse); i {
 			case 0:
 				return &v.state
@@ -1541,7 +1541,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*Hail_Location); i {
 			case 0:
 				return &v.state
@@ -1553,7 +1553,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHailResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1565,7 +1565,7 @@ func file_traits_hail_proto_init() {
 				return nil
 			}
 		}
-		file_traits_hail_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_hail_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*PullHailsResponse_Change); i {
 			case 0:
 				return &v.state
