@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/input_select.proto
 
 package traits
@@ -826,7 +826,7 @@ func file_traits_input_select_proto_rawDescGZIP() []byte {
 
 var file_traits_input_select_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_input_select_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
-var file_traits_input_select_proto_goTypes = []interface{}{
+var file_traits_input_select_proto_goTypes = []any{
 	(InputSupport_Feature)(0),        // 0: smartcore.traits.InputSupport.Feature
 	(*Input)(nil),                    // 1: smartcore.traits.Input
 	(*InputSupport)(nil),             // 2: smartcore.traits.InputSupport
@@ -875,7 +875,7 @@ func file_traits_input_select_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_input_select_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Input); i {
 			case 0:
 				return &v.state
@@ -887,7 +887,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*InputSupport); i {
 			case 0:
 				return &v.state
@@ -899,7 +899,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*AvPort); i {
 			case 0:
 				return &v.state
@@ -911,7 +911,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateInputRequest); i {
 			case 0:
 				return &v.state
@@ -923,7 +923,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*GetInputRequest); i {
 			case 0:
 				return &v.state
@@ -935,7 +935,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullInputRequest); i {
 			case 0:
 				return &v.state
@@ -947,7 +947,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullInputResponse); i {
 			case 0:
 				return &v.state
@@ -959,7 +959,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeInputRequest); i {
 			case 0:
 				return &v.state
@@ -971,7 +971,7 @@ func file_traits_input_select_proto_init() {
 				return nil
 			}
 		}
-		file_traits_input_select_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_input_select_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullInputResponse_Change); i {
 			case 0:
 				return &v.state
