@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/light.proto
 
 package traits
@@ -764,7 +764,7 @@ func file_traits_light_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_light_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
-var file_traits_light_proto_goTypes = []interface{}{
+var file_traits_light_proto_goTypes = []any{
 	(*Brightness)(nil),                    // 0: smartcore.traits.Brightness
 	(*LightPreset)(nil),                   // 1: smartcore.traits.LightPreset
 	(*BrightnessSupport)(nil),             // 2: smartcore.traits.BrightnessSupport
@@ -815,7 +815,7 @@ func file_traits_light_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_light_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Brightness); i {
 			case 0:
 				return &v.state
@@ -827,7 +827,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*LightPreset); i {
 			case 0:
 				return &v.state
@@ -839,7 +839,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*BrightnessSupport); i {
 			case 0:
 				return &v.state
@@ -851,7 +851,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateBrightnessRequest); i {
 			case 0:
 				return &v.state
@@ -863,7 +863,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*GetBrightnessRequest); i {
 			case 0:
 				return &v.state
@@ -875,7 +875,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullBrightnessRequest); i {
 			case 0:
 				return &v.state
@@ -887,7 +887,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullBrightnessResponse); i {
 			case 0:
 				return &v.state
@@ -899,7 +899,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeBrightnessRequest); i {
 			case 0:
 				return &v.state
@@ -911,7 +911,7 @@ func file_traits_light_proto_init() {
 				return nil
 			}
 		}
-		file_traits_light_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_light_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullBrightnessResponse_Change); i {
 			case 0:
 				return &v.state
