@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/lock_unlock.proto
 
 package traits
@@ -558,7 +558,7 @@ func file_traits_lock_unlock_proto_rawDescGZIP() []byte {
 
 var file_traits_lock_unlock_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_lock_unlock_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_traits_lock_unlock_proto_goTypes = []interface{}{
+var file_traits_lock_unlock_proto_goTypes = []any{
 	(LockUnlock_Position)(0),              // 0: smartcore.traits.LockUnlock.Position
 	(*LockUnlock)(nil),                    // 1: smartcore.traits.LockUnlock
 	(*GetLockUnlockRequest)(nil),          // 2: smartcore.traits.GetLockUnlockRequest
@@ -597,7 +597,7 @@ func file_traits_lock_unlock_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_lock_unlock_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_lock_unlock_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*LockUnlock); i {
 			case 0:
 				return &v.state
@@ -609,7 +609,7 @@ func file_traits_lock_unlock_proto_init() {
 				return nil
 			}
 		}
-		file_traits_lock_unlock_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_lock_unlock_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*GetLockUnlockRequest); i {
 			case 0:
 				return &v.state
@@ -621,7 +621,7 @@ func file_traits_lock_unlock_proto_init() {
 				return nil
 			}
 		}
-		file_traits_lock_unlock_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_lock_unlock_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateLockUnlockRequest); i {
 			case 0:
 				return &v.state
@@ -633,7 +633,7 @@ func file_traits_lock_unlock_proto_init() {
 				return nil
 			}
 		}
-		file_traits_lock_unlock_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_lock_unlock_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullLockUnlockRequest); i {
 			case 0:
 				return &v.state
@@ -645,7 +645,7 @@ func file_traits_lock_unlock_proto_init() {
 				return nil
 			}
 		}
-		file_traits_lock_unlock_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_lock_unlock_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullLockUnlockResponse); i {
 			case 0:
 				return &v.state
@@ -657,7 +657,7 @@ func file_traits_lock_unlock_proto_init() {
 				return nil
 			}
 		}
-		file_traits_lock_unlock_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_lock_unlock_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullLockUnlockResponse_Change); i {
 			case 0:
 				return &v.state
