@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/metadata.proto
 
 package traits
@@ -1561,7 +1561,7 @@ func file_traits_metadata_proto_rawDescGZIP() []byte {
 
 var file_traits_metadata_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_metadata_proto_msgTypes = make([]protoimpl.MessageInfo, 25)
-var file_traits_metadata_proto_goTypes = []interface{}{
+var file_traits_metadata_proto_goTypes = []any{
 	(Metadata_NIC_Assignment)(0),        // 0: smartcore.traits.Metadata.NIC.Assignment
 	(*Metadata)(nil),                    // 1: smartcore.traits.Metadata
 	(*TraitMetadata)(nil),               // 2: smartcore.traits.TraitMetadata
@@ -1639,7 +1639,7 @@ func file_traits_metadata_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_metadata_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata); i {
 			case 0:
 				return &v.state
@@ -1651,7 +1651,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*TraitMetadata); i {
 			case 0:
 				return &v.state
@@ -1663,7 +1663,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetMetadataRequest); i {
 			case 0:
 				return &v.state
@@ -1675,7 +1675,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullMetadataRequest); i {
 			case 0:
 				return &v.state
@@ -1687,7 +1687,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullMetadataResponse); i {
 			case 0:
 				return &v.state
@@ -1699,7 +1699,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_Appearance); i {
 			case 0:
 				return &v.state
@@ -1711,7 +1711,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_Location); i {
 			case 0:
 				return &v.state
@@ -1723,7 +1723,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_ID); i {
 			case 0:
 				return &v.state
@@ -1735,7 +1735,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_Product); i {
 			case 0:
 				return &v.state
@@ -1747,7 +1747,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_Revision); i {
 			case 0:
 				return &v.state
@@ -1759,7 +1759,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_Installation); i {
 			case 0:
 				return &v.state
@@ -1771,7 +1771,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_NIC); i {
 			case 0:
 				return &v.state
@@ -1783,7 +1783,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_Membership); i {
 			case 0:
 				return &v.state
@@ -1795,7 +1795,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[17].Exporter = func(v any, i int) any {
 			switch v := v.(*Metadata_Product_Kind); i {
 			case 0:
 				return &v.state
@@ -1807,7 +1807,7 @@ func file_traits_metadata_proto_init() {
 				return nil
 			}
 		}
-		file_traits_metadata_proto_msgTypes[24].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_metadata_proto_msgTypes[24].Exporter = func(v any, i int) any {
 			switch v := v.(*PullMetadataResponse_Change); i {
 			case 0:
 				return &v.state
