@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/mode.proto
 
 package traits
@@ -857,7 +857,7 @@ func file_traits_mode_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_mode_proto_msgTypes = make([]protoimpl.MessageInfo, 14)
-var file_traits_mode_proto_goTypes = []interface{}{
+var file_traits_mode_proto_goTypes = []any{
 	(*ModeValues)(nil),                    // 0: smartcore.traits.ModeValues
 	(*ModeValuesRelative)(nil),            // 1: smartcore.traits.ModeValuesRelative
 	(*Modes)(nil),                         // 2: smartcore.traits.Modes
@@ -912,7 +912,7 @@ func file_traits_mode_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_mode_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*ModeValues); i {
 			case 0:
 				return &v.state
@@ -924,7 +924,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ModeValuesRelative); i {
 			case 0:
 				return &v.state
@@ -936,7 +936,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*Modes); i {
 			case 0:
 				return &v.state
@@ -948,7 +948,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ModesSupport); i {
 			case 0:
 				return &v.state
@@ -960,7 +960,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*GetModeValuesRequest); i {
 			case 0:
 				return &v.state
@@ -972,7 +972,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateModeValuesRequest); i {
 			case 0:
 				return &v.state
@@ -984,7 +984,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullModeValuesRequest); i {
 			case 0:
 				return &v.state
@@ -996,7 +996,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullModeValuesResponse); i {
 			case 0:
 				return &v.state
@@ -1008,7 +1008,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeModesRequest); i {
 			case 0:
 				return &v.state
@@ -1020,7 +1020,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*Modes_Value); i {
 			case 0:
 				return &v.state
@@ -1032,7 +1032,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*Modes_Mode); i {
 			case 0:
 				return &v.state
@@ -1044,7 +1044,7 @@ func file_traits_mode_proto_init() {
 				return nil
 			}
 		}
-		file_traits_mode_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_mode_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*PullModeValuesResponse_Change); i {
 			case 0:
 				return &v.state
