@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/motion_sensor.proto
 
 package traits
@@ -607,7 +607,7 @@ func file_traits_motion_sensor_proto_rawDescGZIP() []byte {
 
 var file_traits_motion_sensor_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_motion_sensor_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_traits_motion_sensor_proto_goTypes = []interface{}{
+var file_traits_motion_sensor_proto_goTypes = []any{
 	(MotionDetection_State)(0),                 // 0: smartcore.traits.MotionDetection.State
 	(*MotionDetection)(nil),                    // 1: smartcore.traits.MotionDetection
 	(*MotionDetectionSupport)(nil),             // 2: smartcore.traits.MotionDetectionSupport
@@ -650,7 +650,7 @@ func file_traits_motion_sensor_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_motion_sensor_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_motion_sensor_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*MotionDetection); i {
 			case 0:
 				return &v.state
@@ -662,7 +662,7 @@ func file_traits_motion_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_motion_sensor_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_motion_sensor_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*MotionDetectionSupport); i {
 			case 0:
 				return &v.state
@@ -674,7 +674,7 @@ func file_traits_motion_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_motion_sensor_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_motion_sensor_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetMotionDetectionRequest); i {
 			case 0:
 				return &v.state
@@ -686,7 +686,7 @@ func file_traits_motion_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_motion_sensor_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_motion_sensor_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullMotionDetectionRequest); i {
 			case 0:
 				return &v.state
@@ -698,7 +698,7 @@ func file_traits_motion_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_motion_sensor_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_motion_sensor_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullMotionDetectionResponse); i {
 			case 0:
 				return &v.state
@@ -710,7 +710,7 @@ func file_traits_motion_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_motion_sensor_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_motion_sensor_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeMotionDetectionRequest); i {
 			case 0:
 				return &v.state
@@ -722,7 +722,7 @@ func file_traits_motion_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_motion_sensor_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_motion_sensor_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullMotionDetectionResponse_Change); i {
 			case 0:
 				return &v.state
