@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/occupancy_sensor.proto
 
 package traits
@@ -639,7 +639,7 @@ func file_traits_occupancy_sensor_proto_rawDescGZIP() []byte {
 
 var file_traits_occupancy_sensor_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_occupancy_sensor_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_traits_occupancy_sensor_proto_goTypes = []interface{}{
+var file_traits_occupancy_sensor_proto_goTypes = []any{
 	(Occupancy_State)(0),                 // 0: smartcore.traits.Occupancy.State
 	(*Occupancy)(nil),                    // 1: smartcore.traits.Occupancy
 	(*OccupancySupport)(nil),             // 2: smartcore.traits.OccupancySupport
@@ -680,7 +680,7 @@ func file_traits_occupancy_sensor_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_occupancy_sensor_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_occupancy_sensor_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Occupancy); i {
 			case 0:
 				return &v.state
@@ -692,7 +692,7 @@ func file_traits_occupancy_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_occupancy_sensor_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_occupancy_sensor_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*OccupancySupport); i {
 			case 0:
 				return &v.state
@@ -704,7 +704,7 @@ func file_traits_occupancy_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_occupancy_sensor_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_occupancy_sensor_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetOccupancyRequest); i {
 			case 0:
 				return &v.state
@@ -716,7 +716,7 @@ func file_traits_occupancy_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_occupancy_sensor_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_occupancy_sensor_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOccupancyRequest); i {
 			case 0:
 				return &v.state
@@ -728,7 +728,7 @@ func file_traits_occupancy_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_occupancy_sensor_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_occupancy_sensor_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOccupancyResponse); i {
 			case 0:
 				return &v.state
@@ -740,7 +740,7 @@ func file_traits_occupancy_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_occupancy_sensor_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_occupancy_sensor_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeOccupancyRequest); i {
 			case 0:
 				return &v.state
@@ -752,7 +752,7 @@ func file_traits_occupancy_sensor_proto_init() {
 				return nil
 			}
 		}
-		file_traits_occupancy_sensor_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_occupancy_sensor_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOccupancyResponse_Change); i {
 			case 0:
 				return &v.state
