@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/on_off.proto
 
 package traits
@@ -628,7 +628,7 @@ func file_traits_on_off_proto_rawDescGZIP() []byte {
 
 var file_traits_on_off_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_on_off_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_traits_on_off_proto_goTypes = []interface{}{
+var file_traits_on_off_proto_goTypes = []any{
 	(OnOff_State)(0),                 // 0: smartcore.traits.OnOff.State
 	(*OnOff)(nil),                    // 1: smartcore.traits.OnOff
 	(*OnOffSupport)(nil),             // 2: smartcore.traits.OnOffSupport
@@ -673,7 +673,7 @@ func file_traits_on_off_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_on_off_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*OnOff); i {
 			case 0:
 				return &v.state
@@ -685,7 +685,7 @@ func file_traits_on_off_proto_init() {
 				return nil
 			}
 		}
-		file_traits_on_off_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*OnOffSupport); i {
 			case 0:
 				return &v.state
@@ -697,7 +697,7 @@ func file_traits_on_off_proto_init() {
 				return nil
 			}
 		}
-		file_traits_on_off_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetOnOffRequest); i {
 			case 0:
 				return &v.state
@@ -709,7 +709,7 @@ func file_traits_on_off_proto_init() {
 				return nil
 			}
 		}
-		file_traits_on_off_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateOnOffRequest); i {
 			case 0:
 				return &v.state
@@ -721,7 +721,7 @@ func file_traits_on_off_proto_init() {
 				return nil
 			}
 		}
-		file_traits_on_off_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOnOffRequest); i {
 			case 0:
 				return &v.state
@@ -733,7 +733,7 @@ func file_traits_on_off_proto_init() {
 				return nil
 			}
 		}
-		file_traits_on_off_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOnOffResponse); i {
 			case 0:
 				return &v.state
@@ -745,7 +745,7 @@ func file_traits_on_off_proto_init() {
 				return nil
 			}
 		}
-		file_traits_on_off_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeOnOffRequest); i {
 			case 0:
 				return &v.state
@@ -757,7 +757,7 @@ func file_traits_on_off_proto_init() {
 				return nil
 			}
 		}
-		file_traits_on_off_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_on_off_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOnOffResponse_Change); i {
 			case 0:
 				return &v.state
