@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/open_close.proto
 
 package traits
@@ -968,7 +968,7 @@ func file_traits_open_close_proto_rawDescGZIP() []byte {
 
 var file_traits_open_close_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
 var file_traits_open_close_proto_msgTypes = make([]protoimpl.MessageInfo, 10)
-var file_traits_open_close_proto_goTypes = []interface{}{
+var file_traits_open_close_proto_goTypes = []any{
 	(OpenClosePosition_Direction)(0),              // 0: smartcore.traits.OpenClosePosition.Direction
 	(OpenClosePosition_Resistance)(0),             // 1: smartcore.traits.OpenClosePosition.Resistance
 	(*OpenClosePositions)(nil),                    // 2: smartcore.traits.OpenClosePositions
@@ -1025,7 +1025,7 @@ func file_traits_open_close_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_open_close_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*OpenClosePositions); i {
 			case 0:
 				return &v.state
@@ -1037,7 +1037,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*OpenClosePosition); i {
 			case 0:
 				return &v.state
@@ -1049,7 +1049,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PositionsSupport); i {
 			case 0:
 				return &v.state
@@ -1061,7 +1061,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetOpenClosePositionsRequest); i {
 			case 0:
 				return &v.state
@@ -1073,7 +1073,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateOpenClosePositionsRequest); i {
 			case 0:
 				return &v.state
@@ -1085,7 +1085,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*StopOpenCloseRequest); i {
 			case 0:
 				return &v.state
@@ -1097,7 +1097,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOpenClosePositionsRequest); i {
 			case 0:
 				return &v.state
@@ -1109,7 +1109,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOpenClosePositionsResponse); i {
 			case 0:
 				return &v.state
@@ -1121,7 +1121,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribePositionsRequest); i {
 			case 0:
 				return &v.state
@@ -1133,7 +1133,7 @@ func file_traits_open_close_proto_init() {
 				return nil
 			}
 		}
-		file_traits_open_close_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_open_close_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*PullOpenClosePositionsResponse_Change); i {
 			case 0:
 				return &v.state
