@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/parent.proto
 
 package traits
@@ -591,7 +591,7 @@ func file_traits_parent_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_parent_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_traits_parent_proto_goTypes = []interface{}{
+var file_traits_parent_proto_goTypes = []any{
 	(*Child)(nil),                       // 0: smartcore.traits.Child
 	(*Trait)(nil),                       // 1: smartcore.traits.Trait
 	(*ListChildrenRequest)(nil),         // 2: smartcore.traits.ListChildrenRequest
@@ -630,7 +630,7 @@ func file_traits_parent_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_parent_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_parent_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Child); i {
 			case 0:
 				return &v.state
@@ -642,7 +642,7 @@ func file_traits_parent_proto_init() {
 				return nil
 			}
 		}
-		file_traits_parent_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_parent_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Trait); i {
 			case 0:
 				return &v.state
@@ -654,7 +654,7 @@ func file_traits_parent_proto_init() {
 				return nil
 			}
 		}
-		file_traits_parent_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_parent_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ListChildrenRequest); i {
 			case 0:
 				return &v.state
@@ -666,7 +666,7 @@ func file_traits_parent_proto_init() {
 				return nil
 			}
 		}
-		file_traits_parent_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_parent_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ListChildrenResponse); i {
 			case 0:
 				return &v.state
@@ -678,7 +678,7 @@ func file_traits_parent_proto_init() {
 				return nil
 			}
 		}
-		file_traits_parent_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_parent_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullChildrenRequest); i {
 			case 0:
 				return &v.state
@@ -690,7 +690,7 @@ func file_traits_parent_proto_init() {
 				return nil
 			}
 		}
-		file_traits_parent_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_parent_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullChildrenResponse); i {
 			case 0:
 				return &v.state
@@ -702,7 +702,7 @@ func file_traits_parent_proto_init() {
 				return nil
 			}
 		}
-		file_traits_parent_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_parent_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullChildrenResponse_Change); i {
 			case 0:
 				return &v.state
