@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/ptz.proto
 
 package traits
@@ -1186,7 +1186,7 @@ func file_traits_ptz_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_ptz_proto_msgTypes = make([]protoimpl.MessageInfo, 15)
-var file_traits_ptz_proto_goTypes = []interface{}{
+var file_traits_ptz_proto_goTypes = []any{
 	(*Ptz)(nil),                    // 0: smartcore.traits.Ptz
 	(*PtzVector)(nil),              // 1: smartcore.traits.PtzVector
 	(*PtzBounds)(nil),              // 2: smartcore.traits.PtzBounds
@@ -1253,7 +1253,7 @@ func file_traits_ptz_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_ptz_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Ptz); i {
 			case 0:
 				return &v.state
@@ -1265,7 +1265,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PtzVector); i {
 			case 0:
 				return &v.state
@@ -1277,7 +1277,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PtzBounds); i {
 			case 0:
 				return &v.state
@@ -1289,7 +1289,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PtzMovement); i {
 			case 0:
 				return &v.state
@@ -1301,7 +1301,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PtzPosition); i {
 			case 0:
 				return &v.state
@@ -1313,7 +1313,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PtzPreset); i {
 			case 0:
 				return &v.state
@@ -1325,7 +1325,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PtzSupport); i {
 			case 0:
 				return &v.state
@@ -1337,7 +1337,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*GetPtzRequest); i {
 			case 0:
 				return &v.state
@@ -1349,7 +1349,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdatePtzRequest); i {
 			case 0:
 				return &v.state
@@ -1361,7 +1361,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*StopPtzRequest); i {
 			case 0:
 				return &v.state
@@ -1373,7 +1373,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*CreatePtzPresetRequest); i {
 			case 0:
 				return &v.state
@@ -1385,7 +1385,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPtzRequest); i {
 			case 0:
 				return &v.state
@@ -1397,7 +1397,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPtzResponse); i {
 			case 0:
 				return &v.state
@@ -1409,7 +1409,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribePtzRequest); i {
 			case 0:
 				return &v.state
@@ -1421,7 +1421,7 @@ func file_traits_ptz_proto_init() {
 				return nil
 			}
 		}
-		file_traits_ptz_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_ptz_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPtzResponse_Change); i {
 			case 0:
 				return &v.state
