@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/publication.proto
 
 package traits
@@ -1441,7 +1441,7 @@ func file_traits_publication_proto_rawDescGZIP() []byte {
 
 var file_traits_publication_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_publication_proto_msgTypes = make([]protoimpl.MessageInfo, 15)
-var file_traits_publication_proto_goTypes = []interface{}{
+var file_traits_publication_proto_goTypes = []any{
 	(Publication_Audience_Receipt)(0),       // 0: smartcore.traits.Publication.Audience.Receipt
 	(*Publication)(nil),                     // 1: smartcore.traits.Publication
 	(*CreatePublicationRequest)(nil),        // 2: smartcore.traits.CreatePublicationRequest
@@ -1513,7 +1513,7 @@ func file_traits_publication_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_publication_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Publication); i {
 			case 0:
 				return &v.state
@@ -1525,7 +1525,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CreatePublicationRequest); i {
 			case 0:
 				return &v.state
@@ -1537,7 +1537,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetPublicationRequest); i {
 			case 0:
 				return &v.state
@@ -1549,7 +1549,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdatePublicationRequest); i {
 			case 0:
 				return &v.state
@@ -1561,7 +1561,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*DeletePublicationRequest); i {
 			case 0:
 				return &v.state
@@ -1573,7 +1573,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPublicationRequest); i {
 			case 0:
 				return &v.state
@@ -1585,7 +1585,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPublicationResponse); i {
 			case 0:
 				return &v.state
@@ -1597,7 +1597,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*ListPublicationsRequest); i {
 			case 0:
 				return &v.state
@@ -1609,7 +1609,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*ListPublicationsResponse); i {
 			case 0:
 				return &v.state
@@ -1621,7 +1621,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPublicationsRequest); i {
 			case 0:
 				return &v.state
@@ -1633,7 +1633,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPublicationsResponse); i {
 			case 0:
 				return &v.state
@@ -1645,7 +1645,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*AcknowledgePublicationRequest); i {
 			case 0:
 				return &v.state
@@ -1657,7 +1657,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*Publication_Audience); i {
 			case 0:
 				return &v.state
@@ -1669,7 +1669,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPublicationResponse_Change); i {
 			case 0:
 				return &v.state
@@ -1681,7 +1681,7 @@ func file_traits_publication_proto_init() {
 				return nil
 			}
 		}
-		file_traits_publication_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_publication_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*PullPublicationsResponse_Change); i {
 			case 0:
 				return &v.state
