@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/speaker.proto
 
 package traits
@@ -496,7 +496,7 @@ func file_traits_speaker_proto_rawDescGZIP() []byte {
 }
 
 var file_traits_speaker_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_traits_speaker_proto_goTypes = []interface{}{
+var file_traits_speaker_proto_goTypes = []any{
 	(*VolumeSupport)(nil),              // 0: smartcore.traits.VolumeSupport
 	(*GetSpeakerVolumeRequest)(nil),    // 1: smartcore.traits.GetSpeakerVolumeRequest
 	(*UpdateSpeakerVolumeRequest)(nil), // 2: smartcore.traits.UpdateSpeakerVolumeRequest
@@ -540,7 +540,7 @@ func file_traits_speaker_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_speaker_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_speaker_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*VolumeSupport); i {
 			case 0:
 				return &v.state
@@ -552,7 +552,7 @@ func file_traits_speaker_proto_init() {
 				return nil
 			}
 		}
-		file_traits_speaker_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_speaker_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*GetSpeakerVolumeRequest); i {
 			case 0:
 				return &v.state
@@ -564,7 +564,7 @@ func file_traits_speaker_proto_init() {
 				return nil
 			}
 		}
-		file_traits_speaker_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_speaker_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateSpeakerVolumeRequest); i {
 			case 0:
 				return &v.state
@@ -576,7 +576,7 @@ func file_traits_speaker_proto_init() {
 				return nil
 			}
 		}
-		file_traits_speaker_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_speaker_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullSpeakerVolumeRequest); i {
 			case 0:
 				return &v.state
@@ -588,7 +588,7 @@ func file_traits_speaker_proto_init() {
 				return nil
 			}
 		}
-		file_traits_speaker_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_speaker_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullSpeakerVolumeResponse); i {
 			case 0:
 				return &v.state
@@ -600,7 +600,7 @@ func file_traits_speaker_proto_init() {
 				return nil
 			}
 		}
-		file_traits_speaker_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_speaker_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*DescribeVolumeRequest); i {
 			case 0:
 				return &v.state
