@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: traits/vending.proto
 
 package traits
@@ -1925,7 +1925,7 @@ func file_traits_vending_proto_rawDescGZIP() []byte {
 
 var file_traits_vending_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_traits_vending_proto_msgTypes = make([]protoimpl.MessageInfo, 23)
-var file_traits_vending_proto_goTypes = []interface{}{
+var file_traits_vending_proto_goTypes = []any{
 	(Consumable_Unit)(0),                   // 0: smartcore.traits.Consumable.Unit
 	(*Consumable)(nil),                     // 1: smartcore.traits.Consumable
 	(*ListConsumablesRequest)(nil),         // 2: smartcore.traits.ListConsumablesRequest
@@ -2024,7 +2024,7 @@ func file_traits_vending_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_traits_vending_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Consumable); i {
 			case 0:
 				return &v.state
@@ -2036,7 +2036,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ListConsumablesRequest); i {
 			case 0:
 				return &v.state
@@ -2048,7 +2048,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ListConsumablesResponse); i {
 			case 0:
 				return &v.state
@@ -2060,7 +2060,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PullConsumablesRequest); i {
 			case 0:
 				return &v.state
@@ -2072,7 +2072,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PullConsumablesResponse); i {
 			case 0:
 				return &v.state
@@ -2084,7 +2084,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*GetStockRequest); i {
 			case 0:
 				return &v.state
@@ -2096,7 +2096,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateStockRequest); i {
 			case 0:
 				return &v.state
@@ -2108,7 +2108,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PullStockRequest); i {
 			case 0:
 				return &v.state
@@ -2120,7 +2120,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PullStockResponse); i {
 			case 0:
 				return &v.state
@@ -2132,7 +2132,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*ListInventoryRequest); i {
 			case 0:
 				return &v.state
@@ -2144,7 +2144,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*ListInventoryResponse); i {
 			case 0:
 				return &v.state
@@ -2156,7 +2156,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*PullInventoryRequest); i {
 			case 0:
 				return &v.state
@@ -2168,7 +2168,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*PullInventoryResponse); i {
 			case 0:
 				return &v.state
@@ -2180,7 +2180,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*DispenseRequest); i {
 			case 0:
 				return &v.state
@@ -2192,7 +2192,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*StopDispenseRequest); i {
 			case 0:
 				return &v.state
@@ -2204,7 +2204,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*Consumable_Portion); i {
 			case 0:
 				return &v.state
@@ -2216,7 +2216,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[16].Exporter = func(v any, i int) any {
 			switch v := v.(*Consumable_Quantity); i {
 			case 0:
 				return &v.state
@@ -2228,7 +2228,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[19].Exporter = func(v any, i int) any {
 			switch v := v.(*Consumable_Stock); i {
 			case 0:
 				return &v.state
@@ -2240,7 +2240,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[20].Exporter = func(v any, i int) any {
 			switch v := v.(*PullConsumablesResponse_Change); i {
 			case 0:
 				return &v.state
@@ -2252,7 +2252,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[21].Exporter = func(v any, i int) any {
 			switch v := v.(*PullStockResponse_Change); i {
 			case 0:
 				return &v.state
@@ -2264,7 +2264,7 @@ func file_traits_vending_proto_init() {
 				return nil
 			}
 		}
-		file_traits_vending_proto_msgTypes[22].Exporter = func(v interface{}, i int) interface{} {
+		file_traits_vending_proto_msgTypes[22].Exporter = func(v any, i int) any {
 			switch v := v.(*PullInventoryResponse_Change); i {
 			case 0:
 				return &v.state
