@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/connection.proto
 
 package types
@@ -176,7 +176,7 @@ func file_types_connection_proto_rawDescGZIP() []byte {
 }
 
 var file_types_connection_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
-var file_types_connection_proto_goTypes = []interface{}{
+var file_types_connection_proto_goTypes = []any{
 	(Connectivity)(0), // 0: smartcore.types.Connectivity
 	(CommStatus)(0),   // 1: smartcore.types.CommStatus
 }
