@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/image.proto
 
 package types
@@ -419,7 +419,7 @@ func file_types_image_proto_rawDescGZIP() []byte {
 
 var file_types_image_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_types_image_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
-var file_types_image_proto_goTypes = []interface{}{
+var file_types_image_proto_goTypes = []any{
 	(Image_Source_Purpose)(0), // 0: smartcore.types.Image.Source.Purpose
 	(*Image)(nil),             // 1: smartcore.types.Image
 	(*Image_Content)(nil),     // 2: smartcore.types.Image.Content
@@ -442,7 +442,7 @@ func file_types_image_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_types_image_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_types_image_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Image); i {
 			case 0:
 				return &v.state
@@ -454,7 +454,7 @@ func file_types_image_proto_init() {
 				return nil
 			}
 		}
-		file_types_image_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_types_image_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Image_Content); i {
 			case 0:
 				return &v.state
@@ -466,7 +466,7 @@ func file_types_image_proto_init() {
 				return nil
 			}
 		}
-		file_types_image_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_types_image_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*Image_Source); i {
 			case 0:
 				return &v.state
@@ -479,7 +479,7 @@ func file_types_image_proto_init() {
 			}
 		}
 	}
-	file_types_image_proto_msgTypes[1].OneofWrappers = []interface{}{
+	file_types_image_proto_msgTypes[1].OneofWrappers = []any{
 		(*Image_Content_Body)(nil),
 		(*Image_Content_Url)(nil),
 		(*Image_Content_Ref)(nil),
