@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/info.proto
 
 package types
@@ -234,7 +234,7 @@ func file_types_info_proto_rawDescGZIP() []byte {
 
 var file_types_info_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_types_info_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_types_info_proto_goTypes = []interface{}{
+var file_types_info_proto_goTypes = []any{
 	(PullSupport)(0),              // 0: smartcore.types.PullSupport
 	(*ResourceSupport)(nil),       // 1: smartcore.types.ResourceSupport
 	(*fieldmaskpb.FieldMask)(nil), // 2: google.protobuf.FieldMask
@@ -257,7 +257,7 @@ func file_types_info_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_types_info_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_types_info_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*ResourceSupport); i {
 			case 0:
 				return &v.state
