@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/number.proto
 
 package types
@@ -532,7 +532,7 @@ func file_types_number_proto_rawDescGZIP() []byte {
 
 var file_types_number_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_types_number_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_types_number_proto_goTypes = []interface{}{
+var file_types_number_proto_goTypes = []any{
 	(InvalidNumberBehaviour)(0), // 0: smartcore.types.InvalidNumberBehaviour
 	(*NumberCapping)(nil),       // 1: smartcore.types.NumberCapping
 	(*Int32Bounds)(nil),         // 2: smartcore.types.Int32Bounds
@@ -565,7 +565,7 @@ func file_types_number_proto_init() {
 	}
 	file_types_tween_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_types_number_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_types_number_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*NumberCapping); i {
 			case 0:
 				return &v.state
@@ -577,7 +577,7 @@ func file_types_number_proto_init() {
 				return nil
 			}
 		}
-		file_types_number_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_types_number_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Int32Bounds); i {
 			case 0:
 				return &v.state
@@ -589,7 +589,7 @@ func file_types_number_proto_init() {
 				return nil
 			}
 		}
-		file_types_number_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_types_number_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*Int32Attributes); i {
 			case 0:
 				return &v.state
@@ -601,7 +601,7 @@ func file_types_number_proto_init() {
 				return nil
 			}
 		}
-		file_types_number_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_types_number_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*FloatBounds); i {
 			case 0:
 				return &v.state
@@ -613,7 +613,7 @@ func file_types_number_proto_init() {
 				return nil
 			}
 		}
-		file_types_number_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_types_number_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*FloatAttributes); i {
 			case 0:
 				return &v.state
@@ -626,8 +626,8 @@ func file_types_number_proto_init() {
 			}
 		}
 	}
-	file_types_number_proto_msgTypes[1].OneofWrappers = []interface{}{}
-	file_types_number_proto_msgTypes[3].OneofWrappers = []interface{}{}
+	file_types_number_proto_msgTypes[1].OneofWrappers = []any{}
+	file_types_number_proto_msgTypes[3].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
