@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/page_token.proto
 
 package types
@@ -144,7 +144,7 @@ func file_types_page_token_proto_rawDescGZIP() []byte {
 }
 
 var file_types_page_token_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_types_page_token_proto_goTypes = []interface{}{
+var file_types_page_token_proto_goTypes = []any{
 	(*PageToken)(nil), // 0: smartcore.types.PageToken
 }
 var file_types_page_token_proto_depIdxs = []int32{
@@ -161,7 +161,7 @@ func file_types_page_token_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_types_page_token_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_types_page_token_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*PageToken); i {
 			case 0:
 				return &v.state
@@ -174,7 +174,7 @@ func file_types_page_token_proto_init() {
 			}
 		}
 	}
-	file_types_page_token_proto_msgTypes[0].OneofWrappers = []interface{}{
+	file_types_page_token_proto_msgTypes[0].OneofWrappers = []any{
 		(*PageToken_LastOffset)(nil),
 		(*PageToken_LastResourceName)(nil),
 	}
