@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/time/period.proto
 
 package time
@@ -118,7 +118,7 @@ func file_types_time_period_proto_rawDescGZIP() []byte {
 }
 
 var file_types_time_period_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_types_time_period_proto_goTypes = []interface{}{
+var file_types_time_period_proto_goTypes = []any{
 	(*Period)(nil),                // 0: smartcore.types.time.Period
 	(*timestamppb.Timestamp)(nil), // 1: google.protobuf.Timestamp
 }
@@ -138,7 +138,7 @@ func file_types_time_period_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_types_time_period_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_types_time_period_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Period); i {
 			case 0:
 				return &v.state
