@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/time/unit.proto
 
 package time
@@ -137,7 +137,7 @@ func file_types_time_unit_proto_rawDescGZIP() []byte {
 }
 
 var file_types_time_unit_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_types_time_unit_proto_goTypes = []interface{}{
+var file_types_time_unit_proto_goTypes = []any{
 	(Unit)(0), // 0: smartcore.types.time.Unit
 }
 var file_types_time_unit_proto_depIdxs = []int32{
