@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.2
 // source: types/unit.proto
 
 package types
@@ -384,7 +384,7 @@ func file_types_unit_proto_rawDescGZIP() []byte {
 
 var file_types_unit_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
 var file_types_unit_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
-var file_types_unit_proto_goTypes = []interface{}{
+var file_types_unit_proto_goTypes = []any{
 	(TemperatureUnit)(0),          // 0: smartcore.types.TemperatureUnit
 	(MuteSupport)(0),              // 1: smartcore.types.MuteSupport
 	(*Temperature)(nil),           // 2: smartcore.types.Temperature
@@ -411,7 +411,7 @@ func file_types_unit_proto_init() {
 	}
 	file_types_tween_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_types_unit_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_types_unit_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Temperature); i {
 			case 0:
 				return &v.state
@@ -423,7 +423,7 @@ func file_types_unit_proto_init() {
 				return nil
 			}
 		}
-		file_types_unit_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_types_unit_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*AudioLevel); i {
 			case 0:
 				return &v.state
@@ -435,7 +435,7 @@ func file_types_unit_proto_init() {
 				return nil
 			}
 		}
-		file_types_unit_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_types_unit_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*AudioLevelChange); i {
 			case 0:
 				return &v.state
