@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: info/auth.proto
 
 package info
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	AuthProvider_AddAccount_FullMethodName               = "/smartcore.info.AuthProvider/AddAccount"
@@ -28,6 +28,10 @@ const (
 // AuthProviderClient is the client API for AuthProvider service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// The auth provider service enables granular control of security permissions for a particular Smart Core controller or
+// system. It does this using the concept of 'Accounts', each of which has a certain set of permissions for which
+// device(s) and/or trait(s)/service(s) they are allowed to access.
 type AuthProviderClient interface {
 	AddAccount(ctx context.Context, in *AddAccountRequest, opts ...grpc.CallOption) (*AddAccountResponse, error)
 	RemoveAccount(ctx context.Context, in *RemoveAccountRequest, opts ...grpc.CallOption) (*RemoveAccountResponse, error)
@@ -44,8 +48,9 @@ func NewAuthProviderClient(cc grpc.ClientConnInterface) AuthProviderClient {
 }
 
 func (c *authProviderClient) AddAccount(ctx context.Context, in *AddAccountRequest, opts ...grpc.CallOption) (*AddAccountResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AddAccountResponse)
-	err := c.cc.Invoke(ctx, AuthProvider_AddAccount_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AuthProvider_AddAccount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -53,8 +58,9 @@ func (c *authProviderClient) AddAccount(ctx context.Context, in *AddAccountReque
 }
 
 func (c *authProviderClient) RemoveAccount(ctx context.Context, in *RemoveAccountRequest, opts ...grpc.CallOption) (*RemoveAccountResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(RemoveAccountResponse)
-	err := c.cc.Invoke(ctx, AuthProvider_RemoveAccount_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AuthProvider_RemoveAccount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -62,8 +68,9 @@ func (c *authProviderClient) RemoveAccount(ctx context.Context, in *RemoveAccoun
 }
 
 func (c *authProviderClient) UpdateAccountPermissions(ctx context.Context, in *UpdateAccountPermissionsRequest, opts ...grpc.CallOption) (*UpdateAccountPermissionsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(UpdateAccountPermissionsResponse)
-	err := c.cc.Invoke(ctx, AuthProvider_UpdateAccountPermissions_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AuthProvider_UpdateAccountPermissions_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -71,8 +78,9 @@ func (c *authProviderClient) UpdateAccountPermissions(ctx context.Context, in *U
 }
 
 func (c *authProviderClient) GenerateToken(ctx context.Context, in *GenerateTokenRequest, opts ...grpc.CallOption) (*GenerateTokenResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GenerateTokenResponse)
-	err := c.cc.Invoke(ctx, AuthProvider_GenerateToken_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AuthProvider_GenerateToken_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -81,7 +89,11 @@ func (c *authProviderClient) GenerateToken(ctx context.Context, in *GenerateToke
 
 // AuthProviderServer is the server API for AuthProvider service.
 // All implementations must embed UnimplementedAuthProviderServer
-// for forward compatibility
+// for forward compatibility.
+//
+// The auth provider service enables granular control of security permissions for a particular Smart Core controller or
+// system. It does this using the concept of 'Accounts', each of which has a certain set of permissions for which
+// device(s) and/or trait(s)/service(s) they are allowed to access.
 type AuthProviderServer interface {
 	AddAccount(context.Context, *AddAccountRequest) (*AddAccountResponse, error)
 	RemoveAccount(context.Context, *RemoveAccountRequest) (*RemoveAccountResponse, error)
@@ -90,9 +102,12 @@ type AuthProviderServer interface {
 	mustEmbedUnimplementedAuthProviderServer()
 }
 
-// UnimplementedAuthProviderServer must be embedded to have forward compatible implementations.
-type UnimplementedAuthProviderServer struct {
-}
+// UnimplementedAuthProviderServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedAuthProviderServer struct{}
 
 func (UnimplementedAuthProviderServer) AddAccount(context.Context, *AddAccountRequest) (*AddAccountResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method AddAccount not implemented")
@@ -107,6 +122,7 @@ func (UnimplementedAuthProviderServer) GenerateToken(context.Context, *GenerateT
 	return nil, status.Errorf(codes.Unimplemented, "method GenerateToken not implemented")
 }
 func (UnimplementedAuthProviderServer) mustEmbedUnimplementedAuthProviderServer() {}
+func (UnimplementedAuthProviderServer) testEmbeddedByValue()                      {}
 
 // UnsafeAuthProviderServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to AuthProviderServer will
@@ -116,6 +132,13 @@ type UnsafeAuthProviderServer interface {
 }
 
 func RegisterAuthProviderServer(s grpc.ServiceRegistrar, srv AuthProviderServer) {
+	// If the following call pancis, it indicates UnimplementedAuthProviderServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&AuthProvider_ServiceDesc, srv)
 }
 
