@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: info/health.proto
 
 package info
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	Health_GetHealthState_FullMethodName   = "/smartcore.info.Health/GetHealthState"
@@ -26,9 +26,11 @@ const (
 // HealthClient is the client API for Health service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describe the health of a device.
 type HealthClient interface {
 	GetHealthState(ctx context.Context, in *GetHealthStateRequest, opts ...grpc.CallOption) (*HealthState, error)
-	PullHealthStates(ctx context.Context, in *PullHealthStatesRequest, opts ...grpc.CallOption) (Health_PullHealthStatesClient, error)
+	PullHealthStates(ctx context.Context, in *PullHealthStatesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullHealthStatesResponse], error)
 }
 
 type healthClient struct {
@@ -40,20 +42,22 @@ func NewHealthClient(cc grpc.ClientConnInterface) HealthClient {
 }
 
 func (c *healthClient) GetHealthState(ctx context.Context, in *GetHealthStateRequest, opts ...grpc.CallOption) (*HealthState, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(HealthState)
-	err := c.cc.Invoke(ctx, Health_GetHealthState_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Health_GetHealthState_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *healthClient) PullHealthStates(ctx context.Context, in *PullHealthStatesRequest, opts ...grpc.CallOption) (Health_PullHealthStatesClient, error) {
-	stream, err := c.cc.NewStream(ctx, &Health_ServiceDesc.Streams[0], Health_PullHealthStates_FullMethodName, opts...)
+func (c *healthClient) PullHealthStates(ctx context.Context, in *PullHealthStatesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullHealthStatesResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &Health_ServiceDesc.Streams[0], Health_PullHealthStates_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &healthPullHealthStatesClient{stream}
+	x := &grpc.GenericClientStream[PullHealthStatesRequest, PullHealthStatesResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -63,43 +67,35 @@ func (c *healthClient) PullHealthStates(ctx context.Context, in *PullHealthState
 	return x, nil
 }
 
-type Health_PullHealthStatesClient interface {
-	Recv() (*PullHealthStatesResponse, error)
-	grpc.ClientStream
-}
-
-type healthPullHealthStatesClient struct {
-	grpc.ClientStream
-}
-
-func (x *healthPullHealthStatesClient) Recv() (*PullHealthStatesResponse, error) {
-	m := new(PullHealthStatesResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Health_PullHealthStatesClient = grpc.ServerStreamingClient[PullHealthStatesResponse]
 
 // HealthServer is the server API for Health service.
 // All implementations must embed UnimplementedHealthServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describe the health of a device.
 type HealthServer interface {
 	GetHealthState(context.Context, *GetHealthStateRequest) (*HealthState, error)
-	PullHealthStates(*PullHealthStatesRequest, Health_PullHealthStatesServer) error
+	PullHealthStates(*PullHealthStatesRequest, grpc.ServerStreamingServer[PullHealthStatesResponse]) error
 	mustEmbedUnimplementedHealthServer()
 }
 
-// UnimplementedHealthServer must be embedded to have forward compatible implementations.
-type UnimplementedHealthServer struct {
-}
+// UnimplementedHealthServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedHealthServer struct{}
 
 func (UnimplementedHealthServer) GetHealthState(context.Context, *GetHealthStateRequest) (*HealthState, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetHealthState not implemented")
 }
-func (UnimplementedHealthServer) PullHealthStates(*PullHealthStatesRequest, Health_PullHealthStatesServer) error {
+func (UnimplementedHealthServer) PullHealthStates(*PullHealthStatesRequest, grpc.ServerStreamingServer[PullHealthStatesResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullHealthStates not implemented")
 }
 func (UnimplementedHealthServer) mustEmbedUnimplementedHealthServer() {}
+func (UnimplementedHealthServer) testEmbeddedByValue()                {}
 
 // UnsafeHealthServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to HealthServer will
@@ -109,6 +105,13 @@ type UnsafeHealthServer interface {
 }
 
 func RegisterHealthServer(s grpc.ServiceRegistrar, srv HealthServer) {
+	// If the following call pancis, it indicates UnimplementedHealthServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Health_ServiceDesc, srv)
 }
 
@@ -135,21 +138,11 @@ func _Health_PullHealthStates_Handler(srv interface{}, stream grpc.ServerStream)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(HealthServer).PullHealthStates(m, &healthPullHealthStatesServer{stream})
+	return srv.(HealthServer).PullHealthStates(m, &grpc.GenericServerStream[PullHealthStatesRequest, PullHealthStatesResponse]{ServerStream: stream})
 }
 
-type Health_PullHealthStatesServer interface {
-	Send(*PullHealthStatesResponse) error
-	grpc.ServerStream
-}
-
-type healthPullHealthStatesServer struct {
-	grpc.ServerStream
-}
-
-func (x *healthPullHealthStatesServer) Send(m *PullHealthStatesResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Health_PullHealthStatesServer = grpc.ServerStreamingServer[PullHealthStatesResponse]
 
 // Health_ServiceDesc is the grpc.ServiceDesc for Health service.
 // It's only intended for direct use with grpc.RegisterService,
