@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: info/info.proto
 
 package info
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	Info_ListDevices_FullMethodName = "/smartcore.info.Info/ListDevices"
@@ -26,11 +26,15 @@ const (
 // InfoClient is the client API for Info service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Allow exposing information about an endpoint related to devices.
+//
+// Deprecated: use the Parent trait for listing devices, use Metadata trait for display name.
 type InfoClient interface {
 	// Get devices that this service knows about
 	ListDevices(ctx context.Context, in *ListDevicesRequest, opts ...grpc.CallOption) (*ListDevicesResponse, error)
 	// Open a server stream that responds with changes to the set of devices described in the request
-	PullDevices(ctx context.Context, in *PullDevicesRequest, opts ...grpc.CallOption) (Info_PullDevicesClient, error)
+	PullDevices(ctx context.Context, in *PullDevicesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullDevicesResponse], error)
 }
 
 type infoClient struct {
@@ -42,20 +46,22 @@ func NewInfoClient(cc grpc.ClientConnInterface) InfoClient {
 }
 
 func (c *infoClient) ListDevices(ctx context.Context, in *ListDevicesRequest, opts ...grpc.CallOption) (*ListDevicesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListDevicesResponse)
-	err := c.cc.Invoke(ctx, Info_ListDevices_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Info_ListDevices_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *infoClient) PullDevices(ctx context.Context, in *PullDevicesRequest, opts ...grpc.CallOption) (Info_PullDevicesClient, error) {
-	stream, err := c.cc.NewStream(ctx, &Info_ServiceDesc.Streams[0], Info_PullDevices_FullMethodName, opts...)
+func (c *infoClient) PullDevices(ctx context.Context, in *PullDevicesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullDevicesResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &Info_ServiceDesc.Streams[0], Info_PullDevices_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &infoPullDevicesClient{stream}
+	x := &grpc.GenericClientStream[PullDevicesRequest, PullDevicesResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -65,45 +71,39 @@ func (c *infoClient) PullDevices(ctx context.Context, in *PullDevicesRequest, op
 	return x, nil
 }
 
-type Info_PullDevicesClient interface {
-	Recv() (*PullDevicesResponse, error)
-	grpc.ClientStream
-}
-
-type infoPullDevicesClient struct {
-	grpc.ClientStream
-}
-
-func (x *infoPullDevicesClient) Recv() (*PullDevicesResponse, error) {
-	m := new(PullDevicesResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Info_PullDevicesClient = grpc.ServerStreamingClient[PullDevicesResponse]
 
 // InfoServer is the server API for Info service.
 // All implementations must embed UnimplementedInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Allow exposing information about an endpoint related to devices.
+//
+// Deprecated: use the Parent trait for listing devices, use Metadata trait for display name.
 type InfoServer interface {
 	// Get devices that this service knows about
 	ListDevices(context.Context, *ListDevicesRequest) (*ListDevicesResponse, error)
 	// Open a server stream that responds with changes to the set of devices described in the request
-	PullDevices(*PullDevicesRequest, Info_PullDevicesServer) error
+	PullDevices(*PullDevicesRequest, grpc.ServerStreamingServer[PullDevicesResponse]) error
 	mustEmbedUnimplementedInfoServer()
 }
 
-// UnimplementedInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedInfoServer struct {
-}
+// UnimplementedInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedInfoServer struct{}
 
 func (UnimplementedInfoServer) ListDevices(context.Context, *ListDevicesRequest) (*ListDevicesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListDevices not implemented")
 }
-func (UnimplementedInfoServer) PullDevices(*PullDevicesRequest, Info_PullDevicesServer) error {
+func (UnimplementedInfoServer) PullDevices(*PullDevicesRequest, grpc.ServerStreamingServer[PullDevicesResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullDevices not implemented")
 }
 func (UnimplementedInfoServer) mustEmbedUnimplementedInfoServer() {}
+func (UnimplementedInfoServer) testEmbeddedByValue()              {}
 
 // UnsafeInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to InfoServer will
@@ -113,6 +113,13 @@ type UnsafeInfoServer interface {
 }
 
 func RegisterInfoServer(s grpc.ServiceRegistrar, srv InfoServer) {
+	// If the following call pancis, it indicates UnimplementedInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Info_ServiceDesc, srv)
 }
 
@@ -139,21 +146,11 @@ func _Info_PullDevices_Handler(srv interface{}, stream grpc.ServerStream) error 
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(InfoServer).PullDevices(m, &infoPullDevicesServer{stream})
+	return srv.(InfoServer).PullDevices(m, &grpc.GenericServerStream[PullDevicesRequest, PullDevicesResponse]{ServerStream: stream})
 }
 
-type Info_PullDevicesServer interface {
-	Send(*PullDevicesResponse) error
-	grpc.ServerStream
-}
-
-type infoPullDevicesServer struct {
-	grpc.ServerStream
-}
-
-func (x *infoPullDevicesServer) Send(m *PullDevicesResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Info_PullDevicesServer = grpc.ServerStreamingServer[PullDevicesResponse]
 
 // Info_ServiceDesc is the grpc.ServiceDesc for Info service.
 // It's only intended for direct use with grpc.RegisterService,
