@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/air_quality_sensor.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	AirQualitySensorApi_GetAirQuality_FullMethodName  = "/smartcore.traits.AirQualitySensorApi/GetAirQuality"
@@ -26,11 +26,13 @@ const (
 // AirQualitySensorApiClient is the client API for AirQualitySensorApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Trait for devices that can report either numeric or descriptive readings about air quality and comfort.
 type AirQualitySensorApiClient interface {
 	// Get the current state of all reading this device supports
 	GetAirQuality(ctx context.Context, in *GetAirQualityRequest, opts ...grpc.CallOption) (*AirQuality, error)
 	// Be notified of changes to any reading
-	PullAirQuality(ctx context.Context, in *PullAirQualityRequest, opts ...grpc.CallOption) (AirQualitySensorApi_PullAirQualityClient, error)
+	PullAirQuality(ctx context.Context, in *PullAirQualityRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullAirQualityResponse], error)
 }
 
 type airQualitySensorApiClient struct {
@@ -42,20 +44,22 @@ func NewAirQualitySensorApiClient(cc grpc.ClientConnInterface) AirQualitySensorA
 }
 
 func (c *airQualitySensorApiClient) GetAirQuality(ctx context.Context, in *GetAirQualityRequest, opts ...grpc.CallOption) (*AirQuality, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AirQuality)
-	err := c.cc.Invoke(ctx, AirQualitySensorApi_GetAirQuality_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AirQualitySensorApi_GetAirQuality_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *airQualitySensorApiClient) PullAirQuality(ctx context.Context, in *PullAirQualityRequest, opts ...grpc.CallOption) (AirQualitySensorApi_PullAirQualityClient, error) {
-	stream, err := c.cc.NewStream(ctx, &AirQualitySensorApi_ServiceDesc.Streams[0], AirQualitySensorApi_PullAirQuality_FullMethodName, opts...)
+func (c *airQualitySensorApiClient) PullAirQuality(ctx context.Context, in *PullAirQualityRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullAirQualityResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &AirQualitySensorApi_ServiceDesc.Streams[0], AirQualitySensorApi_PullAirQuality_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &airQualitySensorApiPullAirQualityClient{stream}
+	x := &grpc.GenericClientStream[PullAirQualityRequest, PullAirQualityResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -65,45 +69,37 @@ func (c *airQualitySensorApiClient) PullAirQuality(ctx context.Context, in *Pull
 	return x, nil
 }
 
-type AirQualitySensorApi_PullAirQualityClient interface {
-	Recv() (*PullAirQualityResponse, error)
-	grpc.ClientStream
-}
-
-type airQualitySensorApiPullAirQualityClient struct {
-	grpc.ClientStream
-}
-
-func (x *airQualitySensorApiPullAirQualityClient) Recv() (*PullAirQualityResponse, error) {
-	m := new(PullAirQualityResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type AirQualitySensorApi_PullAirQualityClient = grpc.ServerStreamingClient[PullAirQualityResponse]
 
 // AirQualitySensorApiServer is the server API for AirQualitySensorApi service.
 // All implementations must embed UnimplementedAirQualitySensorApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Trait for devices that can report either numeric or descriptive readings about air quality and comfort.
 type AirQualitySensorApiServer interface {
 	// Get the current state of all reading this device supports
 	GetAirQuality(context.Context, *GetAirQualityRequest) (*AirQuality, error)
 	// Be notified of changes to any reading
-	PullAirQuality(*PullAirQualityRequest, AirQualitySensorApi_PullAirQualityServer) error
+	PullAirQuality(*PullAirQualityRequest, grpc.ServerStreamingServer[PullAirQualityResponse]) error
 	mustEmbedUnimplementedAirQualitySensorApiServer()
 }
 
-// UnimplementedAirQualitySensorApiServer must be embedded to have forward compatible implementations.
-type UnimplementedAirQualitySensorApiServer struct {
-}
+// UnimplementedAirQualitySensorApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedAirQualitySensorApiServer struct{}
 
 func (UnimplementedAirQualitySensorApiServer) GetAirQuality(context.Context, *GetAirQualityRequest) (*AirQuality, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetAirQuality not implemented")
 }
-func (UnimplementedAirQualitySensorApiServer) PullAirQuality(*PullAirQualityRequest, AirQualitySensorApi_PullAirQualityServer) error {
+func (UnimplementedAirQualitySensorApiServer) PullAirQuality(*PullAirQualityRequest, grpc.ServerStreamingServer[PullAirQualityResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullAirQuality not implemented")
 }
 func (UnimplementedAirQualitySensorApiServer) mustEmbedUnimplementedAirQualitySensorApiServer() {}
+func (UnimplementedAirQualitySensorApiServer) testEmbeddedByValue()                             {}
 
 // UnsafeAirQualitySensorApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to AirQualitySensorApiServer will
@@ -113,6 +109,13 @@ type UnsafeAirQualitySensorApiServer interface {
 }
 
 func RegisterAirQualitySensorApiServer(s grpc.ServiceRegistrar, srv AirQualitySensorApiServer) {
+	// If the following call pancis, it indicates UnimplementedAirQualitySensorApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&AirQualitySensorApi_ServiceDesc, srv)
 }
 
@@ -139,21 +142,11 @@ func _AirQualitySensorApi_PullAirQuality_Handler(srv interface{}, stream grpc.Se
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(AirQualitySensorApiServer).PullAirQuality(m, &airQualitySensorApiPullAirQualityServer{stream})
+	return srv.(AirQualitySensorApiServer).PullAirQuality(m, &grpc.GenericServerStream[PullAirQualityRequest, PullAirQualityResponse]{ServerStream: stream})
 }
 
-type AirQualitySensorApi_PullAirQualityServer interface {
-	Send(*PullAirQualityResponse) error
-	grpc.ServerStream
-}
-
-type airQualitySensorApiPullAirQualityServer struct {
-	grpc.ServerStream
-}
-
-func (x *airQualitySensorApiPullAirQualityServer) Send(m *PullAirQualityResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type AirQualitySensorApi_PullAirQualityServer = grpc.ServerStreamingServer[PullAirQualityResponse]
 
 // AirQualitySensorApi_ServiceDesc is the grpc.ServiceDesc for AirQualitySensorApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -184,6 +177,8 @@ const (
 // AirQualitySensorInfoClient is the client API for AirQualitySensorInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type AirQualitySensorInfoClient interface {
 	// Get information about how a named device implements AirQuality measurements
 	DescribeAirQuality(ctx context.Context, in *DescribeAirQualityRequest, opts ...grpc.CallOption) (*AirQualitySupport, error)
@@ -198,8 +193,9 @@ func NewAirQualitySensorInfoClient(cc grpc.ClientConnInterface) AirQualitySensor
 }
 
 func (c *airQualitySensorInfoClient) DescribeAirQuality(ctx context.Context, in *DescribeAirQualityRequest, opts ...grpc.CallOption) (*AirQualitySupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AirQualitySupport)
-	err := c.cc.Invoke(ctx, AirQualitySensorInfo_DescribeAirQuality_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AirQualitySensorInfo_DescribeAirQuality_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -208,21 +204,27 @@ func (c *airQualitySensorInfoClient) DescribeAirQuality(ctx context.Context, in 
 
 // AirQualitySensorInfoServer is the server API for AirQualitySensorInfo service.
 // All implementations must embed UnimplementedAirQualitySensorInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type AirQualitySensorInfoServer interface {
 	// Get information about how a named device implements AirQuality measurements
 	DescribeAirQuality(context.Context, *DescribeAirQualityRequest) (*AirQualitySupport, error)
 	mustEmbedUnimplementedAirQualitySensorInfoServer()
 }
 
-// UnimplementedAirQualitySensorInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedAirQualitySensorInfoServer struct {
-}
+// UnimplementedAirQualitySensorInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedAirQualitySensorInfoServer struct{}
 
 func (UnimplementedAirQualitySensorInfoServer) DescribeAirQuality(context.Context, *DescribeAirQualityRequest) (*AirQualitySupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeAirQuality not implemented")
 }
 func (UnimplementedAirQualitySensorInfoServer) mustEmbedUnimplementedAirQualitySensorInfoServer() {}
+func (UnimplementedAirQualitySensorInfoServer) testEmbeddedByValue()                              {}
 
 // UnsafeAirQualitySensorInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to AirQualitySensorInfoServer will
@@ -232,6 +234,13 @@ type UnsafeAirQualitySensorInfoServer interface {
 }
 
 func RegisterAirQualitySensorInfoServer(s grpc.ServiceRegistrar, srv AirQualitySensorInfoServer) {
+	// If the following call pancis, it indicates UnimplementedAirQualitySensorInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&AirQualitySensorInfo_ServiceDesc, srv)
 }
 
