@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/air_temperature.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	AirTemperatureApi_GetAirTemperature_FullMethodName    = "/smartcore.traits.AirTemperatureApi/GetAirTemperature"
@@ -27,6 +27,9 @@ const (
 // AirTemperatureApiClient is the client API for AirTemperatureApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Device trait allowing query and control of the air temperature for a space. Like a thermostat and associated HVAC
+// system.
 type AirTemperatureApiClient interface {
 	// Get the current state of the device
 	GetAirTemperature(ctx context.Context, in *GetAirTemperatureRequest, opts ...grpc.CallOption) (*AirTemperature, error)
@@ -36,7 +39,7 @@ type AirTemperatureApiClient interface {
 	// Request notification of change to the device state. The messages in the response stream may not be complete
 	// but will indicate the changes as they occur. They should be merged with the full state as fetched by the
 	// GetAirTemperature method.
-	PullAirTemperature(ctx context.Context, in *PullAirTemperatureRequest, opts ...grpc.CallOption) (AirTemperatureApi_PullAirTemperatureClient, error)
+	PullAirTemperature(ctx context.Context, in *PullAirTemperatureRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullAirTemperatureResponse], error)
 }
 
 type airTemperatureApiClient struct {
@@ -48,8 +51,9 @@ func NewAirTemperatureApiClient(cc grpc.ClientConnInterface) AirTemperatureApiCl
 }
 
 func (c *airTemperatureApiClient) GetAirTemperature(ctx context.Context, in *GetAirTemperatureRequest, opts ...grpc.CallOption) (*AirTemperature, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AirTemperature)
-	err := c.cc.Invoke(ctx, AirTemperatureApi_GetAirTemperature_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AirTemperatureApi_GetAirTemperature_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -57,20 +61,22 @@ func (c *airTemperatureApiClient) GetAirTemperature(ctx context.Context, in *Get
 }
 
 func (c *airTemperatureApiClient) UpdateAirTemperature(ctx context.Context, in *UpdateAirTemperatureRequest, opts ...grpc.CallOption) (*AirTemperature, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AirTemperature)
-	err := c.cc.Invoke(ctx, AirTemperatureApi_UpdateAirTemperature_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AirTemperatureApi_UpdateAirTemperature_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *airTemperatureApiClient) PullAirTemperature(ctx context.Context, in *PullAirTemperatureRequest, opts ...grpc.CallOption) (AirTemperatureApi_PullAirTemperatureClient, error) {
-	stream, err := c.cc.NewStream(ctx, &AirTemperatureApi_ServiceDesc.Streams[0], AirTemperatureApi_PullAirTemperature_FullMethodName, opts...)
+func (c *airTemperatureApiClient) PullAirTemperature(ctx context.Context, in *PullAirTemperatureRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullAirTemperatureResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &AirTemperatureApi_ServiceDesc.Streams[0], AirTemperatureApi_PullAirTemperature_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &airTemperatureApiPullAirTemperatureClient{stream}
+	x := &grpc.GenericClientStream[PullAirTemperatureRequest, PullAirTemperatureResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -80,26 +86,15 @@ func (c *airTemperatureApiClient) PullAirTemperature(ctx context.Context, in *Pu
 	return x, nil
 }
 
-type AirTemperatureApi_PullAirTemperatureClient interface {
-	Recv() (*PullAirTemperatureResponse, error)
-	grpc.ClientStream
-}
-
-type airTemperatureApiPullAirTemperatureClient struct {
-	grpc.ClientStream
-}
-
-func (x *airTemperatureApiPullAirTemperatureClient) Recv() (*PullAirTemperatureResponse, error) {
-	m := new(PullAirTemperatureResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type AirTemperatureApi_PullAirTemperatureClient = grpc.ServerStreamingClient[PullAirTemperatureResponse]
 
 // AirTemperatureApiServer is the server API for AirTemperatureApi service.
 // All implementations must embed UnimplementedAirTemperatureApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Device trait allowing query and control of the air temperature for a space. Like a thermostat and associated HVAC
+// system.
 type AirTemperatureApiServer interface {
 	// Get the current state of the device
 	GetAirTemperature(context.Context, *GetAirTemperatureRequest) (*AirTemperature, error)
@@ -109,13 +104,16 @@ type AirTemperatureApiServer interface {
 	// Request notification of change to the device state. The messages in the response stream may not be complete
 	// but will indicate the changes as they occur. They should be merged with the full state as fetched by the
 	// GetAirTemperature method.
-	PullAirTemperature(*PullAirTemperatureRequest, AirTemperatureApi_PullAirTemperatureServer) error
+	PullAirTemperature(*PullAirTemperatureRequest, grpc.ServerStreamingServer[PullAirTemperatureResponse]) error
 	mustEmbedUnimplementedAirTemperatureApiServer()
 }
 
-// UnimplementedAirTemperatureApiServer must be embedded to have forward compatible implementations.
-type UnimplementedAirTemperatureApiServer struct {
-}
+// UnimplementedAirTemperatureApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedAirTemperatureApiServer struct{}
 
 func (UnimplementedAirTemperatureApiServer) GetAirTemperature(context.Context, *GetAirTemperatureRequest) (*AirTemperature, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetAirTemperature not implemented")
@@ -123,10 +121,11 @@ func (UnimplementedAirTemperatureApiServer) GetAirTemperature(context.Context, *
 func (UnimplementedAirTemperatureApiServer) UpdateAirTemperature(context.Context, *UpdateAirTemperatureRequest) (*AirTemperature, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateAirTemperature not implemented")
 }
-func (UnimplementedAirTemperatureApiServer) PullAirTemperature(*PullAirTemperatureRequest, AirTemperatureApi_PullAirTemperatureServer) error {
+func (UnimplementedAirTemperatureApiServer) PullAirTemperature(*PullAirTemperatureRequest, grpc.ServerStreamingServer[PullAirTemperatureResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullAirTemperature not implemented")
 }
 func (UnimplementedAirTemperatureApiServer) mustEmbedUnimplementedAirTemperatureApiServer() {}
+func (UnimplementedAirTemperatureApiServer) testEmbeddedByValue()                           {}
 
 // UnsafeAirTemperatureApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to AirTemperatureApiServer will
@@ -136,6 +135,13 @@ type UnsafeAirTemperatureApiServer interface {
 }
 
 func RegisterAirTemperatureApiServer(s grpc.ServiceRegistrar, srv AirTemperatureApiServer) {
+	// If the following call pancis, it indicates UnimplementedAirTemperatureApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&AirTemperatureApi_ServiceDesc, srv)
 }
 
@@ -180,21 +186,11 @@ func _AirTemperatureApi_PullAirTemperature_Handler(srv interface{}, stream grpc.
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(AirTemperatureApiServer).PullAirTemperature(m, &airTemperatureApiPullAirTemperatureServer{stream})
+	return srv.(AirTemperatureApiServer).PullAirTemperature(m, &grpc.GenericServerStream[PullAirTemperatureRequest, PullAirTemperatureResponse]{ServerStream: stream})
 }
 
-type AirTemperatureApi_PullAirTemperatureServer interface {
-	Send(*PullAirTemperatureResponse) error
-	grpc.ServerStream
-}
-
-type airTemperatureApiPullAirTemperatureServer struct {
-	grpc.ServerStream
-}
-
-func (x *airTemperatureApiPullAirTemperatureServer) Send(m *PullAirTemperatureResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type AirTemperatureApi_PullAirTemperatureServer = grpc.ServerStreamingServer[PullAirTemperatureResponse]
 
 // AirTemperatureApi_ServiceDesc is the grpc.ServiceDesc for AirTemperatureApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -229,6 +225,8 @@ const (
 // AirTemperatureInfoClient is the client API for AirTemperatureInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type AirTemperatureInfoClient interface {
 	// Get information about how a named device implements AirTemperature features
 	DescribeAirTemperature(ctx context.Context, in *DescribeAirTemperatureRequest, opts ...grpc.CallOption) (*AirTemperatureSupport, error)
@@ -243,8 +241,9 @@ func NewAirTemperatureInfoClient(cc grpc.ClientConnInterface) AirTemperatureInfo
 }
 
 func (c *airTemperatureInfoClient) DescribeAirTemperature(ctx context.Context, in *DescribeAirTemperatureRequest, opts ...grpc.CallOption) (*AirTemperatureSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AirTemperatureSupport)
-	err := c.cc.Invoke(ctx, AirTemperatureInfo_DescribeAirTemperature_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AirTemperatureInfo_DescribeAirTemperature_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -253,21 +252,27 @@ func (c *airTemperatureInfoClient) DescribeAirTemperature(ctx context.Context, i
 
 // AirTemperatureInfoServer is the server API for AirTemperatureInfo service.
 // All implementations must embed UnimplementedAirTemperatureInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type AirTemperatureInfoServer interface {
 	// Get information about how a named device implements AirTemperature features
 	DescribeAirTemperature(context.Context, *DescribeAirTemperatureRequest) (*AirTemperatureSupport, error)
 	mustEmbedUnimplementedAirTemperatureInfoServer()
 }
 
-// UnimplementedAirTemperatureInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedAirTemperatureInfoServer struct {
-}
+// UnimplementedAirTemperatureInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedAirTemperatureInfoServer struct{}
 
 func (UnimplementedAirTemperatureInfoServer) DescribeAirTemperature(context.Context, *DescribeAirTemperatureRequest) (*AirTemperatureSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeAirTemperature not implemented")
 }
 func (UnimplementedAirTemperatureInfoServer) mustEmbedUnimplementedAirTemperatureInfoServer() {}
+func (UnimplementedAirTemperatureInfoServer) testEmbeddedByValue()                            {}
 
 // UnsafeAirTemperatureInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to AirTemperatureInfoServer will
@@ -277,6 +282,13 @@ type UnsafeAirTemperatureInfoServer interface {
 }
 
 func RegisterAirTemperatureInfoServer(s grpc.ServiceRegistrar, srv AirTemperatureInfoServer) {
+	// If the following call pancis, it indicates UnimplementedAirTemperatureInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&AirTemperatureInfo_ServiceDesc, srv)
 }
 
