@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/booking.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	BookingApi_ListBookings_FullMethodName    = "/smartcore.traits.BookingApi/ListBookings"
@@ -30,6 +30,9 @@ const (
 // BookingApiClient is the client API for BookingApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// The BookingApi defines an interface for interacting with bookable resources, for example rooms.
+// Bookings can be listed, edited, and created.
 type BookingApiClient interface {
 	// list bookings for a given bookable
 	ListBookings(ctx context.Context, in *ListBookingsRequest, opts ...grpc.CallOption) (*ListBookingsResponse, error)
@@ -42,7 +45,7 @@ type BookingApiClient interface {
 	// update an existing booking
 	UpdateBooking(ctx context.Context, in *UpdateBookingRequest, opts ...grpc.CallOption) (*UpdateBookingResponse, error)
 	// request updates to booking changes for a given bookable
-	PullBookings(ctx context.Context, in *ListBookingsRequest, opts ...grpc.CallOption) (BookingApi_PullBookingsClient, error)
+	PullBookings(ctx context.Context, in *ListBookingsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullBookingsResponse], error)
 }
 
 type bookingApiClient struct {
@@ -54,8 +57,9 @@ func NewBookingApiClient(cc grpc.ClientConnInterface) BookingApiClient {
 }
 
 func (c *bookingApiClient) ListBookings(ctx context.Context, in *ListBookingsRequest, opts ...grpc.CallOption) (*ListBookingsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListBookingsResponse)
-	err := c.cc.Invoke(ctx, BookingApi_ListBookings_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BookingApi_ListBookings_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -63,8 +67,9 @@ func (c *bookingApiClient) ListBookings(ctx context.Context, in *ListBookingsReq
 }
 
 func (c *bookingApiClient) CheckInBooking(ctx context.Context, in *CheckInBookingRequest, opts ...grpc.CallOption) (*CheckInBookingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CheckInBookingResponse)
-	err := c.cc.Invoke(ctx, BookingApi_CheckInBooking_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BookingApi_CheckInBooking_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -72,8 +77,9 @@ func (c *bookingApiClient) CheckInBooking(ctx context.Context, in *CheckInBookin
 }
 
 func (c *bookingApiClient) CheckOutBooking(ctx context.Context, in *CheckOutBookingRequest, opts ...grpc.CallOption) (*CheckOutBookingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CheckOutBookingResponse)
-	err := c.cc.Invoke(ctx, BookingApi_CheckOutBooking_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BookingApi_CheckOutBooking_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -81,8 +87,9 @@ func (c *bookingApiClient) CheckOutBooking(ctx context.Context, in *CheckOutBook
 }
 
 func (c *bookingApiClient) CreateBooking(ctx context.Context, in *CreateBookingRequest, opts ...grpc.CallOption) (*CreateBookingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreateBookingResponse)
-	err := c.cc.Invoke(ctx, BookingApi_CreateBooking_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BookingApi_CreateBooking_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -90,20 +97,22 @@ func (c *bookingApiClient) CreateBooking(ctx context.Context, in *CreateBookingR
 }
 
 func (c *bookingApiClient) UpdateBooking(ctx context.Context, in *UpdateBookingRequest, opts ...grpc.CallOption) (*UpdateBookingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(UpdateBookingResponse)
-	err := c.cc.Invoke(ctx, BookingApi_UpdateBooking_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BookingApi_UpdateBooking_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *bookingApiClient) PullBookings(ctx context.Context, in *ListBookingsRequest, opts ...grpc.CallOption) (BookingApi_PullBookingsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &BookingApi_ServiceDesc.Streams[0], BookingApi_PullBookings_FullMethodName, opts...)
+func (c *bookingApiClient) PullBookings(ctx context.Context, in *ListBookingsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullBookingsResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &BookingApi_ServiceDesc.Streams[0], BookingApi_PullBookings_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &bookingApiPullBookingsClient{stream}
+	x := &grpc.GenericClientStream[ListBookingsRequest, PullBookingsResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -113,26 +122,15 @@ func (c *bookingApiClient) PullBookings(ctx context.Context, in *ListBookingsReq
 	return x, nil
 }
 
-type BookingApi_PullBookingsClient interface {
-	Recv() (*PullBookingsResponse, error)
-	grpc.ClientStream
-}
-
-type bookingApiPullBookingsClient struct {
-	grpc.ClientStream
-}
-
-func (x *bookingApiPullBookingsClient) Recv() (*PullBookingsResponse, error) {
-	m := new(PullBookingsResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type BookingApi_PullBookingsClient = grpc.ServerStreamingClient[PullBookingsResponse]
 
 // BookingApiServer is the server API for BookingApi service.
 // All implementations must embed UnimplementedBookingApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// The BookingApi defines an interface for interacting with bookable resources, for example rooms.
+// Bookings can be listed, edited, and created.
 type BookingApiServer interface {
 	// list bookings for a given bookable
 	ListBookings(context.Context, *ListBookingsRequest) (*ListBookingsResponse, error)
@@ -145,13 +143,16 @@ type BookingApiServer interface {
 	// update an existing booking
 	UpdateBooking(context.Context, *UpdateBookingRequest) (*UpdateBookingResponse, error)
 	// request updates to booking changes for a given bookable
-	PullBookings(*ListBookingsRequest, BookingApi_PullBookingsServer) error
+	PullBookings(*ListBookingsRequest, grpc.ServerStreamingServer[PullBookingsResponse]) error
 	mustEmbedUnimplementedBookingApiServer()
 }
 
-// UnimplementedBookingApiServer must be embedded to have forward compatible implementations.
-type UnimplementedBookingApiServer struct {
-}
+// UnimplementedBookingApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedBookingApiServer struct{}
 
 func (UnimplementedBookingApiServer) ListBookings(context.Context, *ListBookingsRequest) (*ListBookingsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListBookings not implemented")
@@ -168,10 +169,11 @@ func (UnimplementedBookingApiServer) CreateBooking(context.Context, *CreateBooki
 func (UnimplementedBookingApiServer) UpdateBooking(context.Context, *UpdateBookingRequest) (*UpdateBookingResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateBooking not implemented")
 }
-func (UnimplementedBookingApiServer) PullBookings(*ListBookingsRequest, BookingApi_PullBookingsServer) error {
+func (UnimplementedBookingApiServer) PullBookings(*ListBookingsRequest, grpc.ServerStreamingServer[PullBookingsResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullBookings not implemented")
 }
 func (UnimplementedBookingApiServer) mustEmbedUnimplementedBookingApiServer() {}
+func (UnimplementedBookingApiServer) testEmbeddedByValue()                    {}
 
 // UnsafeBookingApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to BookingApiServer will
@@ -181,6 +183,13 @@ type UnsafeBookingApiServer interface {
 }
 
 func RegisterBookingApiServer(s grpc.ServiceRegistrar, srv BookingApiServer) {
+	// If the following call pancis, it indicates UnimplementedBookingApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&BookingApi_ServiceDesc, srv)
 }
 
@@ -279,21 +288,11 @@ func _BookingApi_PullBookings_Handler(srv interface{}, stream grpc.ServerStream)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(BookingApiServer).PullBookings(m, &bookingApiPullBookingsServer{stream})
+	return srv.(BookingApiServer).PullBookings(m, &grpc.GenericServerStream[ListBookingsRequest, PullBookingsResponse]{ServerStream: stream})
 }
 
-type BookingApi_PullBookingsServer interface {
-	Send(*PullBookingsResponse) error
-	grpc.ServerStream
-}
-
-type bookingApiPullBookingsServer struct {
-	grpc.ServerStream
-}
-
-func (x *bookingApiPullBookingsServer) Send(m *PullBookingsResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type BookingApi_PullBookingsServer = grpc.ServerStreamingServer[PullBookingsResponse]
 
 // BookingApi_ServiceDesc is the grpc.ServiceDesc for BookingApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -340,6 +339,8 @@ const (
 // BookingInfoClient is the client API for BookingInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type BookingInfoClient interface {
 	// Get information about how a named device implements Booking features
 	DescribeBooking(ctx context.Context, in *DescribeBookingRequest, opts ...grpc.CallOption) (*BookingSupport, error)
@@ -354,8 +355,9 @@ func NewBookingInfoClient(cc grpc.ClientConnInterface) BookingInfoClient {
 }
 
 func (c *bookingInfoClient) DescribeBooking(ctx context.Context, in *DescribeBookingRequest, opts ...grpc.CallOption) (*BookingSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(BookingSupport)
-	err := c.cc.Invoke(ctx, BookingInfo_DescribeBooking_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BookingInfo_DescribeBooking_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -364,21 +366,27 @@ func (c *bookingInfoClient) DescribeBooking(ctx context.Context, in *DescribeBoo
 
 // BookingInfoServer is the server API for BookingInfo service.
 // All implementations must embed UnimplementedBookingInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type BookingInfoServer interface {
 	// Get information about how a named device implements Booking features
 	DescribeBooking(context.Context, *DescribeBookingRequest) (*BookingSupport, error)
 	mustEmbedUnimplementedBookingInfoServer()
 }
 
-// UnimplementedBookingInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedBookingInfoServer struct {
-}
+// UnimplementedBookingInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedBookingInfoServer struct{}
 
 func (UnimplementedBookingInfoServer) DescribeBooking(context.Context, *DescribeBookingRequest) (*BookingSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeBooking not implemented")
 }
 func (UnimplementedBookingInfoServer) mustEmbedUnimplementedBookingInfoServer() {}
+func (UnimplementedBookingInfoServer) testEmbeddedByValue()                     {}
 
 // UnsafeBookingInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to BookingInfoServer will
@@ -388,6 +396,13 @@ type UnsafeBookingInfoServer interface {
 }
 
 func RegisterBookingInfoServer(s grpc.ServiceRegistrar, srv BookingInfoServer) {
+	// If the following call pancis, it indicates UnimplementedBookingInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&BookingInfo_ServiceDesc, srv)
 }
 
