@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/brightness_sensor.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	BrightnessSensorApi_GetAmbientBrightness_FullMethodName  = "/smartcore.traits.BrightnessSensorApi/GetAmbientBrightness"
@@ -26,11 +26,13 @@ const (
 // BrightnessSensorApiClient is the client API for BrightnessSensorApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// BrightnessSensorApi reports on the measured ambient brightness for a space.
 type BrightnessSensorApiClient interface {
 	// GetAmbientBrightness returns the current brightness level as measured by the device.
 	GetAmbientBrightness(ctx context.Context, in *GetAmbientBrightnessRequest, opts ...grpc.CallOption) (*AmbientBrightness, error)
 	// PullAmbientBrightness returns a stream of brightness changes as the device measures them.
-	PullAmbientBrightness(ctx context.Context, in *PullAmbientBrightnessRequest, opts ...grpc.CallOption) (BrightnessSensorApi_PullAmbientBrightnessClient, error)
+	PullAmbientBrightness(ctx context.Context, in *PullAmbientBrightnessRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullAmbientBrightnessResponse], error)
 }
 
 type brightnessSensorApiClient struct {
@@ -42,20 +44,22 @@ func NewBrightnessSensorApiClient(cc grpc.ClientConnInterface) BrightnessSensorA
 }
 
 func (c *brightnessSensorApiClient) GetAmbientBrightness(ctx context.Context, in *GetAmbientBrightnessRequest, opts ...grpc.CallOption) (*AmbientBrightness, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AmbientBrightness)
-	err := c.cc.Invoke(ctx, BrightnessSensorApi_GetAmbientBrightness_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BrightnessSensorApi_GetAmbientBrightness_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *brightnessSensorApiClient) PullAmbientBrightness(ctx context.Context, in *PullAmbientBrightnessRequest, opts ...grpc.CallOption) (BrightnessSensorApi_PullAmbientBrightnessClient, error) {
-	stream, err := c.cc.NewStream(ctx, &BrightnessSensorApi_ServiceDesc.Streams[0], BrightnessSensorApi_PullAmbientBrightness_FullMethodName, opts...)
+func (c *brightnessSensorApiClient) PullAmbientBrightness(ctx context.Context, in *PullAmbientBrightnessRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullAmbientBrightnessResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &BrightnessSensorApi_ServiceDesc.Streams[0], BrightnessSensorApi_PullAmbientBrightness_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &brightnessSensorApiPullAmbientBrightnessClient{stream}
+	x := &grpc.GenericClientStream[PullAmbientBrightnessRequest, PullAmbientBrightnessResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -65,45 +69,37 @@ func (c *brightnessSensorApiClient) PullAmbientBrightness(ctx context.Context, i
 	return x, nil
 }
 
-type BrightnessSensorApi_PullAmbientBrightnessClient interface {
-	Recv() (*PullAmbientBrightnessResponse, error)
-	grpc.ClientStream
-}
-
-type brightnessSensorApiPullAmbientBrightnessClient struct {
-	grpc.ClientStream
-}
-
-func (x *brightnessSensorApiPullAmbientBrightnessClient) Recv() (*PullAmbientBrightnessResponse, error) {
-	m := new(PullAmbientBrightnessResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type BrightnessSensorApi_PullAmbientBrightnessClient = grpc.ServerStreamingClient[PullAmbientBrightnessResponse]
 
 // BrightnessSensorApiServer is the server API for BrightnessSensorApi service.
 // All implementations must embed UnimplementedBrightnessSensorApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// BrightnessSensorApi reports on the measured ambient brightness for a space.
 type BrightnessSensorApiServer interface {
 	// GetAmbientBrightness returns the current brightness level as measured by the device.
 	GetAmbientBrightness(context.Context, *GetAmbientBrightnessRequest) (*AmbientBrightness, error)
 	// PullAmbientBrightness returns a stream of brightness changes as the device measures them.
-	PullAmbientBrightness(*PullAmbientBrightnessRequest, BrightnessSensorApi_PullAmbientBrightnessServer) error
+	PullAmbientBrightness(*PullAmbientBrightnessRequest, grpc.ServerStreamingServer[PullAmbientBrightnessResponse]) error
 	mustEmbedUnimplementedBrightnessSensorApiServer()
 }
 
-// UnimplementedBrightnessSensorApiServer must be embedded to have forward compatible implementations.
-type UnimplementedBrightnessSensorApiServer struct {
-}
+// UnimplementedBrightnessSensorApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedBrightnessSensorApiServer struct{}
 
 func (UnimplementedBrightnessSensorApiServer) GetAmbientBrightness(context.Context, *GetAmbientBrightnessRequest) (*AmbientBrightness, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetAmbientBrightness not implemented")
 }
-func (UnimplementedBrightnessSensorApiServer) PullAmbientBrightness(*PullAmbientBrightnessRequest, BrightnessSensorApi_PullAmbientBrightnessServer) error {
+func (UnimplementedBrightnessSensorApiServer) PullAmbientBrightness(*PullAmbientBrightnessRequest, grpc.ServerStreamingServer[PullAmbientBrightnessResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullAmbientBrightness not implemented")
 }
 func (UnimplementedBrightnessSensorApiServer) mustEmbedUnimplementedBrightnessSensorApiServer() {}
+func (UnimplementedBrightnessSensorApiServer) testEmbeddedByValue()                             {}
 
 // UnsafeBrightnessSensorApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to BrightnessSensorApiServer will
@@ -113,6 +109,13 @@ type UnsafeBrightnessSensorApiServer interface {
 }
 
 func RegisterBrightnessSensorApiServer(s grpc.ServiceRegistrar, srv BrightnessSensorApiServer) {
+	// If the following call pancis, it indicates UnimplementedBrightnessSensorApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&BrightnessSensorApi_ServiceDesc, srv)
 }
 
@@ -139,21 +142,11 @@ func _BrightnessSensorApi_PullAmbientBrightness_Handler(srv interface{}, stream 
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(BrightnessSensorApiServer).PullAmbientBrightness(m, &brightnessSensorApiPullAmbientBrightnessServer{stream})
+	return srv.(BrightnessSensorApiServer).PullAmbientBrightness(m, &grpc.GenericServerStream[PullAmbientBrightnessRequest, PullAmbientBrightnessResponse]{ServerStream: stream})
 }
 
-type BrightnessSensorApi_PullAmbientBrightnessServer interface {
-	Send(*PullAmbientBrightnessResponse) error
-	grpc.ServerStream
-}
-
-type brightnessSensorApiPullAmbientBrightnessServer struct {
-	grpc.ServerStream
-}
-
-func (x *brightnessSensorApiPullAmbientBrightnessServer) Send(m *PullAmbientBrightnessResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type BrightnessSensorApi_PullAmbientBrightnessServer = grpc.ServerStreamingServer[PullAmbientBrightnessResponse]
 
 // BrightnessSensorApi_ServiceDesc is the grpc.ServiceDesc for BrightnessSensorApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -184,6 +177,8 @@ const (
 // BrightnessSensorInfoClient is the client API for BrightnessSensorInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type BrightnessSensorInfoClient interface {
 	// Get information about how a named device implements AmbientBrightness measurements
 	DescribeAmbientBrightness(ctx context.Context, in *DescribeAmbientBrightnessRequest, opts ...grpc.CallOption) (*AmbientBrightnessSupport, error)
@@ -198,8 +193,9 @@ func NewBrightnessSensorInfoClient(cc grpc.ClientConnInterface) BrightnessSensor
 }
 
 func (c *brightnessSensorInfoClient) DescribeAmbientBrightness(ctx context.Context, in *DescribeAmbientBrightnessRequest, opts ...grpc.CallOption) (*AmbientBrightnessSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AmbientBrightnessSupport)
-	err := c.cc.Invoke(ctx, BrightnessSensorInfo_DescribeAmbientBrightness_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BrightnessSensorInfo_DescribeAmbientBrightness_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -208,21 +204,27 @@ func (c *brightnessSensorInfoClient) DescribeAmbientBrightness(ctx context.Conte
 
 // BrightnessSensorInfoServer is the server API for BrightnessSensorInfo service.
 // All implementations must embed UnimplementedBrightnessSensorInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type BrightnessSensorInfoServer interface {
 	// Get information about how a named device implements AmbientBrightness measurements
 	DescribeAmbientBrightness(context.Context, *DescribeAmbientBrightnessRequest) (*AmbientBrightnessSupport, error)
 	mustEmbedUnimplementedBrightnessSensorInfoServer()
 }
 
-// UnimplementedBrightnessSensorInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedBrightnessSensorInfoServer struct {
-}
+// UnimplementedBrightnessSensorInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedBrightnessSensorInfoServer struct{}
 
 func (UnimplementedBrightnessSensorInfoServer) DescribeAmbientBrightness(context.Context, *DescribeAmbientBrightnessRequest) (*AmbientBrightnessSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeAmbientBrightness not implemented")
 }
 func (UnimplementedBrightnessSensorInfoServer) mustEmbedUnimplementedBrightnessSensorInfoServer() {}
+func (UnimplementedBrightnessSensorInfoServer) testEmbeddedByValue()                              {}
 
 // UnsafeBrightnessSensorInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to BrightnessSensorInfoServer will
@@ -232,6 +234,13 @@ type UnsafeBrightnessSensorInfoServer interface {
 }
 
 func RegisterBrightnessSensorInfoServer(s grpc.ServiceRegistrar, srv BrightnessSensorInfoServer) {
+	// If the following call pancis, it indicates UnimplementedBrightnessSensorInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&BrightnessSensorInfo_ServiceDesc, srv)
 }
 
