@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/channel.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	ChannelApi_GetChosenChannel_FullMethodName  = "/smartcore.traits.ChannelApi/GetChosenChannel"
@@ -29,6 +29,8 @@ const (
 // ChannelApiClient is the client API for ChannelApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// ChannelApi allows for querying and updating the chosen channel on a device like a TV.
 type ChannelApiClient interface {
 	// GetChosenChannel returns the currently chosen channel for the device.
 	GetChosenChannel(ctx context.Context, in *GetChosenChannelRequest, opts ...grpc.CallOption) (*Channel, error)
@@ -39,7 +41,7 @@ type ChannelApiClient interface {
 	// ReturnChannel will choose the previously chosen channel. Returns the chosen channel
 	ReturnChannel(ctx context.Context, in *ReturnChannelRequest, opts ...grpc.CallOption) (*Channel, error)
 	// PullChosenChannel subscribes to chosen channel changes
-	PullChosenChannel(ctx context.Context, in *PullChosenChannelRequest, opts ...grpc.CallOption) (ChannelApi_PullChosenChannelClient, error)
+	PullChosenChannel(ctx context.Context, in *PullChosenChannelRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullChosenChannelResponse], error)
 }
 
 type channelApiClient struct {
@@ -51,8 +53,9 @@ func NewChannelApiClient(cc grpc.ClientConnInterface) ChannelApiClient {
 }
 
 func (c *channelApiClient) GetChosenChannel(ctx context.Context, in *GetChosenChannelRequest, opts ...grpc.CallOption) (*Channel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Channel)
-	err := c.cc.Invoke(ctx, ChannelApi_GetChosenChannel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ChannelApi_GetChosenChannel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -60,8 +63,9 @@ func (c *channelApiClient) GetChosenChannel(ctx context.Context, in *GetChosenCh
 }
 
 func (c *channelApiClient) ChooseChannel(ctx context.Context, in *ChooseChannelRequest, opts ...grpc.CallOption) (*Channel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Channel)
-	err := c.cc.Invoke(ctx, ChannelApi_ChooseChannel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ChannelApi_ChooseChannel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -69,8 +73,9 @@ func (c *channelApiClient) ChooseChannel(ctx context.Context, in *ChooseChannelR
 }
 
 func (c *channelApiClient) AdjustChannel(ctx context.Context, in *AdjustChannelRequest, opts ...grpc.CallOption) (*Channel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Channel)
-	err := c.cc.Invoke(ctx, ChannelApi_AdjustChannel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ChannelApi_AdjustChannel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -78,20 +83,22 @@ func (c *channelApiClient) AdjustChannel(ctx context.Context, in *AdjustChannelR
 }
 
 func (c *channelApiClient) ReturnChannel(ctx context.Context, in *ReturnChannelRequest, opts ...grpc.CallOption) (*Channel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Channel)
-	err := c.cc.Invoke(ctx, ChannelApi_ReturnChannel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ChannelApi_ReturnChannel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *channelApiClient) PullChosenChannel(ctx context.Context, in *PullChosenChannelRequest, opts ...grpc.CallOption) (ChannelApi_PullChosenChannelClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ChannelApi_ServiceDesc.Streams[0], ChannelApi_PullChosenChannel_FullMethodName, opts...)
+func (c *channelApiClient) PullChosenChannel(ctx context.Context, in *PullChosenChannelRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullChosenChannelResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ChannelApi_ServiceDesc.Streams[0], ChannelApi_PullChosenChannel_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &channelApiPullChosenChannelClient{stream}
+	x := &grpc.GenericClientStream[PullChosenChannelRequest, PullChosenChannelResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -101,26 +108,14 @@ func (c *channelApiClient) PullChosenChannel(ctx context.Context, in *PullChosen
 	return x, nil
 }
 
-type ChannelApi_PullChosenChannelClient interface {
-	Recv() (*PullChosenChannelResponse, error)
-	grpc.ClientStream
-}
-
-type channelApiPullChosenChannelClient struct {
-	grpc.ClientStream
-}
-
-func (x *channelApiPullChosenChannelClient) Recv() (*PullChosenChannelResponse, error) {
-	m := new(PullChosenChannelResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ChannelApi_PullChosenChannelClient = grpc.ServerStreamingClient[PullChosenChannelResponse]
 
 // ChannelApiServer is the server API for ChannelApi service.
 // All implementations must embed UnimplementedChannelApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// ChannelApi allows for querying and updating the chosen channel on a device like a TV.
 type ChannelApiServer interface {
 	// GetChosenChannel returns the currently chosen channel for the device.
 	GetChosenChannel(context.Context, *GetChosenChannelRequest) (*Channel, error)
@@ -131,13 +126,16 @@ type ChannelApiServer interface {
 	// ReturnChannel will choose the previously chosen channel. Returns the chosen channel
 	ReturnChannel(context.Context, *ReturnChannelRequest) (*Channel, error)
 	// PullChosenChannel subscribes to chosen channel changes
-	PullChosenChannel(*PullChosenChannelRequest, ChannelApi_PullChosenChannelServer) error
+	PullChosenChannel(*PullChosenChannelRequest, grpc.ServerStreamingServer[PullChosenChannelResponse]) error
 	mustEmbedUnimplementedChannelApiServer()
 }
 
-// UnimplementedChannelApiServer must be embedded to have forward compatible implementations.
-type UnimplementedChannelApiServer struct {
-}
+// UnimplementedChannelApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedChannelApiServer struct{}
 
 func (UnimplementedChannelApiServer) GetChosenChannel(context.Context, *GetChosenChannelRequest) (*Channel, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetChosenChannel not implemented")
@@ -151,10 +149,11 @@ func (UnimplementedChannelApiServer) AdjustChannel(context.Context, *AdjustChann
 func (UnimplementedChannelApiServer) ReturnChannel(context.Context, *ReturnChannelRequest) (*Channel, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ReturnChannel not implemented")
 }
-func (UnimplementedChannelApiServer) PullChosenChannel(*PullChosenChannelRequest, ChannelApi_PullChosenChannelServer) error {
+func (UnimplementedChannelApiServer) PullChosenChannel(*PullChosenChannelRequest, grpc.ServerStreamingServer[PullChosenChannelResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullChosenChannel not implemented")
 }
 func (UnimplementedChannelApiServer) mustEmbedUnimplementedChannelApiServer() {}
+func (UnimplementedChannelApiServer) testEmbeddedByValue()                    {}
 
 // UnsafeChannelApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ChannelApiServer will
@@ -164,6 +163,13 @@ type UnsafeChannelApiServer interface {
 }
 
 func RegisterChannelApiServer(s grpc.ServiceRegistrar, srv ChannelApiServer) {
+	// If the following call pancis, it indicates UnimplementedChannelApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ChannelApi_ServiceDesc, srv)
 }
 
@@ -244,21 +250,11 @@ func _ChannelApi_PullChosenChannel_Handler(srv interface{}, stream grpc.ServerSt
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ChannelApiServer).PullChosenChannel(m, &channelApiPullChosenChannelServer{stream})
+	return srv.(ChannelApiServer).PullChosenChannel(m, &grpc.GenericServerStream[PullChosenChannelRequest, PullChosenChannelResponse]{ServerStream: stream})
 }
 
-type ChannelApi_PullChosenChannelServer interface {
-	Send(*PullChosenChannelResponse) error
-	grpc.ServerStream
-}
-
-type channelApiPullChosenChannelServer struct {
-	grpc.ServerStream
-}
-
-func (x *channelApiPullChosenChannelServer) Send(m *PullChosenChannelResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ChannelApi_PullChosenChannelServer = grpc.ServerStreamingServer[PullChosenChannelResponse]
 
 // ChannelApi_ServiceDesc is the grpc.ServiceDesc for ChannelApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -301,6 +297,8 @@ const (
 // ChannelInfoClient is the client API for ChannelInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ChannelInfoClient interface {
 	// Get information about how a named device implements channel choosing and adjustment
 	DescribeChosenChannel(ctx context.Context, in *DescribeChosenChannelRequest, opts ...grpc.CallOption) (*ChosenChannelSupport, error)
@@ -315,8 +313,9 @@ func NewChannelInfoClient(cc grpc.ClientConnInterface) ChannelInfoClient {
 }
 
 func (c *channelInfoClient) DescribeChosenChannel(ctx context.Context, in *DescribeChosenChannelRequest, opts ...grpc.CallOption) (*ChosenChannelSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ChosenChannelSupport)
-	err := c.cc.Invoke(ctx, ChannelInfo_DescribeChosenChannel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ChannelInfo_DescribeChosenChannel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -325,21 +324,27 @@ func (c *channelInfoClient) DescribeChosenChannel(ctx context.Context, in *Descr
 
 // ChannelInfoServer is the server API for ChannelInfo service.
 // All implementations must embed UnimplementedChannelInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ChannelInfoServer interface {
 	// Get information about how a named device implements channel choosing and adjustment
 	DescribeChosenChannel(context.Context, *DescribeChosenChannelRequest) (*ChosenChannelSupport, error)
 	mustEmbedUnimplementedChannelInfoServer()
 }
 
-// UnimplementedChannelInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedChannelInfoServer struct {
-}
+// UnimplementedChannelInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedChannelInfoServer struct{}
 
 func (UnimplementedChannelInfoServer) DescribeChosenChannel(context.Context, *DescribeChosenChannelRequest) (*ChosenChannelSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeChosenChannel not implemented")
 }
 func (UnimplementedChannelInfoServer) mustEmbedUnimplementedChannelInfoServer() {}
+func (UnimplementedChannelInfoServer) testEmbeddedByValue()                     {}
 
 // UnsafeChannelInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ChannelInfoServer will
@@ -349,6 +354,13 @@ type UnsafeChannelInfoServer interface {
 }
 
 func RegisterChannelInfoServer(s grpc.ServiceRegistrar, srv ChannelInfoServer) {
+	// If the following call pancis, it indicates UnimplementedChannelInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ChannelInfo_ServiceDesc, srv)
 }
 
