@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/count.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	CountApi_GetCount_FullMethodName    = "/smartcore.traits.CountApi/GetCount"
@@ -28,6 +28,9 @@ const (
 // CountApiClient is the client API for CountApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// CountApi tracks both a positive and negative count of some quantity. Can be used for tickers, footfall sensors, and
+// devices like that.
 type CountApiClient interface {
 	// Get the current count state
 	GetCount(ctx context.Context, in *GetCountRequest, opts ...grpc.CallOption) (*Count, error)
@@ -36,7 +39,7 @@ type CountApiClient interface {
 	// Update one or more properties of the count.
 	UpdateCount(ctx context.Context, in *UpdateCountRequest, opts ...grpc.CallOption) (*Count, error)
 	// Subscribe to changes to the count.
-	PullCounts(ctx context.Context, in *PullCountsRequest, opts ...grpc.CallOption) (CountApi_PullCountsClient, error)
+	PullCounts(ctx context.Context, in *PullCountsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullCountsResponse], error)
 }
 
 type countApiClient struct {
@@ -48,8 +51,9 @@ func NewCountApiClient(cc grpc.ClientConnInterface) CountApiClient {
 }
 
 func (c *countApiClient) GetCount(ctx context.Context, in *GetCountRequest, opts ...grpc.CallOption) (*Count, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Count)
-	err := c.cc.Invoke(ctx, CountApi_GetCount_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CountApi_GetCount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -57,8 +61,9 @@ func (c *countApiClient) GetCount(ctx context.Context, in *GetCountRequest, opts
 }
 
 func (c *countApiClient) ResetCount(ctx context.Context, in *ResetCountRequest, opts ...grpc.CallOption) (*Count, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Count)
-	err := c.cc.Invoke(ctx, CountApi_ResetCount_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CountApi_ResetCount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -66,20 +71,22 @@ func (c *countApiClient) ResetCount(ctx context.Context, in *ResetCountRequest, 
 }
 
 func (c *countApiClient) UpdateCount(ctx context.Context, in *UpdateCountRequest, opts ...grpc.CallOption) (*Count, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Count)
-	err := c.cc.Invoke(ctx, CountApi_UpdateCount_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CountApi_UpdateCount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *countApiClient) PullCounts(ctx context.Context, in *PullCountsRequest, opts ...grpc.CallOption) (CountApi_PullCountsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CountApi_ServiceDesc.Streams[0], CountApi_PullCounts_FullMethodName, opts...)
+func (c *countApiClient) PullCounts(ctx context.Context, in *PullCountsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullCountsResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CountApi_ServiceDesc.Streams[0], CountApi_PullCounts_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &countApiPullCountsClient{stream}
+	x := &grpc.GenericClientStream[PullCountsRequest, PullCountsResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -89,26 +96,15 @@ func (c *countApiClient) PullCounts(ctx context.Context, in *PullCountsRequest, 
 	return x, nil
 }
 
-type CountApi_PullCountsClient interface {
-	Recv() (*PullCountsResponse, error)
-	grpc.ClientStream
-}
-
-type countApiPullCountsClient struct {
-	grpc.ClientStream
-}
-
-func (x *countApiPullCountsClient) Recv() (*PullCountsResponse, error) {
-	m := new(PullCountsResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CountApi_PullCountsClient = grpc.ServerStreamingClient[PullCountsResponse]
 
 // CountApiServer is the server API for CountApi service.
 // All implementations must embed UnimplementedCountApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// CountApi tracks both a positive and negative count of some quantity. Can be used for tickers, footfall sensors, and
+// devices like that.
 type CountApiServer interface {
 	// Get the current count state
 	GetCount(context.Context, *GetCountRequest) (*Count, error)
@@ -117,13 +113,16 @@ type CountApiServer interface {
 	// Update one or more properties of the count.
 	UpdateCount(context.Context, *UpdateCountRequest) (*Count, error)
 	// Subscribe to changes to the count.
-	PullCounts(*PullCountsRequest, CountApi_PullCountsServer) error
+	PullCounts(*PullCountsRequest, grpc.ServerStreamingServer[PullCountsResponse]) error
 	mustEmbedUnimplementedCountApiServer()
 }
 
-// UnimplementedCountApiServer must be embedded to have forward compatible implementations.
-type UnimplementedCountApiServer struct {
-}
+// UnimplementedCountApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedCountApiServer struct{}
 
 func (UnimplementedCountApiServer) GetCount(context.Context, *GetCountRequest) (*Count, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetCount not implemented")
@@ -134,10 +133,11 @@ func (UnimplementedCountApiServer) ResetCount(context.Context, *ResetCountReques
 func (UnimplementedCountApiServer) UpdateCount(context.Context, *UpdateCountRequest) (*Count, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateCount not implemented")
 }
-func (UnimplementedCountApiServer) PullCounts(*PullCountsRequest, CountApi_PullCountsServer) error {
+func (UnimplementedCountApiServer) PullCounts(*PullCountsRequest, grpc.ServerStreamingServer[PullCountsResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullCounts not implemented")
 }
 func (UnimplementedCountApiServer) mustEmbedUnimplementedCountApiServer() {}
+func (UnimplementedCountApiServer) testEmbeddedByValue()                  {}
 
 // UnsafeCountApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to CountApiServer will
@@ -147,6 +147,13 @@ type UnsafeCountApiServer interface {
 }
 
 func RegisterCountApiServer(s grpc.ServiceRegistrar, srv CountApiServer) {
+	// If the following call pancis, it indicates UnimplementedCountApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&CountApi_ServiceDesc, srv)
 }
 
@@ -209,21 +216,11 @@ func _CountApi_PullCounts_Handler(srv interface{}, stream grpc.ServerStream) err
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CountApiServer).PullCounts(m, &countApiPullCountsServer{stream})
+	return srv.(CountApiServer).PullCounts(m, &grpc.GenericServerStream[PullCountsRequest, PullCountsResponse]{ServerStream: stream})
 }
 
-type CountApi_PullCountsServer interface {
-	Send(*PullCountsResponse) error
-	grpc.ServerStream
-}
-
-type countApiPullCountsServer struct {
-	grpc.ServerStream
-}
-
-func (x *countApiPullCountsServer) Send(m *PullCountsResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CountApi_PullCountsServer = grpc.ServerStreamingServer[PullCountsResponse]
 
 // CountApi_ServiceDesc is the grpc.ServiceDesc for CountApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -262,6 +259,8 @@ const (
 // CountInfoClient is the client API for CountInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type CountInfoClient interface {
 	// Get information about how a named device implements Count features
 	DescribeCount(ctx context.Context, in *DescribeCountRequest, opts ...grpc.CallOption) (*CountSupport, error)
@@ -276,8 +275,9 @@ func NewCountInfoClient(cc grpc.ClientConnInterface) CountInfoClient {
 }
 
 func (c *countInfoClient) DescribeCount(ctx context.Context, in *DescribeCountRequest, opts ...grpc.CallOption) (*CountSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CountSupport)
-	err := c.cc.Invoke(ctx, CountInfo_DescribeCount_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CountInfo_DescribeCount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -286,21 +286,27 @@ func (c *countInfoClient) DescribeCount(ctx context.Context, in *DescribeCountRe
 
 // CountInfoServer is the server API for CountInfo service.
 // All implementations must embed UnimplementedCountInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type CountInfoServer interface {
 	// Get information about how a named device implements Count features
 	DescribeCount(context.Context, *DescribeCountRequest) (*CountSupport, error)
 	mustEmbedUnimplementedCountInfoServer()
 }
 
-// UnimplementedCountInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedCountInfoServer struct {
-}
+// UnimplementedCountInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedCountInfoServer struct{}
 
 func (UnimplementedCountInfoServer) DescribeCount(context.Context, *DescribeCountRequest) (*CountSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeCount not implemented")
 }
 func (UnimplementedCountInfoServer) mustEmbedUnimplementedCountInfoServer() {}
+func (UnimplementedCountInfoServer) testEmbeddedByValue()                   {}
 
 // UnsafeCountInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to CountInfoServer will
@@ -310,6 +316,13 @@ type UnsafeCountInfoServer interface {
 }
 
 func RegisterCountInfoServer(s grpc.ServiceRegistrar, srv CountInfoServer) {
+	// If the following call pancis, it indicates UnimplementedCountInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&CountInfo_ServiceDesc, srv)
 }
 
