@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/electric.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	ElectricApi_GetDemand_FullMethodName        = "/smartcore.traits.ElectricApi/GetDemand"
@@ -32,11 +32,14 @@ const (
 // ElectricApiClient is the client API for ElectricApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Trait for devices that are electrically powered.
+// Can apply to both the end device or an intermediary, for example a TV or an extension cord.
 type ElectricApiClient interface {
 	// GetDemand returns the current electricity demand for the device.
 	GetDemand(ctx context.Context, in *GetDemandRequest, opts ...grpc.CallOption) (*ElectricDemand, error)
 	// PullDemand returns a stream of changes to the current electricity demand for the device.
-	PullDemand(ctx context.Context, in *PullDemandRequest, opts ...grpc.CallOption) (ElectricApi_PullDemandClient, error)
+	PullDemand(ctx context.Context, in *PullDemandRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullDemandResponse], error)
 	GetActiveMode(ctx context.Context, in *GetActiveModeRequest, opts ...grpc.CallOption) (*ElectricMode, error)
 	// Configure the device to use the given mode.
 	// The id of the given mode should match a known mode for the device.
@@ -44,11 +47,11 @@ type ElectricApiClient interface {
 	// ClearActiveMode instructs the device to discard any external mode selection, typically by applying the normal
 	// operation mode for the device.
 	ClearActiveMode(ctx context.Context, in *ClearActiveModeRequest, opts ...grpc.CallOption) (*ElectricMode, error)
-	PullActiveMode(ctx context.Context, in *PullActiveModeRequest, opts ...grpc.CallOption) (ElectricApi_PullActiveModeClient, error)
+	PullActiveMode(ctx context.Context, in *PullActiveModeRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullActiveModeResponse], error)
 	// ListModes lists all electric modes for a device
 	ListModes(ctx context.Context, in *ListModesRequest, opts ...grpc.CallOption) (*ListModesResponse, error)
 	// PullModes returns a stream of changes to the list of supported modes for the device.
-	PullModes(ctx context.Context, in *PullModesRequest, opts ...grpc.CallOption) (ElectricApi_PullModesClient, error)
+	PullModes(ctx context.Context, in *PullModesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullModesResponse], error)
 }
 
 type electricApiClient struct {
@@ -60,20 +63,22 @@ func NewElectricApiClient(cc grpc.ClientConnInterface) ElectricApiClient {
 }
 
 func (c *electricApiClient) GetDemand(ctx context.Context, in *GetDemandRequest, opts ...grpc.CallOption) (*ElectricDemand, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ElectricDemand)
-	err := c.cc.Invoke(ctx, ElectricApi_GetDemand_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ElectricApi_GetDemand_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *electricApiClient) PullDemand(ctx context.Context, in *PullDemandRequest, opts ...grpc.CallOption) (ElectricApi_PullDemandClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ElectricApi_ServiceDesc.Streams[0], ElectricApi_PullDemand_FullMethodName, opts...)
+func (c *electricApiClient) PullDemand(ctx context.Context, in *PullDemandRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullDemandResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ElectricApi_ServiceDesc.Streams[0], ElectricApi_PullDemand_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &electricApiPullDemandClient{stream}
+	x := &grpc.GenericClientStream[PullDemandRequest, PullDemandResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -83,26 +88,13 @@ func (c *electricApiClient) PullDemand(ctx context.Context, in *PullDemandReques
 	return x, nil
 }
 
-type ElectricApi_PullDemandClient interface {
-	Recv() (*PullDemandResponse, error)
-	grpc.ClientStream
-}
-
-type electricApiPullDemandClient struct {
-	grpc.ClientStream
-}
-
-func (x *electricApiPullDemandClient) Recv() (*PullDemandResponse, error) {
-	m := new(PullDemandResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ElectricApi_PullDemandClient = grpc.ServerStreamingClient[PullDemandResponse]
 
 func (c *electricApiClient) GetActiveMode(ctx context.Context, in *GetActiveModeRequest, opts ...grpc.CallOption) (*ElectricMode, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ElectricMode)
-	err := c.cc.Invoke(ctx, ElectricApi_GetActiveMode_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ElectricApi_GetActiveMode_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -110,8 +102,9 @@ func (c *electricApiClient) GetActiveMode(ctx context.Context, in *GetActiveMode
 }
 
 func (c *electricApiClient) UpdateActiveMode(ctx context.Context, in *UpdateActiveModeRequest, opts ...grpc.CallOption) (*ElectricMode, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ElectricMode)
-	err := c.cc.Invoke(ctx, ElectricApi_UpdateActiveMode_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ElectricApi_UpdateActiveMode_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -119,20 +112,22 @@ func (c *electricApiClient) UpdateActiveMode(ctx context.Context, in *UpdateActi
 }
 
 func (c *electricApiClient) ClearActiveMode(ctx context.Context, in *ClearActiveModeRequest, opts ...grpc.CallOption) (*ElectricMode, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ElectricMode)
-	err := c.cc.Invoke(ctx, ElectricApi_ClearActiveMode_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ElectricApi_ClearActiveMode_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *electricApiClient) PullActiveMode(ctx context.Context, in *PullActiveModeRequest, opts ...grpc.CallOption) (ElectricApi_PullActiveModeClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ElectricApi_ServiceDesc.Streams[1], ElectricApi_PullActiveMode_FullMethodName, opts...)
+func (c *electricApiClient) PullActiveMode(ctx context.Context, in *PullActiveModeRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullActiveModeResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ElectricApi_ServiceDesc.Streams[1], ElectricApi_PullActiveMode_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &electricApiPullActiveModeClient{stream}
+	x := &grpc.GenericClientStream[PullActiveModeRequest, PullActiveModeResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -142,38 +137,26 @@ func (c *electricApiClient) PullActiveMode(ctx context.Context, in *PullActiveMo
 	return x, nil
 }
 
-type ElectricApi_PullActiveModeClient interface {
-	Recv() (*PullActiveModeResponse, error)
-	grpc.ClientStream
-}
-
-type electricApiPullActiveModeClient struct {
-	grpc.ClientStream
-}
-
-func (x *electricApiPullActiveModeClient) Recv() (*PullActiveModeResponse, error) {
-	m := new(PullActiveModeResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ElectricApi_PullActiveModeClient = grpc.ServerStreamingClient[PullActiveModeResponse]
 
 func (c *electricApiClient) ListModes(ctx context.Context, in *ListModesRequest, opts ...grpc.CallOption) (*ListModesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListModesResponse)
-	err := c.cc.Invoke(ctx, ElectricApi_ListModes_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ElectricApi_ListModes_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *electricApiClient) PullModes(ctx context.Context, in *PullModesRequest, opts ...grpc.CallOption) (ElectricApi_PullModesClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ElectricApi_ServiceDesc.Streams[2], ElectricApi_PullModes_FullMethodName, opts...)
+func (c *electricApiClient) PullModes(ctx context.Context, in *PullModesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullModesResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ElectricApi_ServiceDesc.Streams[2], ElectricApi_PullModes_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &electricApiPullModesClient{stream}
+	x := &grpc.GenericClientStream[PullModesRequest, PullModesResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -183,31 +166,20 @@ func (c *electricApiClient) PullModes(ctx context.Context, in *PullModesRequest,
 	return x, nil
 }
 
-type ElectricApi_PullModesClient interface {
-	Recv() (*PullModesResponse, error)
-	grpc.ClientStream
-}
-
-type electricApiPullModesClient struct {
-	grpc.ClientStream
-}
-
-func (x *electricApiPullModesClient) Recv() (*PullModesResponse, error) {
-	m := new(PullModesResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ElectricApi_PullModesClient = grpc.ServerStreamingClient[PullModesResponse]
 
 // ElectricApiServer is the server API for ElectricApi service.
 // All implementations must embed UnimplementedElectricApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Trait for devices that are electrically powered.
+// Can apply to both the end device or an intermediary, for example a TV or an extension cord.
 type ElectricApiServer interface {
 	// GetDemand returns the current electricity demand for the device.
 	GetDemand(context.Context, *GetDemandRequest) (*ElectricDemand, error)
 	// PullDemand returns a stream of changes to the current electricity demand for the device.
-	PullDemand(*PullDemandRequest, ElectricApi_PullDemandServer) error
+	PullDemand(*PullDemandRequest, grpc.ServerStreamingServer[PullDemandResponse]) error
 	GetActiveMode(context.Context, *GetActiveModeRequest) (*ElectricMode, error)
 	// Configure the device to use the given mode.
 	// The id of the given mode should match a known mode for the device.
@@ -215,22 +187,25 @@ type ElectricApiServer interface {
 	// ClearActiveMode instructs the device to discard any external mode selection, typically by applying the normal
 	// operation mode for the device.
 	ClearActiveMode(context.Context, *ClearActiveModeRequest) (*ElectricMode, error)
-	PullActiveMode(*PullActiveModeRequest, ElectricApi_PullActiveModeServer) error
+	PullActiveMode(*PullActiveModeRequest, grpc.ServerStreamingServer[PullActiveModeResponse]) error
 	// ListModes lists all electric modes for a device
 	ListModes(context.Context, *ListModesRequest) (*ListModesResponse, error)
 	// PullModes returns a stream of changes to the list of supported modes for the device.
-	PullModes(*PullModesRequest, ElectricApi_PullModesServer) error
+	PullModes(*PullModesRequest, grpc.ServerStreamingServer[PullModesResponse]) error
 	mustEmbedUnimplementedElectricApiServer()
 }
 
-// UnimplementedElectricApiServer must be embedded to have forward compatible implementations.
-type UnimplementedElectricApiServer struct {
-}
+// UnimplementedElectricApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedElectricApiServer struct{}
 
 func (UnimplementedElectricApiServer) GetDemand(context.Context, *GetDemandRequest) (*ElectricDemand, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetDemand not implemented")
 }
-func (UnimplementedElectricApiServer) PullDemand(*PullDemandRequest, ElectricApi_PullDemandServer) error {
+func (UnimplementedElectricApiServer) PullDemand(*PullDemandRequest, grpc.ServerStreamingServer[PullDemandResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullDemand not implemented")
 }
 func (UnimplementedElectricApiServer) GetActiveMode(context.Context, *GetActiveModeRequest) (*ElectricMode, error) {
@@ -242,16 +217,17 @@ func (UnimplementedElectricApiServer) UpdateActiveMode(context.Context, *UpdateA
 func (UnimplementedElectricApiServer) ClearActiveMode(context.Context, *ClearActiveModeRequest) (*ElectricMode, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ClearActiveMode not implemented")
 }
-func (UnimplementedElectricApiServer) PullActiveMode(*PullActiveModeRequest, ElectricApi_PullActiveModeServer) error {
+func (UnimplementedElectricApiServer) PullActiveMode(*PullActiveModeRequest, grpc.ServerStreamingServer[PullActiveModeResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullActiveMode not implemented")
 }
 func (UnimplementedElectricApiServer) ListModes(context.Context, *ListModesRequest) (*ListModesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListModes not implemented")
 }
-func (UnimplementedElectricApiServer) PullModes(*PullModesRequest, ElectricApi_PullModesServer) error {
+func (UnimplementedElectricApiServer) PullModes(*PullModesRequest, grpc.ServerStreamingServer[PullModesResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullModes not implemented")
 }
 func (UnimplementedElectricApiServer) mustEmbedUnimplementedElectricApiServer() {}
+func (UnimplementedElectricApiServer) testEmbeddedByValue()                     {}
 
 // UnsafeElectricApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ElectricApiServer will
@@ -261,6 +237,13 @@ type UnsafeElectricApiServer interface {
 }
 
 func RegisterElectricApiServer(s grpc.ServiceRegistrar, srv ElectricApiServer) {
+	// If the following call pancis, it indicates UnimplementedElectricApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ElectricApi_ServiceDesc, srv)
 }
 
@@ -287,21 +270,11 @@ func _ElectricApi_PullDemand_Handler(srv interface{}, stream grpc.ServerStream) 
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ElectricApiServer).PullDemand(m, &electricApiPullDemandServer{stream})
+	return srv.(ElectricApiServer).PullDemand(m, &grpc.GenericServerStream[PullDemandRequest, PullDemandResponse]{ServerStream: stream})
 }
 
-type ElectricApi_PullDemandServer interface {
-	Send(*PullDemandResponse) error
-	grpc.ServerStream
-}
-
-type electricApiPullDemandServer struct {
-	grpc.ServerStream
-}
-
-func (x *electricApiPullDemandServer) Send(m *PullDemandResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ElectricApi_PullDemandServer = grpc.ServerStreamingServer[PullDemandResponse]
 
 func _ElectricApi_GetActiveMode_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(GetActiveModeRequest)
@@ -362,21 +335,11 @@ func _ElectricApi_PullActiveMode_Handler(srv interface{}, stream grpc.ServerStre
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ElectricApiServer).PullActiveMode(m, &electricApiPullActiveModeServer{stream})
+	return srv.(ElectricApiServer).PullActiveMode(m, &grpc.GenericServerStream[PullActiveModeRequest, PullActiveModeResponse]{ServerStream: stream})
 }
 
-type ElectricApi_PullActiveModeServer interface {
-	Send(*PullActiveModeResponse) error
-	grpc.ServerStream
-}
-
-type electricApiPullActiveModeServer struct {
-	grpc.ServerStream
-}
-
-func (x *electricApiPullActiveModeServer) Send(m *PullActiveModeResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ElectricApi_PullActiveModeServer = grpc.ServerStreamingServer[PullActiveModeResponse]
 
 func _ElectricApi_ListModes_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ListModesRequest)
@@ -401,21 +364,11 @@ func _ElectricApi_PullModes_Handler(srv interface{}, stream grpc.ServerStream) e
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ElectricApiServer).PullModes(m, &electricApiPullModesServer{stream})
+	return srv.(ElectricApiServer).PullModes(m, &grpc.GenericServerStream[PullModesRequest, PullModesResponse]{ServerStream: stream})
 }
 
-type ElectricApi_PullModesServer interface {
-	Send(*PullModesResponse) error
-	grpc.ServerStream
-}
-
-type electricApiPullModesServer struct {
-	grpc.ServerStream
-}
-
-func (x *electricApiPullModesServer) Send(m *PullModesResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ElectricApi_PullModesServer = grpc.ServerStreamingServer[PullModesResponse]
 
 // ElectricApi_ServiceDesc is the grpc.ServiceDesc for ElectricApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -465,11 +418,11 @@ var ElectricApi_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "traits/electric.proto",
 }
 
-const ()
-
 // ElectricInfoClient is the client API for ElectricInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ElectricInfoClient interface {
 }
 
@@ -483,16 +436,22 @@ func NewElectricInfoClient(cc grpc.ClientConnInterface) ElectricInfoClient {
 
 // ElectricInfoServer is the server API for ElectricInfo service.
 // All implementations must embed UnimplementedElectricInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ElectricInfoServer interface {
 	mustEmbedUnimplementedElectricInfoServer()
 }
 
-// UnimplementedElectricInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedElectricInfoServer struct {
-}
+// UnimplementedElectricInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedElectricInfoServer struct{}
 
 func (UnimplementedElectricInfoServer) mustEmbedUnimplementedElectricInfoServer() {}
+func (UnimplementedElectricInfoServer) testEmbeddedByValue()                      {}
 
 // UnsafeElectricInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ElectricInfoServer will
@@ -502,6 +461,13 @@ type UnsafeElectricInfoServer interface {
 }
 
 func RegisterElectricInfoServer(s grpc.ServiceRegistrar, srv ElectricInfoServer) {
+	// If the following call pancis, it indicates UnimplementedElectricInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ElectricInfo_ServiceDesc, srv)
 }
 
