@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/emergency.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	EmergencyApi_GetEmergency_FullMethodName    = "/smartcore.traits.EmergencyApi/GetEmergency"
@@ -27,13 +27,16 @@ const (
 // EmergencyApiClient is the client API for EmergencyApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Report and manages emergencies. Typically not associated with a single device, more suited to buildings or spaces.
+// It's likely that some implementation will not allow write of emergency state.
 type EmergencyApiClient interface {
 	// Get the current emergency state
 	GetEmergency(ctx context.Context, in *GetEmergencyRequest, opts ...grpc.CallOption) (*Emergency, error)
 	// Update the current emergency state
 	UpdateEmergency(ctx context.Context, in *UpdateEmergencyRequest, opts ...grpc.CallOption) (*Emergency, error)
 	// Be notified of changes to the emergency state
-	PullEmergency(ctx context.Context, in *PullEmergencyRequest, opts ...grpc.CallOption) (EmergencyApi_PullEmergencyClient, error)
+	PullEmergency(ctx context.Context, in *PullEmergencyRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullEmergencyResponse], error)
 }
 
 type emergencyApiClient struct {
@@ -45,8 +48,9 @@ func NewEmergencyApiClient(cc grpc.ClientConnInterface) EmergencyApiClient {
 }
 
 func (c *emergencyApiClient) GetEmergency(ctx context.Context, in *GetEmergencyRequest, opts ...grpc.CallOption) (*Emergency, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Emergency)
-	err := c.cc.Invoke(ctx, EmergencyApi_GetEmergency_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EmergencyApi_GetEmergency_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,20 +58,22 @@ func (c *emergencyApiClient) GetEmergency(ctx context.Context, in *GetEmergencyR
 }
 
 func (c *emergencyApiClient) UpdateEmergency(ctx context.Context, in *UpdateEmergencyRequest, opts ...grpc.CallOption) (*Emergency, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Emergency)
-	err := c.cc.Invoke(ctx, EmergencyApi_UpdateEmergency_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EmergencyApi_UpdateEmergency_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *emergencyApiClient) PullEmergency(ctx context.Context, in *PullEmergencyRequest, opts ...grpc.CallOption) (EmergencyApi_PullEmergencyClient, error) {
-	stream, err := c.cc.NewStream(ctx, &EmergencyApi_ServiceDesc.Streams[0], EmergencyApi_PullEmergency_FullMethodName, opts...)
+func (c *emergencyApiClient) PullEmergency(ctx context.Context, in *PullEmergencyRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullEmergencyResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &EmergencyApi_ServiceDesc.Streams[0], EmergencyApi_PullEmergency_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &emergencyApiPullEmergencyClient{stream}
+	x := &grpc.GenericClientStream[PullEmergencyRequest, PullEmergencyResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -77,39 +83,31 @@ func (c *emergencyApiClient) PullEmergency(ctx context.Context, in *PullEmergenc
 	return x, nil
 }
 
-type EmergencyApi_PullEmergencyClient interface {
-	Recv() (*PullEmergencyResponse, error)
-	grpc.ClientStream
-}
-
-type emergencyApiPullEmergencyClient struct {
-	grpc.ClientStream
-}
-
-func (x *emergencyApiPullEmergencyClient) Recv() (*PullEmergencyResponse, error) {
-	m := new(PullEmergencyResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type EmergencyApi_PullEmergencyClient = grpc.ServerStreamingClient[PullEmergencyResponse]
 
 // EmergencyApiServer is the server API for EmergencyApi service.
 // All implementations must embed UnimplementedEmergencyApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Report and manages emergencies. Typically not associated with a single device, more suited to buildings or spaces.
+// It's likely that some implementation will not allow write of emergency state.
 type EmergencyApiServer interface {
 	// Get the current emergency state
 	GetEmergency(context.Context, *GetEmergencyRequest) (*Emergency, error)
 	// Update the current emergency state
 	UpdateEmergency(context.Context, *UpdateEmergencyRequest) (*Emergency, error)
 	// Be notified of changes to the emergency state
-	PullEmergency(*PullEmergencyRequest, EmergencyApi_PullEmergencyServer) error
+	PullEmergency(*PullEmergencyRequest, grpc.ServerStreamingServer[PullEmergencyResponse]) error
 	mustEmbedUnimplementedEmergencyApiServer()
 }
 
-// UnimplementedEmergencyApiServer must be embedded to have forward compatible implementations.
-type UnimplementedEmergencyApiServer struct {
-}
+// UnimplementedEmergencyApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedEmergencyApiServer struct{}
 
 func (UnimplementedEmergencyApiServer) GetEmergency(context.Context, *GetEmergencyRequest) (*Emergency, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetEmergency not implemented")
@@ -117,10 +115,11 @@ func (UnimplementedEmergencyApiServer) GetEmergency(context.Context, *GetEmergen
 func (UnimplementedEmergencyApiServer) UpdateEmergency(context.Context, *UpdateEmergencyRequest) (*Emergency, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateEmergency not implemented")
 }
-func (UnimplementedEmergencyApiServer) PullEmergency(*PullEmergencyRequest, EmergencyApi_PullEmergencyServer) error {
+func (UnimplementedEmergencyApiServer) PullEmergency(*PullEmergencyRequest, grpc.ServerStreamingServer[PullEmergencyResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullEmergency not implemented")
 }
 func (UnimplementedEmergencyApiServer) mustEmbedUnimplementedEmergencyApiServer() {}
+func (UnimplementedEmergencyApiServer) testEmbeddedByValue()                      {}
 
 // UnsafeEmergencyApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to EmergencyApiServer will
@@ -130,6 +129,13 @@ type UnsafeEmergencyApiServer interface {
 }
 
 func RegisterEmergencyApiServer(s grpc.ServiceRegistrar, srv EmergencyApiServer) {
+	// If the following call pancis, it indicates UnimplementedEmergencyApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&EmergencyApi_ServiceDesc, srv)
 }
 
@@ -174,21 +180,11 @@ func _EmergencyApi_PullEmergency_Handler(srv interface{}, stream grpc.ServerStre
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(EmergencyApiServer).PullEmergency(m, &emergencyApiPullEmergencyServer{stream})
+	return srv.(EmergencyApiServer).PullEmergency(m, &grpc.GenericServerStream[PullEmergencyRequest, PullEmergencyResponse]{ServerStream: stream})
 }
 
-type EmergencyApi_PullEmergencyServer interface {
-	Send(*PullEmergencyResponse) error
-	grpc.ServerStream
-}
-
-type emergencyApiPullEmergencyServer struct {
-	grpc.ServerStream
-}
-
-func (x *emergencyApiPullEmergencyServer) Send(m *PullEmergencyResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type EmergencyApi_PullEmergencyServer = grpc.ServerStreamingServer[PullEmergencyResponse]
 
 // EmergencyApi_ServiceDesc is the grpc.ServiceDesc for EmergencyApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -223,6 +219,8 @@ const (
 // EmergencyInfoClient is the client API for EmergencyInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type EmergencyInfoClient interface {
 	// Get information about how a named device implements Emergency features
 	DescribeEmergency(ctx context.Context, in *DescribeEmergencyRequest, opts ...grpc.CallOption) (*EmergencySupport, error)
@@ -237,8 +235,9 @@ func NewEmergencyInfoClient(cc grpc.ClientConnInterface) EmergencyInfoClient {
 }
 
 func (c *emergencyInfoClient) DescribeEmergency(ctx context.Context, in *DescribeEmergencyRequest, opts ...grpc.CallOption) (*EmergencySupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(EmergencySupport)
-	err := c.cc.Invoke(ctx, EmergencyInfo_DescribeEmergency_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EmergencyInfo_DescribeEmergency_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -247,21 +246,27 @@ func (c *emergencyInfoClient) DescribeEmergency(ctx context.Context, in *Describ
 
 // EmergencyInfoServer is the server API for EmergencyInfo service.
 // All implementations must embed UnimplementedEmergencyInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type EmergencyInfoServer interface {
 	// Get information about how a named device implements Emergency features
 	DescribeEmergency(context.Context, *DescribeEmergencyRequest) (*EmergencySupport, error)
 	mustEmbedUnimplementedEmergencyInfoServer()
 }
 
-// UnimplementedEmergencyInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedEmergencyInfoServer struct {
-}
+// UnimplementedEmergencyInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedEmergencyInfoServer struct{}
 
 func (UnimplementedEmergencyInfoServer) DescribeEmergency(context.Context, *DescribeEmergencyRequest) (*EmergencySupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeEmergency not implemented")
 }
 func (UnimplementedEmergencyInfoServer) mustEmbedUnimplementedEmergencyInfoServer() {}
+func (UnimplementedEmergencyInfoServer) testEmbeddedByValue()                       {}
 
 // UnsafeEmergencyInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to EmergencyInfoServer will
@@ -271,6 +276,13 @@ type UnsafeEmergencyInfoServer interface {
 }
 
 func RegisterEmergencyInfoServer(s grpc.ServiceRegistrar, srv EmergencyInfoServer) {
+	// If the following call pancis, it indicates UnimplementedEmergencyInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&EmergencyInfo_ServiceDesc, srv)
 }
 
