@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/energy_storage.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	EnergyStorageApi_GetEnergyLevel_FullMethodName  = "/smartcore.traits.EnergyStorageApi/GetEnergyLevel"
@@ -27,11 +27,14 @@ const (
 // EnergyStorageApiClient is the client API for EnergyStorageApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Trait for devices that can store energy or charge another device.
+// Typically applied devices with batteries or chargers of those devices.
 type EnergyStorageApiClient interface {
 	// GetEnergyLevel retrieves the current charge level and potential charge level for the device.
 	GetEnergyLevel(ctx context.Context, in *GetEnergyLevelRequest, opts ...grpc.CallOption) (*EnergyLevel, error)
 	// PullEnergyLevel subscribes to changes in energy level.
-	PullEnergyLevel(ctx context.Context, in *PullEnergyLevelRequest, opts ...grpc.CallOption) (EnergyStorageApi_PullEnergyLevelClient, error)
+	PullEnergyLevel(ctx context.Context, in *PullEnergyLevelRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullEnergyLevelResponse], error)
 	// Charge starts or stops the device charging.
 	// If the device is not plugged in then this will return a FAILED_PRECONDITION error.
 	// Devices may start or stop charging on their own.
@@ -47,20 +50,22 @@ func NewEnergyStorageApiClient(cc grpc.ClientConnInterface) EnergyStorageApiClie
 }
 
 func (c *energyStorageApiClient) GetEnergyLevel(ctx context.Context, in *GetEnergyLevelRequest, opts ...grpc.CallOption) (*EnergyLevel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(EnergyLevel)
-	err := c.cc.Invoke(ctx, EnergyStorageApi_GetEnergyLevel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnergyStorageApi_GetEnergyLevel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *energyStorageApiClient) PullEnergyLevel(ctx context.Context, in *PullEnergyLevelRequest, opts ...grpc.CallOption) (EnergyStorageApi_PullEnergyLevelClient, error) {
-	stream, err := c.cc.NewStream(ctx, &EnergyStorageApi_ServiceDesc.Streams[0], EnergyStorageApi_PullEnergyLevel_FullMethodName, opts...)
+func (c *energyStorageApiClient) PullEnergyLevel(ctx context.Context, in *PullEnergyLevelRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullEnergyLevelResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &EnergyStorageApi_ServiceDesc.Streams[0], EnergyStorageApi_PullEnergyLevel_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &energyStorageApiPullEnergyLevelClient{stream}
+	x := &grpc.GenericClientStream[PullEnergyLevelRequest, PullEnergyLevelResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -70,26 +75,13 @@ func (c *energyStorageApiClient) PullEnergyLevel(ctx context.Context, in *PullEn
 	return x, nil
 }
 
-type EnergyStorageApi_PullEnergyLevelClient interface {
-	Recv() (*PullEnergyLevelResponse, error)
-	grpc.ClientStream
-}
-
-type energyStorageApiPullEnergyLevelClient struct {
-	grpc.ClientStream
-}
-
-func (x *energyStorageApiPullEnergyLevelClient) Recv() (*PullEnergyLevelResponse, error) {
-	m := new(PullEnergyLevelResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type EnergyStorageApi_PullEnergyLevelClient = grpc.ServerStreamingClient[PullEnergyLevelResponse]
 
 func (c *energyStorageApiClient) Charge(ctx context.Context, in *ChargeRequest, opts ...grpc.CallOption) (*ChargeResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ChargeResponse)
-	err := c.cc.Invoke(ctx, EnergyStorageApi_Charge_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnergyStorageApi_Charge_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -98,12 +90,15 @@ func (c *energyStorageApiClient) Charge(ctx context.Context, in *ChargeRequest, 
 
 // EnergyStorageApiServer is the server API for EnergyStorageApi service.
 // All implementations must embed UnimplementedEnergyStorageApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Trait for devices that can store energy or charge another device.
+// Typically applied devices with batteries or chargers of those devices.
 type EnergyStorageApiServer interface {
 	// GetEnergyLevel retrieves the current charge level and potential charge level for the device.
 	GetEnergyLevel(context.Context, *GetEnergyLevelRequest) (*EnergyLevel, error)
 	// PullEnergyLevel subscribes to changes in energy level.
-	PullEnergyLevel(*PullEnergyLevelRequest, EnergyStorageApi_PullEnergyLevelServer) error
+	PullEnergyLevel(*PullEnergyLevelRequest, grpc.ServerStreamingServer[PullEnergyLevelResponse]) error
 	// Charge starts or stops the device charging.
 	// If the device is not plugged in then this will return a FAILED_PRECONDITION error.
 	// Devices may start or stop charging on their own.
@@ -111,20 +106,24 @@ type EnergyStorageApiServer interface {
 	mustEmbedUnimplementedEnergyStorageApiServer()
 }
 
-// UnimplementedEnergyStorageApiServer must be embedded to have forward compatible implementations.
-type UnimplementedEnergyStorageApiServer struct {
-}
+// UnimplementedEnergyStorageApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedEnergyStorageApiServer struct{}
 
 func (UnimplementedEnergyStorageApiServer) GetEnergyLevel(context.Context, *GetEnergyLevelRequest) (*EnergyLevel, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetEnergyLevel not implemented")
 }
-func (UnimplementedEnergyStorageApiServer) PullEnergyLevel(*PullEnergyLevelRequest, EnergyStorageApi_PullEnergyLevelServer) error {
+func (UnimplementedEnergyStorageApiServer) PullEnergyLevel(*PullEnergyLevelRequest, grpc.ServerStreamingServer[PullEnergyLevelResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullEnergyLevel not implemented")
 }
 func (UnimplementedEnergyStorageApiServer) Charge(context.Context, *ChargeRequest) (*ChargeResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Charge not implemented")
 }
 func (UnimplementedEnergyStorageApiServer) mustEmbedUnimplementedEnergyStorageApiServer() {}
+func (UnimplementedEnergyStorageApiServer) testEmbeddedByValue()                          {}
 
 // UnsafeEnergyStorageApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to EnergyStorageApiServer will
@@ -134,6 +133,13 @@ type UnsafeEnergyStorageApiServer interface {
 }
 
 func RegisterEnergyStorageApiServer(s grpc.ServiceRegistrar, srv EnergyStorageApiServer) {
+	// If the following call pancis, it indicates UnimplementedEnergyStorageApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&EnergyStorageApi_ServiceDesc, srv)
 }
 
@@ -160,21 +166,11 @@ func _EnergyStorageApi_PullEnergyLevel_Handler(srv interface{}, stream grpc.Serv
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(EnergyStorageApiServer).PullEnergyLevel(m, &energyStorageApiPullEnergyLevelServer{stream})
+	return srv.(EnergyStorageApiServer).PullEnergyLevel(m, &grpc.GenericServerStream[PullEnergyLevelRequest, PullEnergyLevelResponse]{ServerStream: stream})
 }
 
-type EnergyStorageApi_PullEnergyLevelServer interface {
-	Send(*PullEnergyLevelResponse) error
-	grpc.ServerStream
-}
-
-type energyStorageApiPullEnergyLevelServer struct {
-	grpc.ServerStream
-}
-
-func (x *energyStorageApiPullEnergyLevelServer) Send(m *PullEnergyLevelResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type EnergyStorageApi_PullEnergyLevelServer = grpc.ServerStreamingServer[PullEnergyLevelResponse]
 
 func _EnergyStorageApi_Charge_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ChargeRequest)
@@ -227,6 +223,8 @@ const (
 // EnergyStorageInfoClient is the client API for EnergyStorageInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type EnergyStorageInfoClient interface {
 	// Get information about how a named device implements EnergyLevel features
 	DescribeEnergyLevel(ctx context.Context, in *DescribeEnergyLevelRequest, opts ...grpc.CallOption) (*EnergyLevelSupport, error)
@@ -241,8 +239,9 @@ func NewEnergyStorageInfoClient(cc grpc.ClientConnInterface) EnergyStorageInfoCl
 }
 
 func (c *energyStorageInfoClient) DescribeEnergyLevel(ctx context.Context, in *DescribeEnergyLevelRequest, opts ...grpc.CallOption) (*EnergyLevelSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(EnergyLevelSupport)
-	err := c.cc.Invoke(ctx, EnergyStorageInfo_DescribeEnergyLevel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnergyStorageInfo_DescribeEnergyLevel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -251,21 +250,27 @@ func (c *energyStorageInfoClient) DescribeEnergyLevel(ctx context.Context, in *D
 
 // EnergyStorageInfoServer is the server API for EnergyStorageInfo service.
 // All implementations must embed UnimplementedEnergyStorageInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type EnergyStorageInfoServer interface {
 	// Get information about how a named device implements EnergyLevel features
 	DescribeEnergyLevel(context.Context, *DescribeEnergyLevelRequest) (*EnergyLevelSupport, error)
 	mustEmbedUnimplementedEnergyStorageInfoServer()
 }
 
-// UnimplementedEnergyStorageInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedEnergyStorageInfoServer struct {
-}
+// UnimplementedEnergyStorageInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedEnergyStorageInfoServer struct{}
 
 func (UnimplementedEnergyStorageInfoServer) DescribeEnergyLevel(context.Context, *DescribeEnergyLevelRequest) (*EnergyLevelSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeEnergyLevel not implemented")
 }
 func (UnimplementedEnergyStorageInfoServer) mustEmbedUnimplementedEnergyStorageInfoServer() {}
+func (UnimplementedEnergyStorageInfoServer) testEmbeddedByValue()                           {}
 
 // UnsafeEnergyStorageInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to EnergyStorageInfoServer will
@@ -275,6 +280,13 @@ type UnsafeEnergyStorageInfoServer interface {
 }
 
 func RegisterEnergyStorageInfoServer(s grpc.ServiceRegistrar, srv EnergyStorageInfoServer) {
+	// If the following call pancis, it indicates UnimplementedEnergyStorageInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&EnergyStorageInfo_ServiceDesc, srv)
 }
 
