@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/enter_leave_sensor.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	EnterLeaveSensorApi_PullEnterLeaveEvents_FullMethodName  = "/smartcore.traits.EnterLeaveSensorApi/PullEnterLeaveEvents"
@@ -27,9 +27,12 @@ const (
 // EnterLeaveSensorApiClient is the client API for EnterLeaveSensorApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// EnterLeaveSensor applies to devices that can track objects entering or leaving.
+// It could apply to footfall counters, turn styles, or car park entry barriers.
 type EnterLeaveSensorApiClient interface {
 	// Subscribe to new enter or leave events.
-	PullEnterLeaveEvents(ctx context.Context, in *PullEnterLeaveEventsRequest, opts ...grpc.CallOption) (EnterLeaveSensorApi_PullEnterLeaveEventsClient, error)
+	PullEnterLeaveEvents(ctx context.Context, in *PullEnterLeaveEventsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullEnterLeaveEventsResponse], error)
 	// Get information about the enter leave state.
 	// The response will not contain a direction or occupant, but may include totals.
 	GetEnterLeaveEvent(ctx context.Context, in *GetEnterLeaveEventRequest, opts ...grpc.CallOption) (*EnterLeaveEvent, error)
@@ -45,12 +48,13 @@ func NewEnterLeaveSensorApiClient(cc grpc.ClientConnInterface) EnterLeaveSensorA
 	return &enterLeaveSensorApiClient{cc}
 }
 
-func (c *enterLeaveSensorApiClient) PullEnterLeaveEvents(ctx context.Context, in *PullEnterLeaveEventsRequest, opts ...grpc.CallOption) (EnterLeaveSensorApi_PullEnterLeaveEventsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &EnterLeaveSensorApi_ServiceDesc.Streams[0], EnterLeaveSensorApi_PullEnterLeaveEvents_FullMethodName, opts...)
+func (c *enterLeaveSensorApiClient) PullEnterLeaveEvents(ctx context.Context, in *PullEnterLeaveEventsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullEnterLeaveEventsResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &EnterLeaveSensorApi_ServiceDesc.Streams[0], EnterLeaveSensorApi_PullEnterLeaveEvents_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &enterLeaveSensorApiPullEnterLeaveEventsClient{stream}
+	x := &grpc.GenericClientStream[PullEnterLeaveEventsRequest, PullEnterLeaveEventsResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -60,26 +64,13 @@ func (c *enterLeaveSensorApiClient) PullEnterLeaveEvents(ctx context.Context, in
 	return x, nil
 }
 
-type EnterLeaveSensorApi_PullEnterLeaveEventsClient interface {
-	Recv() (*PullEnterLeaveEventsResponse, error)
-	grpc.ClientStream
-}
-
-type enterLeaveSensorApiPullEnterLeaveEventsClient struct {
-	grpc.ClientStream
-}
-
-func (x *enterLeaveSensorApiPullEnterLeaveEventsClient) Recv() (*PullEnterLeaveEventsResponse, error) {
-	m := new(PullEnterLeaveEventsResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type EnterLeaveSensorApi_PullEnterLeaveEventsClient = grpc.ServerStreamingClient[PullEnterLeaveEventsResponse]
 
 func (c *enterLeaveSensorApiClient) GetEnterLeaveEvent(ctx context.Context, in *GetEnterLeaveEventRequest, opts ...grpc.CallOption) (*EnterLeaveEvent, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(EnterLeaveEvent)
-	err := c.cc.Invoke(ctx, EnterLeaveSensorApi_GetEnterLeaveEvent_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnterLeaveSensorApi_GetEnterLeaveEvent_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -87,8 +78,9 @@ func (c *enterLeaveSensorApiClient) GetEnterLeaveEvent(ctx context.Context, in *
 }
 
 func (c *enterLeaveSensorApiClient) ResetEnterLeaveTotals(ctx context.Context, in *ResetEnterLeaveTotalsRequest, opts ...grpc.CallOption) (*ResetEnterLeaveTotalsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ResetEnterLeaveTotalsResponse)
-	err := c.cc.Invoke(ctx, EnterLeaveSensorApi_ResetEnterLeaveTotals_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnterLeaveSensorApi_ResetEnterLeaveTotals_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -97,10 +89,13 @@ func (c *enterLeaveSensorApiClient) ResetEnterLeaveTotals(ctx context.Context, i
 
 // EnterLeaveSensorApiServer is the server API for EnterLeaveSensorApi service.
 // All implementations must embed UnimplementedEnterLeaveSensorApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// EnterLeaveSensor applies to devices that can track objects entering or leaving.
+// It could apply to footfall counters, turn styles, or car park entry barriers.
 type EnterLeaveSensorApiServer interface {
 	// Subscribe to new enter or leave events.
-	PullEnterLeaveEvents(*PullEnterLeaveEventsRequest, EnterLeaveSensorApi_PullEnterLeaveEventsServer) error
+	PullEnterLeaveEvents(*PullEnterLeaveEventsRequest, grpc.ServerStreamingServer[PullEnterLeaveEventsResponse]) error
 	// Get information about the enter leave state.
 	// The response will not contain a direction or occupant, but may include totals.
 	GetEnterLeaveEvent(context.Context, *GetEnterLeaveEventRequest) (*EnterLeaveEvent, error)
@@ -109,11 +104,14 @@ type EnterLeaveSensorApiServer interface {
 	mustEmbedUnimplementedEnterLeaveSensorApiServer()
 }
 
-// UnimplementedEnterLeaveSensorApiServer must be embedded to have forward compatible implementations.
-type UnimplementedEnterLeaveSensorApiServer struct {
-}
+// UnimplementedEnterLeaveSensorApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedEnterLeaveSensorApiServer struct{}
 
-func (UnimplementedEnterLeaveSensorApiServer) PullEnterLeaveEvents(*PullEnterLeaveEventsRequest, EnterLeaveSensorApi_PullEnterLeaveEventsServer) error {
+func (UnimplementedEnterLeaveSensorApiServer) PullEnterLeaveEvents(*PullEnterLeaveEventsRequest, grpc.ServerStreamingServer[PullEnterLeaveEventsResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullEnterLeaveEvents not implemented")
 }
 func (UnimplementedEnterLeaveSensorApiServer) GetEnterLeaveEvent(context.Context, *GetEnterLeaveEventRequest) (*EnterLeaveEvent, error) {
@@ -123,6 +121,7 @@ func (UnimplementedEnterLeaveSensorApiServer) ResetEnterLeaveTotals(context.Cont
 	return nil, status.Errorf(codes.Unimplemented, "method ResetEnterLeaveTotals not implemented")
 }
 func (UnimplementedEnterLeaveSensorApiServer) mustEmbedUnimplementedEnterLeaveSensorApiServer() {}
+func (UnimplementedEnterLeaveSensorApiServer) testEmbeddedByValue()                             {}
 
 // UnsafeEnterLeaveSensorApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to EnterLeaveSensorApiServer will
@@ -132,6 +131,13 @@ type UnsafeEnterLeaveSensorApiServer interface {
 }
 
 func RegisterEnterLeaveSensorApiServer(s grpc.ServiceRegistrar, srv EnterLeaveSensorApiServer) {
+	// If the following call pancis, it indicates UnimplementedEnterLeaveSensorApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&EnterLeaveSensorApi_ServiceDesc, srv)
 }
 
@@ -140,21 +146,11 @@ func _EnterLeaveSensorApi_PullEnterLeaveEvents_Handler(srv interface{}, stream g
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(EnterLeaveSensorApiServer).PullEnterLeaveEvents(m, &enterLeaveSensorApiPullEnterLeaveEventsServer{stream})
+	return srv.(EnterLeaveSensorApiServer).PullEnterLeaveEvents(m, &grpc.GenericServerStream[PullEnterLeaveEventsRequest, PullEnterLeaveEventsResponse]{ServerStream: stream})
 }
 
-type EnterLeaveSensorApi_PullEnterLeaveEventsServer interface {
-	Send(*PullEnterLeaveEventsResponse) error
-	grpc.ServerStream
-}
-
-type enterLeaveSensorApiPullEnterLeaveEventsServer struct {
-	grpc.ServerStream
-}
-
-func (x *enterLeaveSensorApiPullEnterLeaveEventsServer) Send(m *PullEnterLeaveEventsResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type EnterLeaveSensorApi_PullEnterLeaveEventsServer = grpc.ServerStreamingServer[PullEnterLeaveEventsResponse]
 
 func _EnterLeaveSensorApi_GetEnterLeaveEvent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(GetEnterLeaveEventRequest)
@@ -218,8 +214,6 @@ var EnterLeaveSensorApi_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "traits/enter_leave_sensor.proto",
 }
 
-const ()
-
 // EnterLeaveSensorInfoClient is the client API for EnterLeaveSensorInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
@@ -236,16 +230,20 @@ func NewEnterLeaveSensorInfoClient(cc grpc.ClientConnInterface) EnterLeaveSensor
 
 // EnterLeaveSensorInfoServer is the server API for EnterLeaveSensorInfo service.
 // All implementations must embed UnimplementedEnterLeaveSensorInfoServer
-// for forward compatibility
+// for forward compatibility.
 type EnterLeaveSensorInfoServer interface {
 	mustEmbedUnimplementedEnterLeaveSensorInfoServer()
 }
 
-// UnimplementedEnterLeaveSensorInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedEnterLeaveSensorInfoServer struct {
-}
+// UnimplementedEnterLeaveSensorInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedEnterLeaveSensorInfoServer struct{}
 
 func (UnimplementedEnterLeaveSensorInfoServer) mustEmbedUnimplementedEnterLeaveSensorInfoServer() {}
+func (UnimplementedEnterLeaveSensorInfoServer) testEmbeddedByValue()                              {}
 
 // UnsafeEnterLeaveSensorInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to EnterLeaveSensorInfoServer will
@@ -255,6 +253,13 @@ type UnsafeEnterLeaveSensorInfoServer interface {
 }
 
 func RegisterEnterLeaveSensorInfoServer(s grpc.ServiceRegistrar, srv EnterLeaveSensorInfoServer) {
+	// If the following call pancis, it indicates UnimplementedEnterLeaveSensorInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&EnterLeaveSensorInfo_ServiceDesc, srv)
 }
 
