@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/extend_retract.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	ExtendRetractApi_GetExtension_FullMethodName          = "/smartcore.traits.ExtendRetractApi/GetExtension"
@@ -29,6 +29,8 @@ const (
 // ExtendRetractApiClient is the client API for ExtendRetractApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Applicable to devices that can be extended or retracted. Like a tv mount or adjustable desk
 type ExtendRetractApiClient interface {
 	// Set the target state for the device
 	GetExtension(ctx context.Context, in *GetExtensionRequest, opts ...grpc.CallOption) (*Extension, error)
@@ -39,7 +41,7 @@ type ExtendRetractApiClient interface {
 	// Create a preset for the extension. If no preset extension value is specified then use the current position
 	CreateExtensionPreset(ctx context.Context, in *CreateExtensionPresetRequest, opts ...grpc.CallOption) (*ExtensionPreset, error)
 	// Get notified of changes to the OnOffState of a device
-	PullExtensions(ctx context.Context, in *PullExtensionsRequest, opts ...grpc.CallOption) (ExtendRetractApi_PullExtensionsClient, error)
+	PullExtensions(ctx context.Context, in *PullExtensionsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullExtensionsResponse], error)
 }
 
 type extendRetractApiClient struct {
@@ -51,8 +53,9 @@ func NewExtendRetractApiClient(cc grpc.ClientConnInterface) ExtendRetractApiClie
 }
 
 func (c *extendRetractApiClient) GetExtension(ctx context.Context, in *GetExtensionRequest, opts ...grpc.CallOption) (*Extension, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Extension)
-	err := c.cc.Invoke(ctx, ExtendRetractApi_GetExtension_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ExtendRetractApi_GetExtension_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -60,8 +63,9 @@ func (c *extendRetractApiClient) GetExtension(ctx context.Context, in *GetExtens
 }
 
 func (c *extendRetractApiClient) UpdateExtension(ctx context.Context, in *UpdateExtensionRequest, opts ...grpc.CallOption) (*Extension, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Extension)
-	err := c.cc.Invoke(ctx, ExtendRetractApi_UpdateExtension_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ExtendRetractApi_UpdateExtension_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -69,8 +73,9 @@ func (c *extendRetractApiClient) UpdateExtension(ctx context.Context, in *Update
 }
 
 func (c *extendRetractApiClient) Stop(ctx context.Context, in *ExtendRetractStopRequest, opts ...grpc.CallOption) (*Extension, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Extension)
-	err := c.cc.Invoke(ctx, ExtendRetractApi_Stop_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ExtendRetractApi_Stop_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -78,20 +83,22 @@ func (c *extendRetractApiClient) Stop(ctx context.Context, in *ExtendRetractStop
 }
 
 func (c *extendRetractApiClient) CreateExtensionPreset(ctx context.Context, in *CreateExtensionPresetRequest, opts ...grpc.CallOption) (*ExtensionPreset, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ExtensionPreset)
-	err := c.cc.Invoke(ctx, ExtendRetractApi_CreateExtensionPreset_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ExtendRetractApi_CreateExtensionPreset_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *extendRetractApiClient) PullExtensions(ctx context.Context, in *PullExtensionsRequest, opts ...grpc.CallOption) (ExtendRetractApi_PullExtensionsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ExtendRetractApi_ServiceDesc.Streams[0], ExtendRetractApi_PullExtensions_FullMethodName, opts...)
+func (c *extendRetractApiClient) PullExtensions(ctx context.Context, in *PullExtensionsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullExtensionsResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ExtendRetractApi_ServiceDesc.Streams[0], ExtendRetractApi_PullExtensions_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &extendRetractApiPullExtensionsClient{stream}
+	x := &grpc.GenericClientStream[PullExtensionsRequest, PullExtensionsResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -101,26 +108,14 @@ func (c *extendRetractApiClient) PullExtensions(ctx context.Context, in *PullExt
 	return x, nil
 }
 
-type ExtendRetractApi_PullExtensionsClient interface {
-	Recv() (*PullExtensionsResponse, error)
-	grpc.ClientStream
-}
-
-type extendRetractApiPullExtensionsClient struct {
-	grpc.ClientStream
-}
-
-func (x *extendRetractApiPullExtensionsClient) Recv() (*PullExtensionsResponse, error) {
-	m := new(PullExtensionsResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ExtendRetractApi_PullExtensionsClient = grpc.ServerStreamingClient[PullExtensionsResponse]
 
 // ExtendRetractApiServer is the server API for ExtendRetractApi service.
 // All implementations must embed UnimplementedExtendRetractApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Applicable to devices that can be extended or retracted. Like a tv mount or adjustable desk
 type ExtendRetractApiServer interface {
 	// Set the target state for the device
 	GetExtension(context.Context, *GetExtensionRequest) (*Extension, error)
@@ -131,13 +126,16 @@ type ExtendRetractApiServer interface {
 	// Create a preset for the extension. If no preset extension value is specified then use the current position
 	CreateExtensionPreset(context.Context, *CreateExtensionPresetRequest) (*ExtensionPreset, error)
 	// Get notified of changes to the OnOffState of a device
-	PullExtensions(*PullExtensionsRequest, ExtendRetractApi_PullExtensionsServer) error
+	PullExtensions(*PullExtensionsRequest, grpc.ServerStreamingServer[PullExtensionsResponse]) error
 	mustEmbedUnimplementedExtendRetractApiServer()
 }
 
-// UnimplementedExtendRetractApiServer must be embedded to have forward compatible implementations.
-type UnimplementedExtendRetractApiServer struct {
-}
+// UnimplementedExtendRetractApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedExtendRetractApiServer struct{}
 
 func (UnimplementedExtendRetractApiServer) GetExtension(context.Context, *GetExtensionRequest) (*Extension, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetExtension not implemented")
@@ -151,10 +149,11 @@ func (UnimplementedExtendRetractApiServer) Stop(context.Context, *ExtendRetractS
 func (UnimplementedExtendRetractApiServer) CreateExtensionPreset(context.Context, *CreateExtensionPresetRequest) (*ExtensionPreset, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateExtensionPreset not implemented")
 }
-func (UnimplementedExtendRetractApiServer) PullExtensions(*PullExtensionsRequest, ExtendRetractApi_PullExtensionsServer) error {
+func (UnimplementedExtendRetractApiServer) PullExtensions(*PullExtensionsRequest, grpc.ServerStreamingServer[PullExtensionsResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullExtensions not implemented")
 }
 func (UnimplementedExtendRetractApiServer) mustEmbedUnimplementedExtendRetractApiServer() {}
+func (UnimplementedExtendRetractApiServer) testEmbeddedByValue()                          {}
 
 // UnsafeExtendRetractApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ExtendRetractApiServer will
@@ -164,6 +163,13 @@ type UnsafeExtendRetractApiServer interface {
 }
 
 func RegisterExtendRetractApiServer(s grpc.ServiceRegistrar, srv ExtendRetractApiServer) {
+	// If the following call pancis, it indicates UnimplementedExtendRetractApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ExtendRetractApi_ServiceDesc, srv)
 }
 
@@ -244,21 +250,11 @@ func _ExtendRetractApi_PullExtensions_Handler(srv interface{}, stream grpc.Serve
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ExtendRetractApiServer).PullExtensions(m, &extendRetractApiPullExtensionsServer{stream})
+	return srv.(ExtendRetractApiServer).PullExtensions(m, &grpc.GenericServerStream[PullExtensionsRequest, PullExtensionsResponse]{ServerStream: stream})
 }
 
-type ExtendRetractApi_PullExtensionsServer interface {
-	Send(*PullExtensionsResponse) error
-	grpc.ServerStream
-}
-
-type extendRetractApiPullExtensionsServer struct {
-	grpc.ServerStream
-}
-
-func (x *extendRetractApiPullExtensionsServer) Send(m *PullExtensionsResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ExtendRetractApi_PullExtensionsServer = grpc.ServerStreamingServer[PullExtensionsResponse]
 
 // ExtendRetractApi_ServiceDesc is the grpc.ServiceDesc for ExtendRetractApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -301,6 +297,8 @@ const (
 // ExtendRetractInfoClient is the client API for ExtendRetractInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ExtendRetractInfoClient interface {
 	// Get information about how a named device implements Extension features
 	DescribeExtension(ctx context.Context, in *DescribeExtensionRequest, opts ...grpc.CallOption) (*ExtensionSupport, error)
@@ -315,8 +313,9 @@ func NewExtendRetractInfoClient(cc grpc.ClientConnInterface) ExtendRetractInfoCl
 }
 
 func (c *extendRetractInfoClient) DescribeExtension(ctx context.Context, in *DescribeExtensionRequest, opts ...grpc.CallOption) (*ExtensionSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ExtensionSupport)
-	err := c.cc.Invoke(ctx, ExtendRetractInfo_DescribeExtension_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ExtendRetractInfo_DescribeExtension_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -325,21 +324,27 @@ func (c *extendRetractInfoClient) DescribeExtension(ctx context.Context, in *Des
 
 // ExtendRetractInfoServer is the server API for ExtendRetractInfo service.
 // All implementations must embed UnimplementedExtendRetractInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ExtendRetractInfoServer interface {
 	// Get information about how a named device implements Extension features
 	DescribeExtension(context.Context, *DescribeExtensionRequest) (*ExtensionSupport, error)
 	mustEmbedUnimplementedExtendRetractInfoServer()
 }
 
-// UnimplementedExtendRetractInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedExtendRetractInfoServer struct {
-}
+// UnimplementedExtendRetractInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedExtendRetractInfoServer struct{}
 
 func (UnimplementedExtendRetractInfoServer) DescribeExtension(context.Context, *DescribeExtensionRequest) (*ExtensionSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeExtension not implemented")
 }
 func (UnimplementedExtendRetractInfoServer) mustEmbedUnimplementedExtendRetractInfoServer() {}
+func (UnimplementedExtendRetractInfoServer) testEmbeddedByValue()                           {}
 
 // UnsafeExtendRetractInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ExtendRetractInfoServer will
@@ -349,6 +354,13 @@ type UnsafeExtendRetractInfoServer interface {
 }
 
 func RegisterExtendRetractInfoServer(s grpc.ServiceRegistrar, srv ExtendRetractInfoServer) {
+	// If the following call pancis, it indicates UnimplementedExtendRetractInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ExtendRetractInfo_ServiceDesc, srv)
 }
 
