@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/fan_speed.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	FanSpeedApi_GetFanSpeed_FullMethodName              = "/smartcore.traits.FanSpeedApi/GetFanSpeed"
@@ -28,13 +28,15 @@ const (
 // FanSpeedApiClient is the client API for FanSpeedApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Trait for devices that can control the speed of a fan.
 type FanSpeedApiClient interface {
 	// GetFanSpeed returns the fan speed for the device.
 	GetFanSpeed(ctx context.Context, in *GetFanSpeedRequest, opts ...grpc.CallOption) (*FanSpeed, error)
 	// UpdateFanSpeed updates the fan speed for the device.
 	UpdateFanSpeed(ctx context.Context, in *UpdateFanSpeedRequest, opts ...grpc.CallOption) (*FanSpeed, error)
 	// PullFanSpeed returns a stream of changes to the current fan speed for the device.
-	PullFanSpeed(ctx context.Context, in *PullFanSpeedRequest, opts ...grpc.CallOption) (FanSpeedApi_PullFanSpeedClient, error)
+	PullFanSpeed(ctx context.Context, in *PullFanSpeedRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullFanSpeedResponse], error)
 	// ReverseFanSpeedDirection reverses the fan speed direction.
 	// If the fan is spinning FORWARD it will be changed to BACKWARD, etc.
 	ReverseFanSpeedDirection(ctx context.Context, in *ReverseFanSpeedDirectionRequest, opts ...grpc.CallOption) (*FanSpeed, error)
@@ -49,8 +51,9 @@ func NewFanSpeedApiClient(cc grpc.ClientConnInterface) FanSpeedApiClient {
 }
 
 func (c *fanSpeedApiClient) GetFanSpeed(ctx context.Context, in *GetFanSpeedRequest, opts ...grpc.CallOption) (*FanSpeed, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FanSpeed)
-	err := c.cc.Invoke(ctx, FanSpeedApi_GetFanSpeed_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, FanSpeedApi_GetFanSpeed_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -58,20 +61,22 @@ func (c *fanSpeedApiClient) GetFanSpeed(ctx context.Context, in *GetFanSpeedRequ
 }
 
 func (c *fanSpeedApiClient) UpdateFanSpeed(ctx context.Context, in *UpdateFanSpeedRequest, opts ...grpc.CallOption) (*FanSpeed, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FanSpeed)
-	err := c.cc.Invoke(ctx, FanSpeedApi_UpdateFanSpeed_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, FanSpeedApi_UpdateFanSpeed_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *fanSpeedApiClient) PullFanSpeed(ctx context.Context, in *PullFanSpeedRequest, opts ...grpc.CallOption) (FanSpeedApi_PullFanSpeedClient, error) {
-	stream, err := c.cc.NewStream(ctx, &FanSpeedApi_ServiceDesc.Streams[0], FanSpeedApi_PullFanSpeed_FullMethodName, opts...)
+func (c *fanSpeedApiClient) PullFanSpeed(ctx context.Context, in *PullFanSpeedRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullFanSpeedResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &FanSpeedApi_ServiceDesc.Streams[0], FanSpeedApi_PullFanSpeed_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &fanSpeedApiPullFanSpeedClient{stream}
+	x := &grpc.GenericClientStream[PullFanSpeedRequest, PullFanSpeedResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -81,26 +86,13 @@ func (c *fanSpeedApiClient) PullFanSpeed(ctx context.Context, in *PullFanSpeedRe
 	return x, nil
 }
 
-type FanSpeedApi_PullFanSpeedClient interface {
-	Recv() (*PullFanSpeedResponse, error)
-	grpc.ClientStream
-}
-
-type fanSpeedApiPullFanSpeedClient struct {
-	grpc.ClientStream
-}
-
-func (x *fanSpeedApiPullFanSpeedClient) Recv() (*PullFanSpeedResponse, error) {
-	m := new(PullFanSpeedResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type FanSpeedApi_PullFanSpeedClient = grpc.ServerStreamingClient[PullFanSpeedResponse]
 
 func (c *fanSpeedApiClient) ReverseFanSpeedDirection(ctx context.Context, in *ReverseFanSpeedDirectionRequest, opts ...grpc.CallOption) (*FanSpeed, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FanSpeed)
-	err := c.cc.Invoke(ctx, FanSpeedApi_ReverseFanSpeedDirection_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, FanSpeedApi_ReverseFanSpeedDirection_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -109,23 +101,28 @@ func (c *fanSpeedApiClient) ReverseFanSpeedDirection(ctx context.Context, in *Re
 
 // FanSpeedApiServer is the server API for FanSpeedApi service.
 // All implementations must embed UnimplementedFanSpeedApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Trait for devices that can control the speed of a fan.
 type FanSpeedApiServer interface {
 	// GetFanSpeed returns the fan speed for the device.
 	GetFanSpeed(context.Context, *GetFanSpeedRequest) (*FanSpeed, error)
 	// UpdateFanSpeed updates the fan speed for the device.
 	UpdateFanSpeed(context.Context, *UpdateFanSpeedRequest) (*FanSpeed, error)
 	// PullFanSpeed returns a stream of changes to the current fan speed for the device.
-	PullFanSpeed(*PullFanSpeedRequest, FanSpeedApi_PullFanSpeedServer) error
+	PullFanSpeed(*PullFanSpeedRequest, grpc.ServerStreamingServer[PullFanSpeedResponse]) error
 	// ReverseFanSpeedDirection reverses the fan speed direction.
 	// If the fan is spinning FORWARD it will be changed to BACKWARD, etc.
 	ReverseFanSpeedDirection(context.Context, *ReverseFanSpeedDirectionRequest) (*FanSpeed, error)
 	mustEmbedUnimplementedFanSpeedApiServer()
 }
 
-// UnimplementedFanSpeedApiServer must be embedded to have forward compatible implementations.
-type UnimplementedFanSpeedApiServer struct {
-}
+// UnimplementedFanSpeedApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedFanSpeedApiServer struct{}
 
 func (UnimplementedFanSpeedApiServer) GetFanSpeed(context.Context, *GetFanSpeedRequest) (*FanSpeed, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetFanSpeed not implemented")
@@ -133,13 +130,14 @@ func (UnimplementedFanSpeedApiServer) GetFanSpeed(context.Context, *GetFanSpeedR
 func (UnimplementedFanSpeedApiServer) UpdateFanSpeed(context.Context, *UpdateFanSpeedRequest) (*FanSpeed, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateFanSpeed not implemented")
 }
-func (UnimplementedFanSpeedApiServer) PullFanSpeed(*PullFanSpeedRequest, FanSpeedApi_PullFanSpeedServer) error {
+func (UnimplementedFanSpeedApiServer) PullFanSpeed(*PullFanSpeedRequest, grpc.ServerStreamingServer[PullFanSpeedResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullFanSpeed not implemented")
 }
 func (UnimplementedFanSpeedApiServer) ReverseFanSpeedDirection(context.Context, *ReverseFanSpeedDirectionRequest) (*FanSpeed, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ReverseFanSpeedDirection not implemented")
 }
 func (UnimplementedFanSpeedApiServer) mustEmbedUnimplementedFanSpeedApiServer() {}
+func (UnimplementedFanSpeedApiServer) testEmbeddedByValue()                     {}
 
 // UnsafeFanSpeedApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to FanSpeedApiServer will
@@ -149,6 +147,13 @@ type UnsafeFanSpeedApiServer interface {
 }
 
 func RegisterFanSpeedApiServer(s grpc.ServiceRegistrar, srv FanSpeedApiServer) {
+	// If the following call pancis, it indicates UnimplementedFanSpeedApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&FanSpeedApi_ServiceDesc, srv)
 }
 
@@ -193,21 +198,11 @@ func _FanSpeedApi_PullFanSpeed_Handler(srv interface{}, stream grpc.ServerStream
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(FanSpeedApiServer).PullFanSpeed(m, &fanSpeedApiPullFanSpeedServer{stream})
+	return srv.(FanSpeedApiServer).PullFanSpeed(m, &grpc.GenericServerStream[PullFanSpeedRequest, PullFanSpeedResponse]{ServerStream: stream})
 }
 
-type FanSpeedApi_PullFanSpeedServer interface {
-	Send(*PullFanSpeedResponse) error
-	grpc.ServerStream
-}
-
-type fanSpeedApiPullFanSpeedServer struct {
-	grpc.ServerStream
-}
-
-func (x *fanSpeedApiPullFanSpeedServer) Send(m *PullFanSpeedResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type FanSpeedApi_PullFanSpeedServer = grpc.ServerStreamingServer[PullFanSpeedResponse]
 
 func _FanSpeedApi_ReverseFanSpeedDirection_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ReverseFanSpeedDirectionRequest)
@@ -264,6 +259,8 @@ const (
 // FanSpeedInfoClient is the client API for FanSpeedInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type FanSpeedInfoClient interface {
 	// Get information about how a named device implements FanSpeed features
 	DescribeFanSpeed(ctx context.Context, in *DescribeFanSpeedRequest, opts ...grpc.CallOption) (*FanSpeedSupport, error)
@@ -278,8 +275,9 @@ func NewFanSpeedInfoClient(cc grpc.ClientConnInterface) FanSpeedInfoClient {
 }
 
 func (c *fanSpeedInfoClient) DescribeFanSpeed(ctx context.Context, in *DescribeFanSpeedRequest, opts ...grpc.CallOption) (*FanSpeedSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FanSpeedSupport)
-	err := c.cc.Invoke(ctx, FanSpeedInfo_DescribeFanSpeed_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, FanSpeedInfo_DescribeFanSpeed_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -288,21 +286,27 @@ func (c *fanSpeedInfoClient) DescribeFanSpeed(ctx context.Context, in *DescribeF
 
 // FanSpeedInfoServer is the server API for FanSpeedInfo service.
 // All implementations must embed UnimplementedFanSpeedInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type FanSpeedInfoServer interface {
 	// Get information about how a named device implements FanSpeed features
 	DescribeFanSpeed(context.Context, *DescribeFanSpeedRequest) (*FanSpeedSupport, error)
 	mustEmbedUnimplementedFanSpeedInfoServer()
 }
 
-// UnimplementedFanSpeedInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedFanSpeedInfoServer struct {
-}
+// UnimplementedFanSpeedInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedFanSpeedInfoServer struct{}
 
 func (UnimplementedFanSpeedInfoServer) DescribeFanSpeed(context.Context, *DescribeFanSpeedRequest) (*FanSpeedSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeFanSpeed not implemented")
 }
 func (UnimplementedFanSpeedInfoServer) mustEmbedUnimplementedFanSpeedInfoServer() {}
+func (UnimplementedFanSpeedInfoServer) testEmbeddedByValue()                      {}
 
 // UnsafeFanSpeedInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to FanSpeedInfoServer will
@@ -312,6 +316,13 @@ type UnsafeFanSpeedInfoServer interface {
 }
 
 func RegisterFanSpeedInfoServer(s grpc.ServiceRegistrar, srv FanSpeedInfoServer) {
+	// If the following call pancis, it indicates UnimplementedFanSpeedInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&FanSpeedInfo_ServiceDesc, srv)
 }
 
