@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/hail.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	HailApi_CreateHail_FullMethodName = "/smartcore.traits.HailApi/CreateHail"
@@ -31,6 +31,14 @@ const (
 // HailApiClient is the client API for HailApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// HailApi describes devices that may be asked to transport passengers from one location to another.
+// To interact with these devices you create a hail, a description of the origin and destination for your journey.
+// The device keeps you updated on the preparation or progress of your journey by updating the fields in Hail.
+//
+// HailApi applies to devices like lifts or taxis, where the transportation is initiated by the caller. It does not
+// apply to scheduled/regular transportation - like busses or escalators - where the presence of the passenger is not
+// required.
 type HailApiClient interface {
 	// Create a new hail against the device.
 	// Creating a hail instructs the device that a passenger would like to be transported from one location to another.
@@ -48,11 +56,11 @@ type HailApiClient interface {
 	// Delete a hail by id.
 	DeleteHail(ctx context.Context, in *DeleteHailRequest, opts ...grpc.CallOption) (*DeleteHailResponse, error)
 	// Subscribe to changes in a specific hail.
-	PullHail(ctx context.Context, in *PullHailRequest, opts ...grpc.CallOption) (HailApi_PullHailClient, error)
+	PullHail(ctx context.Context, in *PullHailRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullHailResponse], error)
 	// List known hails.
 	ListHails(ctx context.Context, in *ListHailsRequest, opts ...grpc.CallOption) (*ListHailsResponse, error)
 	// Subscribe to changes in the known hails.
-	PullHails(ctx context.Context, in *PullHailsRequest, opts ...grpc.CallOption) (HailApi_PullHailsClient, error)
+	PullHails(ctx context.Context, in *PullHailsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullHailsResponse], error)
 }
 
 type hailApiClient struct {
@@ -64,8 +72,9 @@ func NewHailApiClient(cc grpc.ClientConnInterface) HailApiClient {
 }
 
 func (c *hailApiClient) CreateHail(ctx context.Context, in *CreateHailRequest, opts ...grpc.CallOption) (*Hail, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Hail)
-	err := c.cc.Invoke(ctx, HailApi_CreateHail_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, HailApi_CreateHail_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -73,8 +82,9 @@ func (c *hailApiClient) CreateHail(ctx context.Context, in *CreateHailRequest, o
 }
 
 func (c *hailApiClient) GetHail(ctx context.Context, in *GetHailRequest, opts ...grpc.CallOption) (*Hail, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Hail)
-	err := c.cc.Invoke(ctx, HailApi_GetHail_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, HailApi_GetHail_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -82,8 +92,9 @@ func (c *hailApiClient) GetHail(ctx context.Context, in *GetHailRequest, opts ..
 }
 
 func (c *hailApiClient) UpdateHail(ctx context.Context, in *UpdateHailRequest, opts ...grpc.CallOption) (*Hail, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Hail)
-	err := c.cc.Invoke(ctx, HailApi_UpdateHail_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, HailApi_UpdateHail_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -91,20 +102,22 @@ func (c *hailApiClient) UpdateHail(ctx context.Context, in *UpdateHailRequest, o
 }
 
 func (c *hailApiClient) DeleteHail(ctx context.Context, in *DeleteHailRequest, opts ...grpc.CallOption) (*DeleteHailResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteHailResponse)
-	err := c.cc.Invoke(ctx, HailApi_DeleteHail_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, HailApi_DeleteHail_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *hailApiClient) PullHail(ctx context.Context, in *PullHailRequest, opts ...grpc.CallOption) (HailApi_PullHailClient, error) {
-	stream, err := c.cc.NewStream(ctx, &HailApi_ServiceDesc.Streams[0], HailApi_PullHail_FullMethodName, opts...)
+func (c *hailApiClient) PullHail(ctx context.Context, in *PullHailRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullHailResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &HailApi_ServiceDesc.Streams[0], HailApi_PullHail_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &hailApiPullHailClient{stream}
+	x := &grpc.GenericClientStream[PullHailRequest, PullHailResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -114,38 +127,26 @@ func (c *hailApiClient) PullHail(ctx context.Context, in *PullHailRequest, opts 
 	return x, nil
 }
 
-type HailApi_PullHailClient interface {
-	Recv() (*PullHailResponse, error)
-	grpc.ClientStream
-}
-
-type hailApiPullHailClient struct {
-	grpc.ClientStream
-}
-
-func (x *hailApiPullHailClient) Recv() (*PullHailResponse, error) {
-	m := new(PullHailResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type HailApi_PullHailClient = grpc.ServerStreamingClient[PullHailResponse]
 
 func (c *hailApiClient) ListHails(ctx context.Context, in *ListHailsRequest, opts ...grpc.CallOption) (*ListHailsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListHailsResponse)
-	err := c.cc.Invoke(ctx, HailApi_ListHails_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, HailApi_ListHails_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *hailApiClient) PullHails(ctx context.Context, in *PullHailsRequest, opts ...grpc.CallOption) (HailApi_PullHailsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &HailApi_ServiceDesc.Streams[1], HailApi_PullHails_FullMethodName, opts...)
+func (c *hailApiClient) PullHails(ctx context.Context, in *PullHailsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullHailsResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &HailApi_ServiceDesc.Streams[1], HailApi_PullHails_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &hailApiPullHailsClient{stream}
+	x := &grpc.GenericClientStream[PullHailsRequest, PullHailsResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -155,26 +156,20 @@ func (c *hailApiClient) PullHails(ctx context.Context, in *PullHailsRequest, opt
 	return x, nil
 }
 
-type HailApi_PullHailsClient interface {
-	Recv() (*PullHailsResponse, error)
-	grpc.ClientStream
-}
-
-type hailApiPullHailsClient struct {
-	grpc.ClientStream
-}
-
-func (x *hailApiPullHailsClient) Recv() (*PullHailsResponse, error) {
-	m := new(PullHailsResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type HailApi_PullHailsClient = grpc.ServerStreamingClient[PullHailsResponse]
 
 // HailApiServer is the server API for HailApi service.
 // All implementations must embed UnimplementedHailApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// HailApi describes devices that may be asked to transport passengers from one location to another.
+// To interact with these devices you create a hail, a description of the origin and destination for your journey.
+// The device keeps you updated on the preparation or progress of your journey by updating the fields in Hail.
+//
+// HailApi applies to devices like lifts or taxis, where the transportation is initiated by the caller. It does not
+// apply to scheduled/regular transportation - like busses or escalators - where the presence of the passenger is not
+// required.
 type HailApiServer interface {
 	// Create a new hail against the device.
 	// Creating a hail instructs the device that a passenger would like to be transported from one location to another.
@@ -192,17 +187,20 @@ type HailApiServer interface {
 	// Delete a hail by id.
 	DeleteHail(context.Context, *DeleteHailRequest) (*DeleteHailResponse, error)
 	// Subscribe to changes in a specific hail.
-	PullHail(*PullHailRequest, HailApi_PullHailServer) error
+	PullHail(*PullHailRequest, grpc.ServerStreamingServer[PullHailResponse]) error
 	// List known hails.
 	ListHails(context.Context, *ListHailsRequest) (*ListHailsResponse, error)
 	// Subscribe to changes in the known hails.
-	PullHails(*PullHailsRequest, HailApi_PullHailsServer) error
+	PullHails(*PullHailsRequest, grpc.ServerStreamingServer[PullHailsResponse]) error
 	mustEmbedUnimplementedHailApiServer()
 }
 
-// UnimplementedHailApiServer must be embedded to have forward compatible implementations.
-type UnimplementedHailApiServer struct {
-}
+// UnimplementedHailApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedHailApiServer struct{}
 
 func (UnimplementedHailApiServer) CreateHail(context.Context, *CreateHailRequest) (*Hail, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateHail not implemented")
@@ -216,16 +214,17 @@ func (UnimplementedHailApiServer) UpdateHail(context.Context, *UpdateHailRequest
 func (UnimplementedHailApiServer) DeleteHail(context.Context, *DeleteHailRequest) (*DeleteHailResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DeleteHail not implemented")
 }
-func (UnimplementedHailApiServer) PullHail(*PullHailRequest, HailApi_PullHailServer) error {
+func (UnimplementedHailApiServer) PullHail(*PullHailRequest, grpc.ServerStreamingServer[PullHailResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullHail not implemented")
 }
 func (UnimplementedHailApiServer) ListHails(context.Context, *ListHailsRequest) (*ListHailsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListHails not implemented")
 }
-func (UnimplementedHailApiServer) PullHails(*PullHailsRequest, HailApi_PullHailsServer) error {
+func (UnimplementedHailApiServer) PullHails(*PullHailsRequest, grpc.ServerStreamingServer[PullHailsResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullHails not implemented")
 }
 func (UnimplementedHailApiServer) mustEmbedUnimplementedHailApiServer() {}
+func (UnimplementedHailApiServer) testEmbeddedByValue()                 {}
 
 // UnsafeHailApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to HailApiServer will
@@ -235,6 +234,13 @@ type UnsafeHailApiServer interface {
 }
 
 func RegisterHailApiServer(s grpc.ServiceRegistrar, srv HailApiServer) {
+	// If the following call pancis, it indicates UnimplementedHailApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&HailApi_ServiceDesc, srv)
 }
 
@@ -315,21 +321,11 @@ func _HailApi_PullHail_Handler(srv interface{}, stream grpc.ServerStream) error 
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(HailApiServer).PullHail(m, &hailApiPullHailServer{stream})
+	return srv.(HailApiServer).PullHail(m, &grpc.GenericServerStream[PullHailRequest, PullHailResponse]{ServerStream: stream})
 }
 
-type HailApi_PullHailServer interface {
-	Send(*PullHailResponse) error
-	grpc.ServerStream
-}
-
-type hailApiPullHailServer struct {
-	grpc.ServerStream
-}
-
-func (x *hailApiPullHailServer) Send(m *PullHailResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type HailApi_PullHailServer = grpc.ServerStreamingServer[PullHailResponse]
 
 func _HailApi_ListHails_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ListHailsRequest)
@@ -354,21 +350,11 @@ func _HailApi_PullHails_Handler(srv interface{}, stream grpc.ServerStream) error
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(HailApiServer).PullHails(m, &hailApiPullHailsServer{stream})
+	return srv.(HailApiServer).PullHails(m, &grpc.GenericServerStream[PullHailsRequest, PullHailsResponse]{ServerStream: stream})
 }
 
-type HailApi_PullHailsServer interface {
-	Send(*PullHailsResponse) error
-	grpc.ServerStream
-}
-
-type hailApiPullHailsServer struct {
-	grpc.ServerStream
-}
-
-func (x *hailApiPullHailsServer) Send(m *PullHailsResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type HailApi_PullHailsServer = grpc.ServerStreamingServer[PullHailsResponse]
 
 // HailApi_ServiceDesc is the grpc.ServiceDesc for HailApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -413,8 +399,6 @@ var HailApi_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "traits/hail.proto",
 }
 
-const ()
-
 // HailInfoClient is the client API for HailInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
@@ -431,16 +415,20 @@ func NewHailInfoClient(cc grpc.ClientConnInterface) HailInfoClient {
 
 // HailInfoServer is the server API for HailInfo service.
 // All implementations must embed UnimplementedHailInfoServer
-// for forward compatibility
+// for forward compatibility.
 type HailInfoServer interface {
 	mustEmbedUnimplementedHailInfoServer()
 }
 
-// UnimplementedHailInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedHailInfoServer struct {
-}
+// UnimplementedHailInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedHailInfoServer struct{}
 
 func (UnimplementedHailInfoServer) mustEmbedUnimplementedHailInfoServer() {}
+func (UnimplementedHailInfoServer) testEmbeddedByValue()                  {}
 
 // UnsafeHailInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to HailInfoServer will
@@ -450,6 +438,13 @@ type UnsafeHailInfoServer interface {
 }
 
 func RegisterHailInfoServer(s grpc.ServiceRegistrar, srv HailInfoServer) {
+	// If the following call pancis, it indicates UnimplementedHailInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&HailInfo_ServiceDesc, srv)
 }
 
