@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/input_select.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	InputSelectApi_UpdateInput_FullMethodName = "/smartcore.traits.InputSelectApi/UpdateInput"
@@ -27,13 +27,16 @@ const (
 // InputSelectApiClient is the client API for InputSelectApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// A trait to control the input for a device. Typically related to AV devices like displays or presentation switchers.
+// A more complicated device might have multiple outputs, this trait also supports this.
 type InputSelectApiClient interface {
 	// request that the input be selected
 	UpdateInput(ctx context.Context, in *UpdateInputRequest, opts ...grpc.CallOption) (*Input, error)
 	// get the current value of the selected input
 	GetInput(ctx context.Context, in *GetInputRequest, opts ...grpc.CallOption) (*Input, error)
 	// request updates to changes in the selected input
-	PullInput(ctx context.Context, in *PullInputRequest, opts ...grpc.CallOption) (InputSelectApi_PullInputClient, error)
+	PullInput(ctx context.Context, in *PullInputRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullInputResponse], error)
 }
 
 type inputSelectApiClient struct {
@@ -45,8 +48,9 @@ func NewInputSelectApiClient(cc grpc.ClientConnInterface) InputSelectApiClient {
 }
 
 func (c *inputSelectApiClient) UpdateInput(ctx context.Context, in *UpdateInputRequest, opts ...grpc.CallOption) (*Input, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Input)
-	err := c.cc.Invoke(ctx, InputSelectApi_UpdateInput_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, InputSelectApi_UpdateInput_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,20 +58,22 @@ func (c *inputSelectApiClient) UpdateInput(ctx context.Context, in *UpdateInputR
 }
 
 func (c *inputSelectApiClient) GetInput(ctx context.Context, in *GetInputRequest, opts ...grpc.CallOption) (*Input, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Input)
-	err := c.cc.Invoke(ctx, InputSelectApi_GetInput_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, InputSelectApi_GetInput_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *inputSelectApiClient) PullInput(ctx context.Context, in *PullInputRequest, opts ...grpc.CallOption) (InputSelectApi_PullInputClient, error) {
-	stream, err := c.cc.NewStream(ctx, &InputSelectApi_ServiceDesc.Streams[0], InputSelectApi_PullInput_FullMethodName, opts...)
+func (c *inputSelectApiClient) PullInput(ctx context.Context, in *PullInputRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullInputResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &InputSelectApi_ServiceDesc.Streams[0], InputSelectApi_PullInput_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &inputSelectApiPullInputClient{stream}
+	x := &grpc.GenericClientStream[PullInputRequest, PullInputResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -77,39 +83,31 @@ func (c *inputSelectApiClient) PullInput(ctx context.Context, in *PullInputReque
 	return x, nil
 }
 
-type InputSelectApi_PullInputClient interface {
-	Recv() (*PullInputResponse, error)
-	grpc.ClientStream
-}
-
-type inputSelectApiPullInputClient struct {
-	grpc.ClientStream
-}
-
-func (x *inputSelectApiPullInputClient) Recv() (*PullInputResponse, error) {
-	m := new(PullInputResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type InputSelectApi_PullInputClient = grpc.ServerStreamingClient[PullInputResponse]
 
 // InputSelectApiServer is the server API for InputSelectApi service.
 // All implementations must embed UnimplementedInputSelectApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// A trait to control the input for a device. Typically related to AV devices like displays or presentation switchers.
+// A more complicated device might have multiple outputs, this trait also supports this.
 type InputSelectApiServer interface {
 	// request that the input be selected
 	UpdateInput(context.Context, *UpdateInputRequest) (*Input, error)
 	// get the current value of the selected input
 	GetInput(context.Context, *GetInputRequest) (*Input, error)
 	// request updates to changes in the selected input
-	PullInput(*PullInputRequest, InputSelectApi_PullInputServer) error
+	PullInput(*PullInputRequest, grpc.ServerStreamingServer[PullInputResponse]) error
 	mustEmbedUnimplementedInputSelectApiServer()
 }
 
-// UnimplementedInputSelectApiServer must be embedded to have forward compatible implementations.
-type UnimplementedInputSelectApiServer struct {
-}
+// UnimplementedInputSelectApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedInputSelectApiServer struct{}
 
 func (UnimplementedInputSelectApiServer) UpdateInput(context.Context, *UpdateInputRequest) (*Input, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateInput not implemented")
@@ -117,10 +115,11 @@ func (UnimplementedInputSelectApiServer) UpdateInput(context.Context, *UpdateInp
 func (UnimplementedInputSelectApiServer) GetInput(context.Context, *GetInputRequest) (*Input, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetInput not implemented")
 }
-func (UnimplementedInputSelectApiServer) PullInput(*PullInputRequest, InputSelectApi_PullInputServer) error {
+func (UnimplementedInputSelectApiServer) PullInput(*PullInputRequest, grpc.ServerStreamingServer[PullInputResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullInput not implemented")
 }
 func (UnimplementedInputSelectApiServer) mustEmbedUnimplementedInputSelectApiServer() {}
+func (UnimplementedInputSelectApiServer) testEmbeddedByValue()                        {}
 
 // UnsafeInputSelectApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to InputSelectApiServer will
@@ -130,6 +129,13 @@ type UnsafeInputSelectApiServer interface {
 }
 
 func RegisterInputSelectApiServer(s grpc.ServiceRegistrar, srv InputSelectApiServer) {
+	// If the following call pancis, it indicates UnimplementedInputSelectApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&InputSelectApi_ServiceDesc, srv)
 }
 
@@ -174,21 +180,11 @@ func _InputSelectApi_PullInput_Handler(srv interface{}, stream grpc.ServerStream
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(InputSelectApiServer).PullInput(m, &inputSelectApiPullInputServer{stream})
+	return srv.(InputSelectApiServer).PullInput(m, &grpc.GenericServerStream[PullInputRequest, PullInputResponse]{ServerStream: stream})
 }
 
-type InputSelectApi_PullInputServer interface {
-	Send(*PullInputResponse) error
-	grpc.ServerStream
-}
-
-type inputSelectApiPullInputServer struct {
-	grpc.ServerStream
-}
-
-func (x *inputSelectApiPullInputServer) Send(m *PullInputResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type InputSelectApi_PullInputServer = grpc.ServerStreamingServer[PullInputResponse]
 
 // InputSelectApi_ServiceDesc is the grpc.ServiceDesc for InputSelectApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -223,6 +219,8 @@ const (
 // InputSelectInfoClient is the client API for InputSelectInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type InputSelectInfoClient interface {
 	// Get information about how a named device implements Input features
 	DescribeInput(ctx context.Context, in *DescribeInputRequest, opts ...grpc.CallOption) (*InputSupport, error)
@@ -237,8 +235,9 @@ func NewInputSelectInfoClient(cc grpc.ClientConnInterface) InputSelectInfoClient
 }
 
 func (c *inputSelectInfoClient) DescribeInput(ctx context.Context, in *DescribeInputRequest, opts ...grpc.CallOption) (*InputSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(InputSupport)
-	err := c.cc.Invoke(ctx, InputSelectInfo_DescribeInput_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, InputSelectInfo_DescribeInput_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -247,21 +246,27 @@ func (c *inputSelectInfoClient) DescribeInput(ctx context.Context, in *DescribeI
 
 // InputSelectInfoServer is the server API for InputSelectInfo service.
 // All implementations must embed UnimplementedInputSelectInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type InputSelectInfoServer interface {
 	// Get information about how a named device implements Input features
 	DescribeInput(context.Context, *DescribeInputRequest) (*InputSupport, error)
 	mustEmbedUnimplementedInputSelectInfoServer()
 }
 
-// UnimplementedInputSelectInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedInputSelectInfoServer struct {
-}
+// UnimplementedInputSelectInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedInputSelectInfoServer struct{}
 
 func (UnimplementedInputSelectInfoServer) DescribeInput(context.Context, *DescribeInputRequest) (*InputSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeInput not implemented")
 }
 func (UnimplementedInputSelectInfoServer) mustEmbedUnimplementedInputSelectInfoServer() {}
+func (UnimplementedInputSelectInfoServer) testEmbeddedByValue()                         {}
 
 // UnsafeInputSelectInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to InputSelectInfoServer will
@@ -271,6 +276,13 @@ type UnsafeInputSelectInfoServer interface {
 }
 
 func RegisterInputSelectInfoServer(s grpc.ServiceRegistrar, srv InputSelectInfoServer) {
+	// If the following call pancis, it indicates UnimplementedInputSelectInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&InputSelectInfo_ServiceDesc, srv)
 }
 
