@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/light.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	LightApi_UpdateBrightness_FullMethodName = "/smartcore.traits.LightApi/UpdateBrightness"
@@ -27,13 +27,16 @@ const (
 // LightApiClient is the client API for LightApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Control the light output of a device. Brightness values should be normalised to 0-100 as a percentage, though the
+// device may not support all steps within. See BrightnessSupport for step values of the device.
 type LightApiClient interface {
 	// request that the brightness be changed
 	UpdateBrightness(ctx context.Context, in *UpdateBrightnessRequest, opts ...grpc.CallOption) (*Brightness, error)
 	// get the current value of the range
 	GetBrightness(ctx context.Context, in *GetBrightnessRequest, opts ...grpc.CallOption) (*Brightness, error)
 	// request updates to changes in the range value
-	PullBrightness(ctx context.Context, in *PullBrightnessRequest, opts ...grpc.CallOption) (LightApi_PullBrightnessClient, error)
+	PullBrightness(ctx context.Context, in *PullBrightnessRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullBrightnessResponse], error)
 }
 
 type lightApiClient struct {
@@ -45,8 +48,9 @@ func NewLightApiClient(cc grpc.ClientConnInterface) LightApiClient {
 }
 
 func (c *lightApiClient) UpdateBrightness(ctx context.Context, in *UpdateBrightnessRequest, opts ...grpc.CallOption) (*Brightness, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Brightness)
-	err := c.cc.Invoke(ctx, LightApi_UpdateBrightness_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, LightApi_UpdateBrightness_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,20 +58,22 @@ func (c *lightApiClient) UpdateBrightness(ctx context.Context, in *UpdateBrightn
 }
 
 func (c *lightApiClient) GetBrightness(ctx context.Context, in *GetBrightnessRequest, opts ...grpc.CallOption) (*Brightness, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Brightness)
-	err := c.cc.Invoke(ctx, LightApi_GetBrightness_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, LightApi_GetBrightness_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *lightApiClient) PullBrightness(ctx context.Context, in *PullBrightnessRequest, opts ...grpc.CallOption) (LightApi_PullBrightnessClient, error) {
-	stream, err := c.cc.NewStream(ctx, &LightApi_ServiceDesc.Streams[0], LightApi_PullBrightness_FullMethodName, opts...)
+func (c *lightApiClient) PullBrightness(ctx context.Context, in *PullBrightnessRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullBrightnessResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &LightApi_ServiceDesc.Streams[0], LightApi_PullBrightness_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &lightApiPullBrightnessClient{stream}
+	x := &grpc.GenericClientStream[PullBrightnessRequest, PullBrightnessResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -77,39 +83,31 @@ func (c *lightApiClient) PullBrightness(ctx context.Context, in *PullBrightnessR
 	return x, nil
 }
 
-type LightApi_PullBrightnessClient interface {
-	Recv() (*PullBrightnessResponse, error)
-	grpc.ClientStream
-}
-
-type lightApiPullBrightnessClient struct {
-	grpc.ClientStream
-}
-
-func (x *lightApiPullBrightnessClient) Recv() (*PullBrightnessResponse, error) {
-	m := new(PullBrightnessResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type LightApi_PullBrightnessClient = grpc.ServerStreamingClient[PullBrightnessResponse]
 
 // LightApiServer is the server API for LightApi service.
 // All implementations must embed UnimplementedLightApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Control the light output of a device. Brightness values should be normalised to 0-100 as a percentage, though the
+// device may not support all steps within. See BrightnessSupport for step values of the device.
 type LightApiServer interface {
 	// request that the brightness be changed
 	UpdateBrightness(context.Context, *UpdateBrightnessRequest) (*Brightness, error)
 	// get the current value of the range
 	GetBrightness(context.Context, *GetBrightnessRequest) (*Brightness, error)
 	// request updates to changes in the range value
-	PullBrightness(*PullBrightnessRequest, LightApi_PullBrightnessServer) error
+	PullBrightness(*PullBrightnessRequest, grpc.ServerStreamingServer[PullBrightnessResponse]) error
 	mustEmbedUnimplementedLightApiServer()
 }
 
-// UnimplementedLightApiServer must be embedded to have forward compatible implementations.
-type UnimplementedLightApiServer struct {
-}
+// UnimplementedLightApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedLightApiServer struct{}
 
 func (UnimplementedLightApiServer) UpdateBrightness(context.Context, *UpdateBrightnessRequest) (*Brightness, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateBrightness not implemented")
@@ -117,10 +115,11 @@ func (UnimplementedLightApiServer) UpdateBrightness(context.Context, *UpdateBrig
 func (UnimplementedLightApiServer) GetBrightness(context.Context, *GetBrightnessRequest) (*Brightness, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetBrightness not implemented")
 }
-func (UnimplementedLightApiServer) PullBrightness(*PullBrightnessRequest, LightApi_PullBrightnessServer) error {
+func (UnimplementedLightApiServer) PullBrightness(*PullBrightnessRequest, grpc.ServerStreamingServer[PullBrightnessResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullBrightness not implemented")
 }
 func (UnimplementedLightApiServer) mustEmbedUnimplementedLightApiServer() {}
+func (UnimplementedLightApiServer) testEmbeddedByValue()                  {}
 
 // UnsafeLightApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to LightApiServer will
@@ -130,6 +129,13 @@ type UnsafeLightApiServer interface {
 }
 
 func RegisterLightApiServer(s grpc.ServiceRegistrar, srv LightApiServer) {
+	// If the following call pancis, it indicates UnimplementedLightApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&LightApi_ServiceDesc, srv)
 }
 
@@ -174,21 +180,11 @@ func _LightApi_PullBrightness_Handler(srv interface{}, stream grpc.ServerStream)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(LightApiServer).PullBrightness(m, &lightApiPullBrightnessServer{stream})
+	return srv.(LightApiServer).PullBrightness(m, &grpc.GenericServerStream[PullBrightnessRequest, PullBrightnessResponse]{ServerStream: stream})
 }
 
-type LightApi_PullBrightnessServer interface {
-	Send(*PullBrightnessResponse) error
-	grpc.ServerStream
-}
-
-type lightApiPullBrightnessServer struct {
-	grpc.ServerStream
-}
-
-func (x *lightApiPullBrightnessServer) Send(m *PullBrightnessResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type LightApi_PullBrightnessServer = grpc.ServerStreamingServer[PullBrightnessResponse]
 
 // LightApi_ServiceDesc is the grpc.ServiceDesc for LightApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -223,6 +219,8 @@ const (
 // LightInfoClient is the client API for LightInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type LightInfoClient interface {
 	// Get information about how a named device implements Brightness features
 	DescribeBrightness(ctx context.Context, in *DescribeBrightnessRequest, opts ...grpc.CallOption) (*BrightnessSupport, error)
@@ -237,8 +235,9 @@ func NewLightInfoClient(cc grpc.ClientConnInterface) LightInfoClient {
 }
 
 func (c *lightInfoClient) DescribeBrightness(ctx context.Context, in *DescribeBrightnessRequest, opts ...grpc.CallOption) (*BrightnessSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(BrightnessSupport)
-	err := c.cc.Invoke(ctx, LightInfo_DescribeBrightness_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, LightInfo_DescribeBrightness_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -247,21 +246,27 @@ func (c *lightInfoClient) DescribeBrightness(ctx context.Context, in *DescribeBr
 
 // LightInfoServer is the server API for LightInfo service.
 // All implementations must embed UnimplementedLightInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type LightInfoServer interface {
 	// Get information about how a named device implements Brightness features
 	DescribeBrightness(context.Context, *DescribeBrightnessRequest) (*BrightnessSupport, error)
 	mustEmbedUnimplementedLightInfoServer()
 }
 
-// UnimplementedLightInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedLightInfoServer struct {
-}
+// UnimplementedLightInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedLightInfoServer struct{}
 
 func (UnimplementedLightInfoServer) DescribeBrightness(context.Context, *DescribeBrightnessRequest) (*BrightnessSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeBrightness not implemented")
 }
 func (UnimplementedLightInfoServer) mustEmbedUnimplementedLightInfoServer() {}
+func (UnimplementedLightInfoServer) testEmbeddedByValue()                   {}
 
 // UnsafeLightInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to LightInfoServer will
@@ -271,6 +276,13 @@ type UnsafeLightInfoServer interface {
 }
 
 func RegisterLightInfoServer(s grpc.ServiceRegistrar, srv LightInfoServer) {
+	// If the following call pancis, it indicates UnimplementedLightInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&LightInfo_ServiceDesc, srv)
 }
 
