@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/lock_unlock.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	LockUnlockApi_GetLockUnlock_FullMethodName    = "/smartcore.traits.LockUnlockApi/GetLockUnlock"
@@ -27,13 +27,17 @@ const (
 // LockUnlockApiClient is the client API for LockUnlockApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// LockUnlockApi describes devices that can be locked and unlocked, and/or report a locked state.
+// The concept is encoded in the LockUnlock.position, a device may be locked, unlocked, or transitioning into one of
+// these positions.
 type LockUnlockApiClient interface {
 	// Get the current lock position for the device.
 	GetLockUnlock(ctx context.Context, in *GetLockUnlockRequest, opts ...grpc.CallOption) (*LockUnlock, error)
 	// Update the lock position of the device.
 	UpdateLockUnlock(ctx context.Context, in *UpdateLockUnlockRequest, opts ...grpc.CallOption) (*LockUnlock, error)
 	// Subscribe to changes in the lock position for the device.
-	PullLockUnlock(ctx context.Context, in *PullLockUnlockRequest, opts ...grpc.CallOption) (LockUnlockApi_PullLockUnlockClient, error)
+	PullLockUnlock(ctx context.Context, in *PullLockUnlockRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullLockUnlockResponse], error)
 }
 
 type lockUnlockApiClient struct {
@@ -45,8 +49,9 @@ func NewLockUnlockApiClient(cc grpc.ClientConnInterface) LockUnlockApiClient {
 }
 
 func (c *lockUnlockApiClient) GetLockUnlock(ctx context.Context, in *GetLockUnlockRequest, opts ...grpc.CallOption) (*LockUnlock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(LockUnlock)
-	err := c.cc.Invoke(ctx, LockUnlockApi_GetLockUnlock_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, LockUnlockApi_GetLockUnlock_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,20 +59,22 @@ func (c *lockUnlockApiClient) GetLockUnlock(ctx context.Context, in *GetLockUnlo
 }
 
 func (c *lockUnlockApiClient) UpdateLockUnlock(ctx context.Context, in *UpdateLockUnlockRequest, opts ...grpc.CallOption) (*LockUnlock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(LockUnlock)
-	err := c.cc.Invoke(ctx, LockUnlockApi_UpdateLockUnlock_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, LockUnlockApi_UpdateLockUnlock_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *lockUnlockApiClient) PullLockUnlock(ctx context.Context, in *PullLockUnlockRequest, opts ...grpc.CallOption) (LockUnlockApi_PullLockUnlockClient, error) {
-	stream, err := c.cc.NewStream(ctx, &LockUnlockApi_ServiceDesc.Streams[0], LockUnlockApi_PullLockUnlock_FullMethodName, opts...)
+func (c *lockUnlockApiClient) PullLockUnlock(ctx context.Context, in *PullLockUnlockRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullLockUnlockResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &LockUnlockApi_ServiceDesc.Streams[0], LockUnlockApi_PullLockUnlock_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &lockUnlockApiPullLockUnlockClient{stream}
+	x := &grpc.GenericClientStream[PullLockUnlockRequest, PullLockUnlockResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -77,39 +84,32 @@ func (c *lockUnlockApiClient) PullLockUnlock(ctx context.Context, in *PullLockUn
 	return x, nil
 }
 
-type LockUnlockApi_PullLockUnlockClient interface {
-	Recv() (*PullLockUnlockResponse, error)
-	grpc.ClientStream
-}
-
-type lockUnlockApiPullLockUnlockClient struct {
-	grpc.ClientStream
-}
-
-func (x *lockUnlockApiPullLockUnlockClient) Recv() (*PullLockUnlockResponse, error) {
-	m := new(PullLockUnlockResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type LockUnlockApi_PullLockUnlockClient = grpc.ServerStreamingClient[PullLockUnlockResponse]
 
 // LockUnlockApiServer is the server API for LockUnlockApi service.
 // All implementations must embed UnimplementedLockUnlockApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// LockUnlockApi describes devices that can be locked and unlocked, and/or report a locked state.
+// The concept is encoded in the LockUnlock.position, a device may be locked, unlocked, or transitioning into one of
+// these positions.
 type LockUnlockApiServer interface {
 	// Get the current lock position for the device.
 	GetLockUnlock(context.Context, *GetLockUnlockRequest) (*LockUnlock, error)
 	// Update the lock position of the device.
 	UpdateLockUnlock(context.Context, *UpdateLockUnlockRequest) (*LockUnlock, error)
 	// Subscribe to changes in the lock position for the device.
-	PullLockUnlock(*PullLockUnlockRequest, LockUnlockApi_PullLockUnlockServer) error
+	PullLockUnlock(*PullLockUnlockRequest, grpc.ServerStreamingServer[PullLockUnlockResponse]) error
 	mustEmbedUnimplementedLockUnlockApiServer()
 }
 
-// UnimplementedLockUnlockApiServer must be embedded to have forward compatible implementations.
-type UnimplementedLockUnlockApiServer struct {
-}
+// UnimplementedLockUnlockApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedLockUnlockApiServer struct{}
 
 func (UnimplementedLockUnlockApiServer) GetLockUnlock(context.Context, *GetLockUnlockRequest) (*LockUnlock, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetLockUnlock not implemented")
@@ -117,10 +117,11 @@ func (UnimplementedLockUnlockApiServer) GetLockUnlock(context.Context, *GetLockU
 func (UnimplementedLockUnlockApiServer) UpdateLockUnlock(context.Context, *UpdateLockUnlockRequest) (*LockUnlock, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateLockUnlock not implemented")
 }
-func (UnimplementedLockUnlockApiServer) PullLockUnlock(*PullLockUnlockRequest, LockUnlockApi_PullLockUnlockServer) error {
+func (UnimplementedLockUnlockApiServer) PullLockUnlock(*PullLockUnlockRequest, grpc.ServerStreamingServer[PullLockUnlockResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullLockUnlock not implemented")
 }
 func (UnimplementedLockUnlockApiServer) mustEmbedUnimplementedLockUnlockApiServer() {}
+func (UnimplementedLockUnlockApiServer) testEmbeddedByValue()                       {}
 
 // UnsafeLockUnlockApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to LockUnlockApiServer will
@@ -130,6 +131,13 @@ type UnsafeLockUnlockApiServer interface {
 }
 
 func RegisterLockUnlockApiServer(s grpc.ServiceRegistrar, srv LockUnlockApiServer) {
+	// If the following call pancis, it indicates UnimplementedLockUnlockApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&LockUnlockApi_ServiceDesc, srv)
 }
 
@@ -174,21 +182,11 @@ func _LockUnlockApi_PullLockUnlock_Handler(srv interface{}, stream grpc.ServerSt
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(LockUnlockApiServer).PullLockUnlock(m, &lockUnlockApiPullLockUnlockServer{stream})
+	return srv.(LockUnlockApiServer).PullLockUnlock(m, &grpc.GenericServerStream[PullLockUnlockRequest, PullLockUnlockResponse]{ServerStream: stream})
 }
 
-type LockUnlockApi_PullLockUnlockServer interface {
-	Send(*PullLockUnlockResponse) error
-	grpc.ServerStream
-}
-
-type lockUnlockApiPullLockUnlockServer struct {
-	grpc.ServerStream
-}
-
-func (x *lockUnlockApiPullLockUnlockServer) Send(m *PullLockUnlockResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type LockUnlockApi_PullLockUnlockServer = grpc.ServerStreamingServer[PullLockUnlockResponse]
 
 // LockUnlockApi_ServiceDesc is the grpc.ServiceDesc for LockUnlockApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -216,8 +214,6 @@ var LockUnlockApi_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "traits/lock_unlock.proto",
 }
 
-const ()
-
 // LockUnlockInfoClient is the client API for LockUnlockInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
@@ -234,16 +230,20 @@ func NewLockUnlockInfoClient(cc grpc.ClientConnInterface) LockUnlockInfoClient {
 
 // LockUnlockInfoServer is the server API for LockUnlockInfo service.
 // All implementations must embed UnimplementedLockUnlockInfoServer
-// for forward compatibility
+// for forward compatibility.
 type LockUnlockInfoServer interface {
 	mustEmbedUnimplementedLockUnlockInfoServer()
 }
 
-// UnimplementedLockUnlockInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedLockUnlockInfoServer struct {
-}
+// UnimplementedLockUnlockInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedLockUnlockInfoServer struct{}
 
 func (UnimplementedLockUnlockInfoServer) mustEmbedUnimplementedLockUnlockInfoServer() {}
+func (UnimplementedLockUnlockInfoServer) testEmbeddedByValue()                        {}
 
 // UnsafeLockUnlockInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to LockUnlockInfoServer will
@@ -253,6 +253,13 @@ type UnsafeLockUnlockInfoServer interface {
 }
 
 func RegisterLockUnlockInfoServer(s grpc.ServiceRegistrar, srv LockUnlockInfoServer) {
+	// If the following call pancis, it indicates UnimplementedLockUnlockInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&LockUnlockInfo_ServiceDesc, srv)
 }
 
