@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/metadata.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	MetadataApi_GetMetadata_FullMethodName  = "/smartcore.traits.MetadataApi/GetMetadata"
@@ -26,11 +26,13 @@ const (
 // MetadataApiClient is the client API for MetadataApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// MetadataApi describes information about a named device.
 type MetadataApiClient interface {
 	// GetMetadata returns information associated with the given name.
 	GetMetadata(ctx context.Context, in *GetMetadataRequest, opts ...grpc.CallOption) (*Metadata, error)
 	// PullMetadata subscribes to changes in the information associated with the named device.
-	PullMetadata(ctx context.Context, in *PullMetadataRequest, opts ...grpc.CallOption) (MetadataApi_PullMetadataClient, error)
+	PullMetadata(ctx context.Context, in *PullMetadataRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullMetadataResponse], error)
 }
 
 type metadataApiClient struct {
@@ -42,20 +44,22 @@ func NewMetadataApiClient(cc grpc.ClientConnInterface) MetadataApiClient {
 }
 
 func (c *metadataApiClient) GetMetadata(ctx context.Context, in *GetMetadataRequest, opts ...grpc.CallOption) (*Metadata, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Metadata)
-	err := c.cc.Invoke(ctx, MetadataApi_GetMetadata_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MetadataApi_GetMetadata_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *metadataApiClient) PullMetadata(ctx context.Context, in *PullMetadataRequest, opts ...grpc.CallOption) (MetadataApi_PullMetadataClient, error) {
-	stream, err := c.cc.NewStream(ctx, &MetadataApi_ServiceDesc.Streams[0], MetadataApi_PullMetadata_FullMethodName, opts...)
+func (c *metadataApiClient) PullMetadata(ctx context.Context, in *PullMetadataRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullMetadataResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &MetadataApi_ServiceDesc.Streams[0], MetadataApi_PullMetadata_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &metadataApiPullMetadataClient{stream}
+	x := &grpc.GenericClientStream[PullMetadataRequest, PullMetadataResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -65,45 +69,37 @@ func (c *metadataApiClient) PullMetadata(ctx context.Context, in *PullMetadataRe
 	return x, nil
 }
 
-type MetadataApi_PullMetadataClient interface {
-	Recv() (*PullMetadataResponse, error)
-	grpc.ClientStream
-}
-
-type metadataApiPullMetadataClient struct {
-	grpc.ClientStream
-}
-
-func (x *metadataApiPullMetadataClient) Recv() (*PullMetadataResponse, error) {
-	m := new(PullMetadataResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MetadataApi_PullMetadataClient = grpc.ServerStreamingClient[PullMetadataResponse]
 
 // MetadataApiServer is the server API for MetadataApi service.
 // All implementations must embed UnimplementedMetadataApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// MetadataApi describes information about a named device.
 type MetadataApiServer interface {
 	// GetMetadata returns information associated with the given name.
 	GetMetadata(context.Context, *GetMetadataRequest) (*Metadata, error)
 	// PullMetadata subscribes to changes in the information associated with the named device.
-	PullMetadata(*PullMetadataRequest, MetadataApi_PullMetadataServer) error
+	PullMetadata(*PullMetadataRequest, grpc.ServerStreamingServer[PullMetadataResponse]) error
 	mustEmbedUnimplementedMetadataApiServer()
 }
 
-// UnimplementedMetadataApiServer must be embedded to have forward compatible implementations.
-type UnimplementedMetadataApiServer struct {
-}
+// UnimplementedMetadataApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMetadataApiServer struct{}
 
 func (UnimplementedMetadataApiServer) GetMetadata(context.Context, *GetMetadataRequest) (*Metadata, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetMetadata not implemented")
 }
-func (UnimplementedMetadataApiServer) PullMetadata(*PullMetadataRequest, MetadataApi_PullMetadataServer) error {
+func (UnimplementedMetadataApiServer) PullMetadata(*PullMetadataRequest, grpc.ServerStreamingServer[PullMetadataResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullMetadata not implemented")
 }
 func (UnimplementedMetadataApiServer) mustEmbedUnimplementedMetadataApiServer() {}
+func (UnimplementedMetadataApiServer) testEmbeddedByValue()                     {}
 
 // UnsafeMetadataApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MetadataApiServer will
@@ -113,6 +109,13 @@ type UnsafeMetadataApiServer interface {
 }
 
 func RegisterMetadataApiServer(s grpc.ServiceRegistrar, srv MetadataApiServer) {
+	// If the following call pancis, it indicates UnimplementedMetadataApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MetadataApi_ServiceDesc, srv)
 }
 
@@ -139,21 +142,11 @@ func _MetadataApi_PullMetadata_Handler(srv interface{}, stream grpc.ServerStream
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(MetadataApiServer).PullMetadata(m, &metadataApiPullMetadataServer{stream})
+	return srv.(MetadataApiServer).PullMetadata(m, &grpc.GenericServerStream[PullMetadataRequest, PullMetadataResponse]{ServerStream: stream})
 }
 
-type MetadataApi_PullMetadataServer interface {
-	Send(*PullMetadataResponse) error
-	grpc.ServerStream
-}
-
-type metadataApiPullMetadataServer struct {
-	grpc.ServerStream
-}
-
-func (x *metadataApiPullMetadataServer) Send(m *PullMetadataResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MetadataApi_PullMetadataServer = grpc.ServerStreamingServer[PullMetadataResponse]
 
 // MetadataApi_ServiceDesc is the grpc.ServiceDesc for MetadataApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -177,8 +170,6 @@ var MetadataApi_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "traits/metadata.proto",
 }
 
-const ()
-
 // MetadataInfoClient is the client API for MetadataInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
@@ -195,16 +186,20 @@ func NewMetadataInfoClient(cc grpc.ClientConnInterface) MetadataInfoClient {
 
 // MetadataInfoServer is the server API for MetadataInfo service.
 // All implementations must embed UnimplementedMetadataInfoServer
-// for forward compatibility
+// for forward compatibility.
 type MetadataInfoServer interface {
 	mustEmbedUnimplementedMetadataInfoServer()
 }
 
-// UnimplementedMetadataInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedMetadataInfoServer struct {
-}
+// UnimplementedMetadataInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMetadataInfoServer struct{}
 
 func (UnimplementedMetadataInfoServer) mustEmbedUnimplementedMetadataInfoServer() {}
+func (UnimplementedMetadataInfoServer) testEmbeddedByValue()                      {}
 
 // UnsafeMetadataInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MetadataInfoServer will
@@ -214,6 +209,13 @@ type UnsafeMetadataInfoServer interface {
 }
 
 func RegisterMetadataInfoServer(s grpc.ServiceRegistrar, srv MetadataInfoServer) {
+	// If the following call pancis, it indicates UnimplementedMetadataInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MetadataInfo_ServiceDesc, srv)
 }
 
