@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/microphone.proto
 
 package traits
@@ -16,8 +16,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	MicrophoneApi_GetGain_FullMethodName    = "/smartcore.traits.MicrophoneApi/GetGain"
@@ -28,12 +28,17 @@ const (
 // MicrophoneApiClient is the client API for MicrophoneApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Trait for devices that support some kind of audio input. The most basic is a simple muting microphone, it's optional
+// to support gain adjustment on a scale.
+//
+// For the sakes of this trait "gain" means the combination of a gain and a mute state.
 type MicrophoneApiClient interface {
 	// Get the current state of the gain for the device
 	GetGain(ctx context.Context, in *GetMicrophoneGainRequest, opts ...grpc.CallOption) (*types.AudioLevel, error)
 	// update the gain state for the device
 	UpdateGain(ctx context.Context, in *UpdateMicrophoneGainRequest, opts ...grpc.CallOption) (*types.AudioLevel, error)
-	PullGain(ctx context.Context, in *PullMicrophoneGainRequest, opts ...grpc.CallOption) (MicrophoneApi_PullGainClient, error)
+	PullGain(ctx context.Context, in *PullMicrophoneGainRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullMicrophoneGainResponse], error)
 }
 
 type microphoneApiClient struct {
@@ -45,8 +50,9 @@ func NewMicrophoneApiClient(cc grpc.ClientConnInterface) MicrophoneApiClient {
 }
 
 func (c *microphoneApiClient) GetGain(ctx context.Context, in *GetMicrophoneGainRequest, opts ...grpc.CallOption) (*types.AudioLevel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(types.AudioLevel)
-	err := c.cc.Invoke(ctx, MicrophoneApi_GetGain_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MicrophoneApi_GetGain_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,20 +60,22 @@ func (c *microphoneApiClient) GetGain(ctx context.Context, in *GetMicrophoneGain
 }
 
 func (c *microphoneApiClient) UpdateGain(ctx context.Context, in *UpdateMicrophoneGainRequest, opts ...grpc.CallOption) (*types.AudioLevel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(types.AudioLevel)
-	err := c.cc.Invoke(ctx, MicrophoneApi_UpdateGain_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MicrophoneApi_UpdateGain_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *microphoneApiClient) PullGain(ctx context.Context, in *PullMicrophoneGainRequest, opts ...grpc.CallOption) (MicrophoneApi_PullGainClient, error) {
-	stream, err := c.cc.NewStream(ctx, &MicrophoneApi_ServiceDesc.Streams[0], MicrophoneApi_PullGain_FullMethodName, opts...)
+func (c *microphoneApiClient) PullGain(ctx context.Context, in *PullMicrophoneGainRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullMicrophoneGainResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &MicrophoneApi_ServiceDesc.Streams[0], MicrophoneApi_PullGain_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &microphoneApiPullGainClient{stream}
+	x := &grpc.GenericClientStream[PullMicrophoneGainRequest, PullMicrophoneGainResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -77,38 +85,32 @@ func (c *microphoneApiClient) PullGain(ctx context.Context, in *PullMicrophoneGa
 	return x, nil
 }
 
-type MicrophoneApi_PullGainClient interface {
-	Recv() (*PullMicrophoneGainResponse, error)
-	grpc.ClientStream
-}
-
-type microphoneApiPullGainClient struct {
-	grpc.ClientStream
-}
-
-func (x *microphoneApiPullGainClient) Recv() (*PullMicrophoneGainResponse, error) {
-	m := new(PullMicrophoneGainResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MicrophoneApi_PullGainClient = grpc.ServerStreamingClient[PullMicrophoneGainResponse]
 
 // MicrophoneApiServer is the server API for MicrophoneApi service.
 // All implementations must embed UnimplementedMicrophoneApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Trait for devices that support some kind of audio input. The most basic is a simple muting microphone, it's optional
+// to support gain adjustment on a scale.
+//
+// For the sakes of this trait "gain" means the combination of a gain and a mute state.
 type MicrophoneApiServer interface {
 	// Get the current state of the gain for the device
 	GetGain(context.Context, *GetMicrophoneGainRequest) (*types.AudioLevel, error)
 	// update the gain state for the device
 	UpdateGain(context.Context, *UpdateMicrophoneGainRequest) (*types.AudioLevel, error)
-	PullGain(*PullMicrophoneGainRequest, MicrophoneApi_PullGainServer) error
+	PullGain(*PullMicrophoneGainRequest, grpc.ServerStreamingServer[PullMicrophoneGainResponse]) error
 	mustEmbedUnimplementedMicrophoneApiServer()
 }
 
-// UnimplementedMicrophoneApiServer must be embedded to have forward compatible implementations.
-type UnimplementedMicrophoneApiServer struct {
-}
+// UnimplementedMicrophoneApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMicrophoneApiServer struct{}
 
 func (UnimplementedMicrophoneApiServer) GetGain(context.Context, *GetMicrophoneGainRequest) (*types.AudioLevel, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetGain not implemented")
@@ -116,10 +118,11 @@ func (UnimplementedMicrophoneApiServer) GetGain(context.Context, *GetMicrophoneG
 func (UnimplementedMicrophoneApiServer) UpdateGain(context.Context, *UpdateMicrophoneGainRequest) (*types.AudioLevel, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateGain not implemented")
 }
-func (UnimplementedMicrophoneApiServer) PullGain(*PullMicrophoneGainRequest, MicrophoneApi_PullGainServer) error {
+func (UnimplementedMicrophoneApiServer) PullGain(*PullMicrophoneGainRequest, grpc.ServerStreamingServer[PullMicrophoneGainResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullGain not implemented")
 }
 func (UnimplementedMicrophoneApiServer) mustEmbedUnimplementedMicrophoneApiServer() {}
+func (UnimplementedMicrophoneApiServer) testEmbeddedByValue()                       {}
 
 // UnsafeMicrophoneApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MicrophoneApiServer will
@@ -129,6 +132,13 @@ type UnsafeMicrophoneApiServer interface {
 }
 
 func RegisterMicrophoneApiServer(s grpc.ServiceRegistrar, srv MicrophoneApiServer) {
+	// If the following call pancis, it indicates UnimplementedMicrophoneApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MicrophoneApi_ServiceDesc, srv)
 }
 
@@ -173,21 +183,11 @@ func _MicrophoneApi_PullGain_Handler(srv interface{}, stream grpc.ServerStream) 
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(MicrophoneApiServer).PullGain(m, &microphoneApiPullGainServer{stream})
+	return srv.(MicrophoneApiServer).PullGain(m, &grpc.GenericServerStream[PullMicrophoneGainRequest, PullMicrophoneGainResponse]{ServerStream: stream})
 }
 
-type MicrophoneApi_PullGainServer interface {
-	Send(*PullMicrophoneGainResponse) error
-	grpc.ServerStream
-}
-
-type microphoneApiPullGainServer struct {
-	grpc.ServerStream
-}
-
-func (x *microphoneApiPullGainServer) Send(m *PullMicrophoneGainResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MicrophoneApi_PullGainServer = grpc.ServerStreamingServer[PullMicrophoneGainResponse]
 
 // MicrophoneApi_ServiceDesc is the grpc.ServiceDesc for MicrophoneApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -222,6 +222,8 @@ const (
 // MicrophoneInfoClient is the client API for MicrophoneInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type MicrophoneInfoClient interface {
 	// Get information about how a named device implements Gain features
 	DescribeGain(ctx context.Context, in *DescribeGainRequest, opts ...grpc.CallOption) (*GainSupport, error)
@@ -236,8 +238,9 @@ func NewMicrophoneInfoClient(cc grpc.ClientConnInterface) MicrophoneInfoClient {
 }
 
 func (c *microphoneInfoClient) DescribeGain(ctx context.Context, in *DescribeGainRequest, opts ...grpc.CallOption) (*GainSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GainSupport)
-	err := c.cc.Invoke(ctx, MicrophoneInfo_DescribeGain_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MicrophoneInfo_DescribeGain_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -246,21 +249,27 @@ func (c *microphoneInfoClient) DescribeGain(ctx context.Context, in *DescribeGai
 
 // MicrophoneInfoServer is the server API for MicrophoneInfo service.
 // All implementations must embed UnimplementedMicrophoneInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type MicrophoneInfoServer interface {
 	// Get information about how a named device implements Gain features
 	DescribeGain(context.Context, *DescribeGainRequest) (*GainSupport, error)
 	mustEmbedUnimplementedMicrophoneInfoServer()
 }
 
-// UnimplementedMicrophoneInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedMicrophoneInfoServer struct {
-}
+// UnimplementedMicrophoneInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMicrophoneInfoServer struct{}
 
 func (UnimplementedMicrophoneInfoServer) DescribeGain(context.Context, *DescribeGainRequest) (*GainSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeGain not implemented")
 }
 func (UnimplementedMicrophoneInfoServer) mustEmbedUnimplementedMicrophoneInfoServer() {}
+func (UnimplementedMicrophoneInfoServer) testEmbeddedByValue()                        {}
 
 // UnsafeMicrophoneInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MicrophoneInfoServer will
@@ -270,6 +279,13 @@ type UnsafeMicrophoneInfoServer interface {
 }
 
 func RegisterMicrophoneInfoServer(s grpc.ServiceRegistrar, srv MicrophoneInfoServer) {
+	// If the following call pancis, it indicates UnimplementedMicrophoneInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MicrophoneInfo_ServiceDesc, srv)
 }
 
