@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/mode.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	ModeApi_GetModeValues_FullMethodName    = "/smartcore.traits.ModeApi/GetModeValues"
@@ -27,6 +27,9 @@ const (
 // ModeApiClient is the client API for ModeApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Trait for devices that present one or more distinct modes.
+// For example a washing machine might have a "temperature" mode with available values ["delicate", "auto", "whites"].
 type ModeApiClient interface {
 	// GetModeValues returns the current state for all modes for a device.
 	GetModeValues(ctx context.Context, in *GetModeValuesRequest, opts ...grpc.CallOption) (*ModeValues, error)
@@ -38,7 +41,7 @@ type ModeApiClient interface {
 	// underflow or overflow the available modes.
 	UpdateModeValues(ctx context.Context, in *UpdateModeValuesRequest, opts ...grpc.CallOption) (*ModeValues, error)
 	// PullModeValues returns a stream of updates to the modes set on the device.
-	PullModeValues(ctx context.Context, in *PullModeValuesRequest, opts ...grpc.CallOption) (ModeApi_PullModeValuesClient, error)
+	PullModeValues(ctx context.Context, in *PullModeValuesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullModeValuesResponse], error)
 }
 
 type modeApiClient struct {
@@ -50,8 +53,9 @@ func NewModeApiClient(cc grpc.ClientConnInterface) ModeApiClient {
 }
 
 func (c *modeApiClient) GetModeValues(ctx context.Context, in *GetModeValuesRequest, opts ...grpc.CallOption) (*ModeValues, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ModeValues)
-	err := c.cc.Invoke(ctx, ModeApi_GetModeValues_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ModeApi_GetModeValues_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,20 +63,22 @@ func (c *modeApiClient) GetModeValues(ctx context.Context, in *GetModeValuesRequ
 }
 
 func (c *modeApiClient) UpdateModeValues(ctx context.Context, in *UpdateModeValuesRequest, opts ...grpc.CallOption) (*ModeValues, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ModeValues)
-	err := c.cc.Invoke(ctx, ModeApi_UpdateModeValues_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ModeApi_UpdateModeValues_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *modeApiClient) PullModeValues(ctx context.Context, in *PullModeValuesRequest, opts ...grpc.CallOption) (ModeApi_PullModeValuesClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ModeApi_ServiceDesc.Streams[0], ModeApi_PullModeValues_FullMethodName, opts...)
+func (c *modeApiClient) PullModeValues(ctx context.Context, in *PullModeValuesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullModeValuesResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ModeApi_ServiceDesc.Streams[0], ModeApi_PullModeValues_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &modeApiPullModeValuesClient{stream}
+	x := &grpc.GenericClientStream[PullModeValuesRequest, PullModeValuesResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -82,26 +88,15 @@ func (c *modeApiClient) PullModeValues(ctx context.Context, in *PullModeValuesRe
 	return x, nil
 }
 
-type ModeApi_PullModeValuesClient interface {
-	Recv() (*PullModeValuesResponse, error)
-	grpc.ClientStream
-}
-
-type modeApiPullModeValuesClient struct {
-	grpc.ClientStream
-}
-
-func (x *modeApiPullModeValuesClient) Recv() (*PullModeValuesResponse, error) {
-	m := new(PullModeValuesResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ModeApi_PullModeValuesClient = grpc.ServerStreamingClient[PullModeValuesResponse]
 
 // ModeApiServer is the server API for ModeApi service.
 // All implementations must embed UnimplementedModeApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Trait for devices that present one or more distinct modes.
+// For example a washing machine might have a "temperature" mode with available values ["delicate", "auto", "whites"].
 type ModeApiServer interface {
 	// GetModeValues returns the current state for all modes for a device.
 	GetModeValues(context.Context, *GetModeValuesRequest) (*ModeValues, error)
@@ -113,13 +108,16 @@ type ModeApiServer interface {
 	// underflow or overflow the available modes.
 	UpdateModeValues(context.Context, *UpdateModeValuesRequest) (*ModeValues, error)
 	// PullModeValues returns a stream of updates to the modes set on the device.
-	PullModeValues(*PullModeValuesRequest, ModeApi_PullModeValuesServer) error
+	PullModeValues(*PullModeValuesRequest, grpc.ServerStreamingServer[PullModeValuesResponse]) error
 	mustEmbedUnimplementedModeApiServer()
 }
 
-// UnimplementedModeApiServer must be embedded to have forward compatible implementations.
-type UnimplementedModeApiServer struct {
-}
+// UnimplementedModeApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedModeApiServer struct{}
 
 func (UnimplementedModeApiServer) GetModeValues(context.Context, *GetModeValuesRequest) (*ModeValues, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetModeValues not implemented")
@@ -127,10 +125,11 @@ func (UnimplementedModeApiServer) GetModeValues(context.Context, *GetModeValuesR
 func (UnimplementedModeApiServer) UpdateModeValues(context.Context, *UpdateModeValuesRequest) (*ModeValues, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateModeValues not implemented")
 }
-func (UnimplementedModeApiServer) PullModeValues(*PullModeValuesRequest, ModeApi_PullModeValuesServer) error {
+func (UnimplementedModeApiServer) PullModeValues(*PullModeValuesRequest, grpc.ServerStreamingServer[PullModeValuesResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullModeValues not implemented")
 }
 func (UnimplementedModeApiServer) mustEmbedUnimplementedModeApiServer() {}
+func (UnimplementedModeApiServer) testEmbeddedByValue()                 {}
 
 // UnsafeModeApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ModeApiServer will
@@ -140,6 +139,13 @@ type UnsafeModeApiServer interface {
 }
 
 func RegisterModeApiServer(s grpc.ServiceRegistrar, srv ModeApiServer) {
+	// If the following call pancis, it indicates UnimplementedModeApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ModeApi_ServiceDesc, srv)
 }
 
@@ -184,21 +190,11 @@ func _ModeApi_PullModeValues_Handler(srv interface{}, stream grpc.ServerStream) 
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ModeApiServer).PullModeValues(m, &modeApiPullModeValuesServer{stream})
+	return srv.(ModeApiServer).PullModeValues(m, &grpc.GenericServerStream[PullModeValuesRequest, PullModeValuesResponse]{ServerStream: stream})
 }
 
-type ModeApi_PullModeValuesServer interface {
-	Send(*PullModeValuesResponse) error
-	grpc.ServerStream
-}
-
-type modeApiPullModeValuesServer struct {
-	grpc.ServerStream
-}
-
-func (x *modeApiPullModeValuesServer) Send(m *PullModeValuesResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ModeApi_PullModeValuesServer = grpc.ServerStreamingServer[PullModeValuesResponse]
 
 // ModeApi_ServiceDesc is the grpc.ServiceDesc for ModeApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -233,6 +229,8 @@ const (
 // ModeInfoClient is the client API for ModeInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ModeInfoClient interface {
 	// Get information about how a named device implements Modes features
 	DescribeModes(ctx context.Context, in *DescribeModesRequest, opts ...grpc.CallOption) (*ModesSupport, error)
@@ -247,8 +245,9 @@ func NewModeInfoClient(cc grpc.ClientConnInterface) ModeInfoClient {
 }
 
 func (c *modeInfoClient) DescribeModes(ctx context.Context, in *DescribeModesRequest, opts ...grpc.CallOption) (*ModesSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ModesSupport)
-	err := c.cc.Invoke(ctx, ModeInfo_DescribeModes_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ModeInfo_DescribeModes_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -257,21 +256,27 @@ func (c *modeInfoClient) DescribeModes(ctx context.Context, in *DescribeModesReq
 
 // ModeInfoServer is the server API for ModeInfo service.
 // All implementations must embed UnimplementedModeInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type ModeInfoServer interface {
 	// Get information about how a named device implements Modes features
 	DescribeModes(context.Context, *DescribeModesRequest) (*ModesSupport, error)
 	mustEmbedUnimplementedModeInfoServer()
 }
 
-// UnimplementedModeInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedModeInfoServer struct {
-}
+// UnimplementedModeInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedModeInfoServer struct{}
 
 func (UnimplementedModeInfoServer) DescribeModes(context.Context, *DescribeModesRequest) (*ModesSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeModes not implemented")
 }
 func (UnimplementedModeInfoServer) mustEmbedUnimplementedModeInfoServer() {}
+func (UnimplementedModeInfoServer) testEmbeddedByValue()                  {}
 
 // UnsafeModeInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ModeInfoServer will
@@ -281,6 +286,13 @@ type UnsafeModeInfoServer interface {
 }
 
 func RegisterModeInfoServer(s grpc.ServiceRegistrar, srv ModeInfoServer) {
+	// If the following call pancis, it indicates UnimplementedModeInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ModeInfo_ServiceDesc, srv)
 }
 
