@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/motion_sensor.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	MotionSensorApi_GetMotionDetection_FullMethodName   = "/smartcore.traits.MotionSensorApi/GetMotionDetection"
@@ -26,11 +26,13 @@ const (
 // MotionSensorApiClient is the client API for MotionSensorApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Represents a sensor that detects motion. Note that the absence of motion does not directly imply an unoccupied space.
 type MotionSensorApiClient interface {
 	// Get the current motion state for the device.
 	GetMotionDetection(ctx context.Context, in *GetMotionDetectionRequest, opts ...grpc.CallOption) (*MotionDetection, error)
 	// subscribe to changes in the motion state for the device.
-	PullMotionDetections(ctx context.Context, in *PullMotionDetectionRequest, opts ...grpc.CallOption) (MotionSensorApi_PullMotionDetectionsClient, error)
+	PullMotionDetections(ctx context.Context, in *PullMotionDetectionRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullMotionDetectionResponse], error)
 }
 
 type motionSensorApiClient struct {
@@ -42,20 +44,22 @@ func NewMotionSensorApiClient(cc grpc.ClientConnInterface) MotionSensorApiClient
 }
 
 func (c *motionSensorApiClient) GetMotionDetection(ctx context.Context, in *GetMotionDetectionRequest, opts ...grpc.CallOption) (*MotionDetection, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MotionDetection)
-	err := c.cc.Invoke(ctx, MotionSensorApi_GetMotionDetection_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MotionSensorApi_GetMotionDetection_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *motionSensorApiClient) PullMotionDetections(ctx context.Context, in *PullMotionDetectionRequest, opts ...grpc.CallOption) (MotionSensorApi_PullMotionDetectionsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &MotionSensorApi_ServiceDesc.Streams[0], MotionSensorApi_PullMotionDetections_FullMethodName, opts...)
+func (c *motionSensorApiClient) PullMotionDetections(ctx context.Context, in *PullMotionDetectionRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullMotionDetectionResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &MotionSensorApi_ServiceDesc.Streams[0], MotionSensorApi_PullMotionDetections_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &motionSensorApiPullMotionDetectionsClient{stream}
+	x := &grpc.GenericClientStream[PullMotionDetectionRequest, PullMotionDetectionResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -65,45 +69,37 @@ func (c *motionSensorApiClient) PullMotionDetections(ctx context.Context, in *Pu
 	return x, nil
 }
 
-type MotionSensorApi_PullMotionDetectionsClient interface {
-	Recv() (*PullMotionDetectionResponse, error)
-	grpc.ClientStream
-}
-
-type motionSensorApiPullMotionDetectionsClient struct {
-	grpc.ClientStream
-}
-
-func (x *motionSensorApiPullMotionDetectionsClient) Recv() (*PullMotionDetectionResponse, error) {
-	m := new(PullMotionDetectionResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MotionSensorApi_PullMotionDetectionsClient = grpc.ServerStreamingClient[PullMotionDetectionResponse]
 
 // MotionSensorApiServer is the server API for MotionSensorApi service.
 // All implementations must embed UnimplementedMotionSensorApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Represents a sensor that detects motion. Note that the absence of motion does not directly imply an unoccupied space.
 type MotionSensorApiServer interface {
 	// Get the current motion state for the device.
 	GetMotionDetection(context.Context, *GetMotionDetectionRequest) (*MotionDetection, error)
 	// subscribe to changes in the motion state for the device.
-	PullMotionDetections(*PullMotionDetectionRequest, MotionSensorApi_PullMotionDetectionsServer) error
+	PullMotionDetections(*PullMotionDetectionRequest, grpc.ServerStreamingServer[PullMotionDetectionResponse]) error
 	mustEmbedUnimplementedMotionSensorApiServer()
 }
 
-// UnimplementedMotionSensorApiServer must be embedded to have forward compatible implementations.
-type UnimplementedMotionSensorApiServer struct {
-}
+// UnimplementedMotionSensorApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMotionSensorApiServer struct{}
 
 func (UnimplementedMotionSensorApiServer) GetMotionDetection(context.Context, *GetMotionDetectionRequest) (*MotionDetection, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetMotionDetection not implemented")
 }
-func (UnimplementedMotionSensorApiServer) PullMotionDetections(*PullMotionDetectionRequest, MotionSensorApi_PullMotionDetectionsServer) error {
+func (UnimplementedMotionSensorApiServer) PullMotionDetections(*PullMotionDetectionRequest, grpc.ServerStreamingServer[PullMotionDetectionResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullMotionDetections not implemented")
 }
 func (UnimplementedMotionSensorApiServer) mustEmbedUnimplementedMotionSensorApiServer() {}
+func (UnimplementedMotionSensorApiServer) testEmbeddedByValue()                         {}
 
 // UnsafeMotionSensorApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MotionSensorApiServer will
@@ -113,6 +109,13 @@ type UnsafeMotionSensorApiServer interface {
 }
 
 func RegisterMotionSensorApiServer(s grpc.ServiceRegistrar, srv MotionSensorApiServer) {
+	// If the following call pancis, it indicates UnimplementedMotionSensorApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MotionSensorApi_ServiceDesc, srv)
 }
 
@@ -139,21 +142,11 @@ func _MotionSensorApi_PullMotionDetections_Handler(srv interface{}, stream grpc.
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(MotionSensorApiServer).PullMotionDetections(m, &motionSensorApiPullMotionDetectionsServer{stream})
+	return srv.(MotionSensorApiServer).PullMotionDetections(m, &grpc.GenericServerStream[PullMotionDetectionRequest, PullMotionDetectionResponse]{ServerStream: stream})
 }
 
-type MotionSensorApi_PullMotionDetectionsServer interface {
-	Send(*PullMotionDetectionResponse) error
-	grpc.ServerStream
-}
-
-type motionSensorApiPullMotionDetectionsServer struct {
-	grpc.ServerStream
-}
-
-func (x *motionSensorApiPullMotionDetectionsServer) Send(m *PullMotionDetectionResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MotionSensorApi_PullMotionDetectionsServer = grpc.ServerStreamingServer[PullMotionDetectionResponse]
 
 // MotionSensorApi_ServiceDesc is the grpc.ServiceDesc for MotionSensorApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -184,6 +177,8 @@ const (
 // MotionSensorSensorInfoClient is the client API for MotionSensorSensorInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type MotionSensorSensorInfoClient interface {
 	// Get information about how a named device implements MotionDetection measurements
 	DescribeMotionDetection(ctx context.Context, in *DescribeMotionDetectionRequest, opts ...grpc.CallOption) (*MotionDetectionSupport, error)
@@ -198,8 +193,9 @@ func NewMotionSensorSensorInfoClient(cc grpc.ClientConnInterface) MotionSensorSe
 }
 
 func (c *motionSensorSensorInfoClient) DescribeMotionDetection(ctx context.Context, in *DescribeMotionDetectionRequest, opts ...grpc.CallOption) (*MotionDetectionSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MotionDetectionSupport)
-	err := c.cc.Invoke(ctx, MotionSensorSensorInfo_DescribeMotionDetection_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MotionSensorSensorInfo_DescribeMotionDetection_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -208,22 +204,28 @@ func (c *motionSensorSensorInfoClient) DescribeMotionDetection(ctx context.Conte
 
 // MotionSensorSensorInfoServer is the server API for MotionSensorSensorInfo service.
 // All implementations must embed UnimplementedMotionSensorSensorInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type MotionSensorSensorInfoServer interface {
 	// Get information about how a named device implements MotionDetection measurements
 	DescribeMotionDetection(context.Context, *DescribeMotionDetectionRequest) (*MotionDetectionSupport, error)
 	mustEmbedUnimplementedMotionSensorSensorInfoServer()
 }
 
-// UnimplementedMotionSensorSensorInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedMotionSensorSensorInfoServer struct {
-}
+// UnimplementedMotionSensorSensorInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMotionSensorSensorInfoServer struct{}
 
 func (UnimplementedMotionSensorSensorInfoServer) DescribeMotionDetection(context.Context, *DescribeMotionDetectionRequest) (*MotionDetectionSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeMotionDetection not implemented")
 }
 func (UnimplementedMotionSensorSensorInfoServer) mustEmbedUnimplementedMotionSensorSensorInfoServer() {
 }
+func (UnimplementedMotionSensorSensorInfoServer) testEmbeddedByValue() {}
 
 // UnsafeMotionSensorSensorInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MotionSensorSensorInfoServer will
@@ -233,6 +235,13 @@ type UnsafeMotionSensorSensorInfoServer interface {
 }
 
 func RegisterMotionSensorSensorInfoServer(s grpc.ServiceRegistrar, srv MotionSensorSensorInfoServer) {
+	// If the following call pancis, it indicates UnimplementedMotionSensorSensorInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MotionSensorSensorInfo_ServiceDesc, srv)
 }
 
