@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/occupancy_sensor.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	OccupancySensorApi_GetOccupancy_FullMethodName  = "/smartcore.traits.OccupancySensorApi/GetOccupancy"
@@ -26,11 +26,15 @@ const (
 // OccupancySensorApiClient is the client API for OccupancySensorApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// The reporting of occupancy in a space. Similar to a MotionSensor, this reports whether a space is occupied or
+// unoccupied with additional optional information relating to _how_ occupied the space is (i.e. how many people are
+// present)
 type OccupancySensorApiClient interface {
 	// Get the current motion state for the device.
 	GetOccupancy(ctx context.Context, in *GetOccupancyRequest, opts ...grpc.CallOption) (*Occupancy, error)
 	// subscribe to changes in the motion state for the device.
-	PullOccupancy(ctx context.Context, in *PullOccupancyRequest, opts ...grpc.CallOption) (OccupancySensorApi_PullOccupancyClient, error)
+	PullOccupancy(ctx context.Context, in *PullOccupancyRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullOccupancyResponse], error)
 }
 
 type occupancySensorApiClient struct {
@@ -42,20 +46,22 @@ func NewOccupancySensorApiClient(cc grpc.ClientConnInterface) OccupancySensorApi
 }
 
 func (c *occupancySensorApiClient) GetOccupancy(ctx context.Context, in *GetOccupancyRequest, opts ...grpc.CallOption) (*Occupancy, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Occupancy)
-	err := c.cc.Invoke(ctx, OccupancySensorApi_GetOccupancy_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OccupancySensorApi_GetOccupancy_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *occupancySensorApiClient) PullOccupancy(ctx context.Context, in *PullOccupancyRequest, opts ...grpc.CallOption) (OccupancySensorApi_PullOccupancyClient, error) {
-	stream, err := c.cc.NewStream(ctx, &OccupancySensorApi_ServiceDesc.Streams[0], OccupancySensorApi_PullOccupancy_FullMethodName, opts...)
+func (c *occupancySensorApiClient) PullOccupancy(ctx context.Context, in *PullOccupancyRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullOccupancyResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &OccupancySensorApi_ServiceDesc.Streams[0], OccupancySensorApi_PullOccupancy_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &occupancySensorApiPullOccupancyClient{stream}
+	x := &grpc.GenericClientStream[PullOccupancyRequest, PullOccupancyResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -65,45 +71,39 @@ func (c *occupancySensorApiClient) PullOccupancy(ctx context.Context, in *PullOc
 	return x, nil
 }
 
-type OccupancySensorApi_PullOccupancyClient interface {
-	Recv() (*PullOccupancyResponse, error)
-	grpc.ClientStream
-}
-
-type occupancySensorApiPullOccupancyClient struct {
-	grpc.ClientStream
-}
-
-func (x *occupancySensorApiPullOccupancyClient) Recv() (*PullOccupancyResponse, error) {
-	m := new(PullOccupancyResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type OccupancySensorApi_PullOccupancyClient = grpc.ServerStreamingClient[PullOccupancyResponse]
 
 // OccupancySensorApiServer is the server API for OccupancySensorApi service.
 // All implementations must embed UnimplementedOccupancySensorApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// The reporting of occupancy in a space. Similar to a MotionSensor, this reports whether a space is occupied or
+// unoccupied with additional optional information relating to _how_ occupied the space is (i.e. how many people are
+// present)
 type OccupancySensorApiServer interface {
 	// Get the current motion state for the device.
 	GetOccupancy(context.Context, *GetOccupancyRequest) (*Occupancy, error)
 	// subscribe to changes in the motion state for the device.
-	PullOccupancy(*PullOccupancyRequest, OccupancySensorApi_PullOccupancyServer) error
+	PullOccupancy(*PullOccupancyRequest, grpc.ServerStreamingServer[PullOccupancyResponse]) error
 	mustEmbedUnimplementedOccupancySensorApiServer()
 }
 
-// UnimplementedOccupancySensorApiServer must be embedded to have forward compatible implementations.
-type UnimplementedOccupancySensorApiServer struct {
-}
+// UnimplementedOccupancySensorApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedOccupancySensorApiServer struct{}
 
 func (UnimplementedOccupancySensorApiServer) GetOccupancy(context.Context, *GetOccupancyRequest) (*Occupancy, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetOccupancy not implemented")
 }
-func (UnimplementedOccupancySensorApiServer) PullOccupancy(*PullOccupancyRequest, OccupancySensorApi_PullOccupancyServer) error {
+func (UnimplementedOccupancySensorApiServer) PullOccupancy(*PullOccupancyRequest, grpc.ServerStreamingServer[PullOccupancyResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullOccupancy not implemented")
 }
 func (UnimplementedOccupancySensorApiServer) mustEmbedUnimplementedOccupancySensorApiServer() {}
+func (UnimplementedOccupancySensorApiServer) testEmbeddedByValue()                            {}
 
 // UnsafeOccupancySensorApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to OccupancySensorApiServer will
@@ -113,6 +113,13 @@ type UnsafeOccupancySensorApiServer interface {
 }
 
 func RegisterOccupancySensorApiServer(s grpc.ServiceRegistrar, srv OccupancySensorApiServer) {
+	// If the following call pancis, it indicates UnimplementedOccupancySensorApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&OccupancySensorApi_ServiceDesc, srv)
 }
 
@@ -139,21 +146,11 @@ func _OccupancySensorApi_PullOccupancy_Handler(srv interface{}, stream grpc.Serv
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(OccupancySensorApiServer).PullOccupancy(m, &occupancySensorApiPullOccupancyServer{stream})
+	return srv.(OccupancySensorApiServer).PullOccupancy(m, &grpc.GenericServerStream[PullOccupancyRequest, PullOccupancyResponse]{ServerStream: stream})
 }
 
-type OccupancySensorApi_PullOccupancyServer interface {
-	Send(*PullOccupancyResponse) error
-	grpc.ServerStream
-}
-
-type occupancySensorApiPullOccupancyServer struct {
-	grpc.ServerStream
-}
-
-func (x *occupancySensorApiPullOccupancyServer) Send(m *PullOccupancyResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type OccupancySensorApi_PullOccupancyServer = grpc.ServerStreamingServer[PullOccupancyResponse]
 
 // OccupancySensorApi_ServiceDesc is the grpc.ServiceDesc for OccupancySensorApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -184,6 +181,8 @@ const (
 // OccupancySensorInfoClient is the client API for OccupancySensorInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type OccupancySensorInfoClient interface {
 	// Get information about how a named device implements Occupancy measurements
 	DescribeOccupancy(ctx context.Context, in *DescribeOccupancyRequest, opts ...grpc.CallOption) (*OccupancySupport, error)
@@ -198,8 +197,9 @@ func NewOccupancySensorInfoClient(cc grpc.ClientConnInterface) OccupancySensorIn
 }
 
 func (c *occupancySensorInfoClient) DescribeOccupancy(ctx context.Context, in *DescribeOccupancyRequest, opts ...grpc.CallOption) (*OccupancySupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(OccupancySupport)
-	err := c.cc.Invoke(ctx, OccupancySensorInfo_DescribeOccupancy_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OccupancySensorInfo_DescribeOccupancy_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -208,21 +208,27 @@ func (c *occupancySensorInfoClient) DescribeOccupancy(ctx context.Context, in *D
 
 // OccupancySensorInfoServer is the server API for OccupancySensorInfo service.
 // All implementations must embed UnimplementedOccupancySensorInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type OccupancySensorInfoServer interface {
 	// Get information about how a named device implements Occupancy measurements
 	DescribeOccupancy(context.Context, *DescribeOccupancyRequest) (*OccupancySupport, error)
 	mustEmbedUnimplementedOccupancySensorInfoServer()
 }
 
-// UnimplementedOccupancySensorInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedOccupancySensorInfoServer struct {
-}
+// UnimplementedOccupancySensorInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedOccupancySensorInfoServer struct{}
 
 func (UnimplementedOccupancySensorInfoServer) DescribeOccupancy(context.Context, *DescribeOccupancyRequest) (*OccupancySupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeOccupancy not implemented")
 }
 func (UnimplementedOccupancySensorInfoServer) mustEmbedUnimplementedOccupancySensorInfoServer() {}
+func (UnimplementedOccupancySensorInfoServer) testEmbeddedByValue()                             {}
 
 // UnsafeOccupancySensorInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to OccupancySensorInfoServer will
@@ -232,6 +238,13 @@ type UnsafeOccupancySensorInfoServer interface {
 }
 
 func RegisterOccupancySensorInfoServer(s grpc.ServiceRegistrar, srv OccupancySensorInfoServer) {
+	// If the following call pancis, it indicates UnimplementedOccupancySensorInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&OccupancySensorInfo_ServiceDesc, srv)
 }
 
