@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/on_off.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	OnOffApi_GetOnOff_FullMethodName    = "/smartcore.traits.OnOffApi/GetOnOff"
@@ -27,13 +27,15 @@ const (
 // OnOffApiClient is the client API for OnOffApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Applicable to devices that have a basic binary on and off functionality.
 type OnOffApiClient interface {
 	// Get the current state of the device
 	GetOnOff(ctx context.Context, in *GetOnOffRequest, opts ...grpc.CallOption) (*OnOff, error)
 	// Update the device to be on or off
 	UpdateOnOff(ctx context.Context, in *UpdateOnOffRequest, opts ...grpc.CallOption) (*OnOff, error)
 	// Get notified of changes to the OnOff state of a device
-	PullOnOff(ctx context.Context, in *PullOnOffRequest, opts ...grpc.CallOption) (OnOffApi_PullOnOffClient, error)
+	PullOnOff(ctx context.Context, in *PullOnOffRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullOnOffResponse], error)
 }
 
 type onOffApiClient struct {
@@ -45,8 +47,9 @@ func NewOnOffApiClient(cc grpc.ClientConnInterface) OnOffApiClient {
 }
 
 func (c *onOffApiClient) GetOnOff(ctx context.Context, in *GetOnOffRequest, opts ...grpc.CallOption) (*OnOff, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(OnOff)
-	err := c.cc.Invoke(ctx, OnOffApi_GetOnOff_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OnOffApi_GetOnOff_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,20 +57,22 @@ func (c *onOffApiClient) GetOnOff(ctx context.Context, in *GetOnOffRequest, opts
 }
 
 func (c *onOffApiClient) UpdateOnOff(ctx context.Context, in *UpdateOnOffRequest, opts ...grpc.CallOption) (*OnOff, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(OnOff)
-	err := c.cc.Invoke(ctx, OnOffApi_UpdateOnOff_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OnOffApi_UpdateOnOff_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *onOffApiClient) PullOnOff(ctx context.Context, in *PullOnOffRequest, opts ...grpc.CallOption) (OnOffApi_PullOnOffClient, error) {
-	stream, err := c.cc.NewStream(ctx, &OnOffApi_ServiceDesc.Streams[0], OnOffApi_PullOnOff_FullMethodName, opts...)
+func (c *onOffApiClient) PullOnOff(ctx context.Context, in *PullOnOffRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullOnOffResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &OnOffApi_ServiceDesc.Streams[0], OnOffApi_PullOnOff_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &onOffApiPullOnOffClient{stream}
+	x := &grpc.GenericClientStream[PullOnOffRequest, PullOnOffResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -77,39 +82,30 @@ func (c *onOffApiClient) PullOnOff(ctx context.Context, in *PullOnOffRequest, op
 	return x, nil
 }
 
-type OnOffApi_PullOnOffClient interface {
-	Recv() (*PullOnOffResponse, error)
-	grpc.ClientStream
-}
-
-type onOffApiPullOnOffClient struct {
-	grpc.ClientStream
-}
-
-func (x *onOffApiPullOnOffClient) Recv() (*PullOnOffResponse, error) {
-	m := new(PullOnOffResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type OnOffApi_PullOnOffClient = grpc.ServerStreamingClient[PullOnOffResponse]
 
 // OnOffApiServer is the server API for OnOffApi service.
 // All implementations must embed UnimplementedOnOffApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Applicable to devices that have a basic binary on and off functionality.
 type OnOffApiServer interface {
 	// Get the current state of the device
 	GetOnOff(context.Context, *GetOnOffRequest) (*OnOff, error)
 	// Update the device to be on or off
 	UpdateOnOff(context.Context, *UpdateOnOffRequest) (*OnOff, error)
 	// Get notified of changes to the OnOff state of a device
-	PullOnOff(*PullOnOffRequest, OnOffApi_PullOnOffServer) error
+	PullOnOff(*PullOnOffRequest, grpc.ServerStreamingServer[PullOnOffResponse]) error
 	mustEmbedUnimplementedOnOffApiServer()
 }
 
-// UnimplementedOnOffApiServer must be embedded to have forward compatible implementations.
-type UnimplementedOnOffApiServer struct {
-}
+// UnimplementedOnOffApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedOnOffApiServer struct{}
 
 func (UnimplementedOnOffApiServer) GetOnOff(context.Context, *GetOnOffRequest) (*OnOff, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetOnOff not implemented")
@@ -117,10 +113,11 @@ func (UnimplementedOnOffApiServer) GetOnOff(context.Context, *GetOnOffRequest) (
 func (UnimplementedOnOffApiServer) UpdateOnOff(context.Context, *UpdateOnOffRequest) (*OnOff, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateOnOff not implemented")
 }
-func (UnimplementedOnOffApiServer) PullOnOff(*PullOnOffRequest, OnOffApi_PullOnOffServer) error {
+func (UnimplementedOnOffApiServer) PullOnOff(*PullOnOffRequest, grpc.ServerStreamingServer[PullOnOffResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullOnOff not implemented")
 }
 func (UnimplementedOnOffApiServer) mustEmbedUnimplementedOnOffApiServer() {}
+func (UnimplementedOnOffApiServer) testEmbeddedByValue()                  {}
 
 // UnsafeOnOffApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to OnOffApiServer will
@@ -130,6 +127,13 @@ type UnsafeOnOffApiServer interface {
 }
 
 func RegisterOnOffApiServer(s grpc.ServiceRegistrar, srv OnOffApiServer) {
+	// If the following call pancis, it indicates UnimplementedOnOffApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&OnOffApi_ServiceDesc, srv)
 }
 
@@ -174,21 +178,11 @@ func _OnOffApi_PullOnOff_Handler(srv interface{}, stream grpc.ServerStream) erro
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(OnOffApiServer).PullOnOff(m, &onOffApiPullOnOffServer{stream})
+	return srv.(OnOffApiServer).PullOnOff(m, &grpc.GenericServerStream[PullOnOffRequest, PullOnOffResponse]{ServerStream: stream})
 }
 
-type OnOffApi_PullOnOffServer interface {
-	Send(*PullOnOffResponse) error
-	grpc.ServerStream
-}
-
-type onOffApiPullOnOffServer struct {
-	grpc.ServerStream
-}
-
-func (x *onOffApiPullOnOffServer) Send(m *PullOnOffResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type OnOffApi_PullOnOffServer = grpc.ServerStreamingServer[PullOnOffResponse]
 
 // OnOffApi_ServiceDesc is the grpc.ServiceDesc for OnOffApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -223,6 +217,8 @@ const (
 // OnOffInfoClient is the client API for OnOffInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type OnOffInfoClient interface {
 	// Get information about how a named device implements OnOff features
 	DescribeOnOff(ctx context.Context, in *DescribeOnOffRequest, opts ...grpc.CallOption) (*OnOffSupport, error)
@@ -237,8 +233,9 @@ func NewOnOffInfoClient(cc grpc.ClientConnInterface) OnOffInfoClient {
 }
 
 func (c *onOffInfoClient) DescribeOnOff(ctx context.Context, in *DescribeOnOffRequest, opts ...grpc.CallOption) (*OnOffSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(OnOffSupport)
-	err := c.cc.Invoke(ctx, OnOffInfo_DescribeOnOff_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OnOffInfo_DescribeOnOff_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -247,21 +244,27 @@ func (c *onOffInfoClient) DescribeOnOff(ctx context.Context, in *DescribeOnOffRe
 
 // OnOffInfoServer is the server API for OnOffInfo service.
 // All implementations must embed UnimplementedOnOffInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type OnOffInfoServer interface {
 	// Get information about how a named device implements OnOff features
 	DescribeOnOff(context.Context, *DescribeOnOffRequest) (*OnOffSupport, error)
 	mustEmbedUnimplementedOnOffInfoServer()
 }
 
-// UnimplementedOnOffInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedOnOffInfoServer struct {
-}
+// UnimplementedOnOffInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedOnOffInfoServer struct{}
 
 func (UnimplementedOnOffInfoServer) DescribeOnOff(context.Context, *DescribeOnOffRequest) (*OnOffSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeOnOff not implemented")
 }
 func (UnimplementedOnOffInfoServer) mustEmbedUnimplementedOnOffInfoServer() {}
+func (UnimplementedOnOffInfoServer) testEmbeddedByValue()                   {}
 
 // UnsafeOnOffInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to OnOffInfoServer will
@@ -271,6 +274,13 @@ type UnsafeOnOffInfoServer interface {
 }
 
 func RegisterOnOffInfoServer(s grpc.ServiceRegistrar, srv OnOffInfoServer) {
+	// If the following call pancis, it indicates UnimplementedOnOffInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&OnOffInfo_ServiceDesc, srv)
 }
 
