@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/open_close.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	OpenCloseApi_GetPositions_FullMethodName    = "/smartcore.traits.OpenCloseApi/GetPositions"
@@ -28,13 +28,15 @@ const (
 // OpenCloseApiClient is the client API for OpenCloseApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Applicable to devices that can open or close. Think window contact closure, window blind, or garage door.
 type OpenCloseApiClient interface {
 	GetPositions(ctx context.Context, in *GetOpenClosePositionsRequest, opts ...grpc.CallOption) (*OpenClosePositions, error)
 	UpdatePositions(ctx context.Context, in *UpdateOpenClosePositionsRequest, opts ...grpc.CallOption) (*OpenClosePositions, error)
 	// Stop causes any changes being performed by the underlying device to stop. Will return the current state.
 	Stop(ctx context.Context, in *StopOpenCloseRequest, opts ...grpc.CallOption) (*OpenClosePositions, error)
 	// Get notified of changes to the OpenCloseState of a device
-	PullPositions(ctx context.Context, in *PullOpenClosePositionsRequest, opts ...grpc.CallOption) (OpenCloseApi_PullPositionsClient, error)
+	PullPositions(ctx context.Context, in *PullOpenClosePositionsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullOpenClosePositionsResponse], error)
 }
 
 type openCloseApiClient struct {
@@ -46,8 +48,9 @@ func NewOpenCloseApiClient(cc grpc.ClientConnInterface) OpenCloseApiClient {
 }
 
 func (c *openCloseApiClient) GetPositions(ctx context.Context, in *GetOpenClosePositionsRequest, opts ...grpc.CallOption) (*OpenClosePositions, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(OpenClosePositions)
-	err := c.cc.Invoke(ctx, OpenCloseApi_GetPositions_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OpenCloseApi_GetPositions_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -55,8 +58,9 @@ func (c *openCloseApiClient) GetPositions(ctx context.Context, in *GetOpenCloseP
 }
 
 func (c *openCloseApiClient) UpdatePositions(ctx context.Context, in *UpdateOpenClosePositionsRequest, opts ...grpc.CallOption) (*OpenClosePositions, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(OpenClosePositions)
-	err := c.cc.Invoke(ctx, OpenCloseApi_UpdatePositions_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OpenCloseApi_UpdatePositions_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -64,20 +68,22 @@ func (c *openCloseApiClient) UpdatePositions(ctx context.Context, in *UpdateOpen
 }
 
 func (c *openCloseApiClient) Stop(ctx context.Context, in *StopOpenCloseRequest, opts ...grpc.CallOption) (*OpenClosePositions, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(OpenClosePositions)
-	err := c.cc.Invoke(ctx, OpenCloseApi_Stop_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OpenCloseApi_Stop_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *openCloseApiClient) PullPositions(ctx context.Context, in *PullOpenClosePositionsRequest, opts ...grpc.CallOption) (OpenCloseApi_PullPositionsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &OpenCloseApi_ServiceDesc.Streams[0], OpenCloseApi_PullPositions_FullMethodName, opts...)
+func (c *openCloseApiClient) PullPositions(ctx context.Context, in *PullOpenClosePositionsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullOpenClosePositionsResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &OpenCloseApi_ServiceDesc.Streams[0], OpenCloseApi_PullPositions_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &openCloseApiPullPositionsClient{stream}
+	x := &grpc.GenericClientStream[PullOpenClosePositionsRequest, PullOpenClosePositionsResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -87,39 +93,30 @@ func (c *openCloseApiClient) PullPositions(ctx context.Context, in *PullOpenClos
 	return x, nil
 }
 
-type OpenCloseApi_PullPositionsClient interface {
-	Recv() (*PullOpenClosePositionsResponse, error)
-	grpc.ClientStream
-}
-
-type openCloseApiPullPositionsClient struct {
-	grpc.ClientStream
-}
-
-func (x *openCloseApiPullPositionsClient) Recv() (*PullOpenClosePositionsResponse, error) {
-	m := new(PullOpenClosePositionsResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type OpenCloseApi_PullPositionsClient = grpc.ServerStreamingClient[PullOpenClosePositionsResponse]
 
 // OpenCloseApiServer is the server API for OpenCloseApi service.
 // All implementations must embed UnimplementedOpenCloseApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Applicable to devices that can open or close. Think window contact closure, window blind, or garage door.
 type OpenCloseApiServer interface {
 	GetPositions(context.Context, *GetOpenClosePositionsRequest) (*OpenClosePositions, error)
 	UpdatePositions(context.Context, *UpdateOpenClosePositionsRequest) (*OpenClosePositions, error)
 	// Stop causes any changes being performed by the underlying device to stop. Will return the current state.
 	Stop(context.Context, *StopOpenCloseRequest) (*OpenClosePositions, error)
 	// Get notified of changes to the OpenCloseState of a device
-	PullPositions(*PullOpenClosePositionsRequest, OpenCloseApi_PullPositionsServer) error
+	PullPositions(*PullOpenClosePositionsRequest, grpc.ServerStreamingServer[PullOpenClosePositionsResponse]) error
 	mustEmbedUnimplementedOpenCloseApiServer()
 }
 
-// UnimplementedOpenCloseApiServer must be embedded to have forward compatible implementations.
-type UnimplementedOpenCloseApiServer struct {
-}
+// UnimplementedOpenCloseApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedOpenCloseApiServer struct{}
 
 func (UnimplementedOpenCloseApiServer) GetPositions(context.Context, *GetOpenClosePositionsRequest) (*OpenClosePositions, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetPositions not implemented")
@@ -130,10 +127,11 @@ func (UnimplementedOpenCloseApiServer) UpdatePositions(context.Context, *UpdateO
 func (UnimplementedOpenCloseApiServer) Stop(context.Context, *StopOpenCloseRequest) (*OpenClosePositions, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Stop not implemented")
 }
-func (UnimplementedOpenCloseApiServer) PullPositions(*PullOpenClosePositionsRequest, OpenCloseApi_PullPositionsServer) error {
+func (UnimplementedOpenCloseApiServer) PullPositions(*PullOpenClosePositionsRequest, grpc.ServerStreamingServer[PullOpenClosePositionsResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullPositions not implemented")
 }
 func (UnimplementedOpenCloseApiServer) mustEmbedUnimplementedOpenCloseApiServer() {}
+func (UnimplementedOpenCloseApiServer) testEmbeddedByValue()                      {}
 
 // UnsafeOpenCloseApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to OpenCloseApiServer will
@@ -143,6 +141,13 @@ type UnsafeOpenCloseApiServer interface {
 }
 
 func RegisterOpenCloseApiServer(s grpc.ServiceRegistrar, srv OpenCloseApiServer) {
+	// If the following call pancis, it indicates UnimplementedOpenCloseApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&OpenCloseApi_ServiceDesc, srv)
 }
 
@@ -205,21 +210,11 @@ func _OpenCloseApi_PullPositions_Handler(srv interface{}, stream grpc.ServerStre
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(OpenCloseApiServer).PullPositions(m, &openCloseApiPullPositionsServer{stream})
+	return srv.(OpenCloseApiServer).PullPositions(m, &grpc.GenericServerStream[PullOpenClosePositionsRequest, PullOpenClosePositionsResponse]{ServerStream: stream})
 }
 
-type OpenCloseApi_PullPositionsServer interface {
-	Send(*PullOpenClosePositionsResponse) error
-	grpc.ServerStream
-}
-
-type openCloseApiPullPositionsServer struct {
-	grpc.ServerStream
-}
-
-func (x *openCloseApiPullPositionsServer) Send(m *PullOpenClosePositionsResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type OpenCloseApi_PullPositionsServer = grpc.ServerStreamingServer[PullOpenClosePositionsResponse]
 
 // OpenCloseApi_ServiceDesc is the grpc.ServiceDesc for OpenCloseApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -258,6 +253,8 @@ const (
 // OpenCloseInfoClient is the client API for OpenCloseInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type OpenCloseInfoClient interface {
 	// Get information about how a named device implements Positions features
 	DescribePositions(ctx context.Context, in *DescribePositionsRequest, opts ...grpc.CallOption) (*PositionsSupport, error)
@@ -272,8 +269,9 @@ func NewOpenCloseInfoClient(cc grpc.ClientConnInterface) OpenCloseInfoClient {
 }
 
 func (c *openCloseInfoClient) DescribePositions(ctx context.Context, in *DescribePositionsRequest, opts ...grpc.CallOption) (*PositionsSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PositionsSupport)
-	err := c.cc.Invoke(ctx, OpenCloseInfo_DescribePositions_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, OpenCloseInfo_DescribePositions_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -282,21 +280,27 @@ func (c *openCloseInfoClient) DescribePositions(ctx context.Context, in *Describ
 
 // OpenCloseInfoServer is the server API for OpenCloseInfo service.
 // All implementations must embed UnimplementedOpenCloseInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type OpenCloseInfoServer interface {
 	// Get information about how a named device implements Positions features
 	DescribePositions(context.Context, *DescribePositionsRequest) (*PositionsSupport, error)
 	mustEmbedUnimplementedOpenCloseInfoServer()
 }
 
-// UnimplementedOpenCloseInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedOpenCloseInfoServer struct {
-}
+// UnimplementedOpenCloseInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedOpenCloseInfoServer struct{}
 
 func (UnimplementedOpenCloseInfoServer) DescribePositions(context.Context, *DescribePositionsRequest) (*PositionsSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribePositions not implemented")
 }
 func (UnimplementedOpenCloseInfoServer) mustEmbedUnimplementedOpenCloseInfoServer() {}
+func (UnimplementedOpenCloseInfoServer) testEmbeddedByValue()                       {}
 
 // UnsafeOpenCloseInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to OpenCloseInfoServer will
@@ -306,6 +310,13 @@ type UnsafeOpenCloseInfoServer interface {
 }
 
 func RegisterOpenCloseInfoServer(s grpc.ServiceRegistrar, srv OpenCloseInfoServer) {
+	// If the following call pancis, it indicates UnimplementedOpenCloseInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&OpenCloseInfo_ServiceDesc, srv)
 }
 
