@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/parent.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	ParentApi_ListChildren_FullMethodName = "/smartcore.traits.ParentApi/ListChildren"
@@ -26,11 +26,13 @@ const (
 // ParentApiClient is the client API for ParentApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// ParentApi describes the capabilities of a device to have child devices.
 type ParentApiClient interface {
 	// ListChildren returns children associated with the named parent device.
 	ListChildren(ctx context.Context, in *ListChildrenRequest, opts ...grpc.CallOption) (*ListChildrenResponse, error)
 	// PullChildren subscribes to changes in the children associated with the named parent device.
-	PullChildren(ctx context.Context, in *PullChildrenRequest, opts ...grpc.CallOption) (ParentApi_PullChildrenClient, error)
+	PullChildren(ctx context.Context, in *PullChildrenRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullChildrenResponse], error)
 }
 
 type parentApiClient struct {
@@ -42,20 +44,22 @@ func NewParentApiClient(cc grpc.ClientConnInterface) ParentApiClient {
 }
 
 func (c *parentApiClient) ListChildren(ctx context.Context, in *ListChildrenRequest, opts ...grpc.CallOption) (*ListChildrenResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListChildrenResponse)
-	err := c.cc.Invoke(ctx, ParentApi_ListChildren_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ParentApi_ListChildren_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *parentApiClient) PullChildren(ctx context.Context, in *PullChildrenRequest, opts ...grpc.CallOption) (ParentApi_PullChildrenClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ParentApi_ServiceDesc.Streams[0], ParentApi_PullChildren_FullMethodName, opts...)
+func (c *parentApiClient) PullChildren(ctx context.Context, in *PullChildrenRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullChildrenResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ParentApi_ServiceDesc.Streams[0], ParentApi_PullChildren_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &parentApiPullChildrenClient{stream}
+	x := &grpc.GenericClientStream[PullChildrenRequest, PullChildrenResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -65,45 +69,37 @@ func (c *parentApiClient) PullChildren(ctx context.Context, in *PullChildrenRequ
 	return x, nil
 }
 
-type ParentApi_PullChildrenClient interface {
-	Recv() (*PullChildrenResponse, error)
-	grpc.ClientStream
-}
-
-type parentApiPullChildrenClient struct {
-	grpc.ClientStream
-}
-
-func (x *parentApiPullChildrenClient) Recv() (*PullChildrenResponse, error) {
-	m := new(PullChildrenResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ParentApi_PullChildrenClient = grpc.ServerStreamingClient[PullChildrenResponse]
 
 // ParentApiServer is the server API for ParentApi service.
 // All implementations must embed UnimplementedParentApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// ParentApi describes the capabilities of a device to have child devices.
 type ParentApiServer interface {
 	// ListChildren returns children associated with the named parent device.
 	ListChildren(context.Context, *ListChildrenRequest) (*ListChildrenResponse, error)
 	// PullChildren subscribes to changes in the children associated with the named parent device.
-	PullChildren(*PullChildrenRequest, ParentApi_PullChildrenServer) error
+	PullChildren(*PullChildrenRequest, grpc.ServerStreamingServer[PullChildrenResponse]) error
 	mustEmbedUnimplementedParentApiServer()
 }
 
-// UnimplementedParentApiServer must be embedded to have forward compatible implementations.
-type UnimplementedParentApiServer struct {
-}
+// UnimplementedParentApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedParentApiServer struct{}
 
 func (UnimplementedParentApiServer) ListChildren(context.Context, *ListChildrenRequest) (*ListChildrenResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListChildren not implemented")
 }
-func (UnimplementedParentApiServer) PullChildren(*PullChildrenRequest, ParentApi_PullChildrenServer) error {
+func (UnimplementedParentApiServer) PullChildren(*PullChildrenRequest, grpc.ServerStreamingServer[PullChildrenResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullChildren not implemented")
 }
 func (UnimplementedParentApiServer) mustEmbedUnimplementedParentApiServer() {}
+func (UnimplementedParentApiServer) testEmbeddedByValue()                   {}
 
 // UnsafeParentApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ParentApiServer will
@@ -113,6 +109,13 @@ type UnsafeParentApiServer interface {
 }
 
 func RegisterParentApiServer(s grpc.ServiceRegistrar, srv ParentApiServer) {
+	// If the following call pancis, it indicates UnimplementedParentApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ParentApi_ServiceDesc, srv)
 }
 
@@ -139,21 +142,11 @@ func _ParentApi_PullChildren_Handler(srv interface{}, stream grpc.ServerStream) 
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ParentApiServer).PullChildren(m, &parentApiPullChildrenServer{stream})
+	return srv.(ParentApiServer).PullChildren(m, &grpc.GenericServerStream[PullChildrenRequest, PullChildrenResponse]{ServerStream: stream})
 }
 
-type ParentApi_PullChildrenServer interface {
-	Send(*PullChildrenResponse) error
-	grpc.ServerStream
-}
-
-type parentApiPullChildrenServer struct {
-	grpc.ServerStream
-}
-
-func (x *parentApiPullChildrenServer) Send(m *PullChildrenResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ParentApi_PullChildrenServer = grpc.ServerStreamingServer[PullChildrenResponse]
 
 // ParentApi_ServiceDesc is the grpc.ServiceDesc for ParentApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -177,8 +170,6 @@ var ParentApi_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "traits/parent.proto",
 }
 
-const ()
-
 // ParentInfoClient is the client API for ParentInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
@@ -195,16 +186,20 @@ func NewParentInfoClient(cc grpc.ClientConnInterface) ParentInfoClient {
 
 // ParentInfoServer is the server API for ParentInfo service.
 // All implementations must embed UnimplementedParentInfoServer
-// for forward compatibility
+// for forward compatibility.
 type ParentInfoServer interface {
 	mustEmbedUnimplementedParentInfoServer()
 }
 
-// UnimplementedParentInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedParentInfoServer struct {
-}
+// UnimplementedParentInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedParentInfoServer struct{}
 
 func (UnimplementedParentInfoServer) mustEmbedUnimplementedParentInfoServer() {}
+func (UnimplementedParentInfoServer) testEmbeddedByValue()                    {}
 
 // UnsafeParentInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ParentInfoServer will
@@ -214,6 +209,13 @@ type UnsafeParentInfoServer interface {
 }
 
 func RegisterParentInfoServer(s grpc.ServiceRegistrar, srv ParentInfoServer) {
+	// If the following call pancis, it indicates UnimplementedParentInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ParentInfo_ServiceDesc, srv)
 }
 
