@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/ptz.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	PtzApi_GetPtz_FullMethodName       = "/smartcore.traits.PtzApi/GetPtz"
@@ -29,6 +29,8 @@ const (
 // PtzApiClient is the client API for PtzApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Allows interaction with a pan-tilt-zoom device, typically a camera.
 type PtzApiClient interface {
 	// Set the target state for the device
 	GetPtz(ctx context.Context, in *GetPtzRequest, opts ...grpc.CallOption) (*Ptz, error)
@@ -39,7 +41,7 @@ type PtzApiClient interface {
 	// Create a preset for the ptz position. If no preset ptz position is specified then use the current position
 	CreatePreset(ctx context.Context, in *CreatePtzPresetRequest, opts ...grpc.CallOption) (*PtzPreset, error)
 	// Get notified of changes to the OnOffState of a device
-	PullPtz(ctx context.Context, in *PullPtzRequest, opts ...grpc.CallOption) (PtzApi_PullPtzClient, error)
+	PullPtz(ctx context.Context, in *PullPtzRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullPtzResponse], error)
 }
 
 type ptzApiClient struct {
@@ -51,8 +53,9 @@ func NewPtzApiClient(cc grpc.ClientConnInterface) PtzApiClient {
 }
 
 func (c *ptzApiClient) GetPtz(ctx context.Context, in *GetPtzRequest, opts ...grpc.CallOption) (*Ptz, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Ptz)
-	err := c.cc.Invoke(ctx, PtzApi_GetPtz_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PtzApi_GetPtz_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -60,8 +63,9 @@ func (c *ptzApiClient) GetPtz(ctx context.Context, in *GetPtzRequest, opts ...gr
 }
 
 func (c *ptzApiClient) UpdatePtz(ctx context.Context, in *UpdatePtzRequest, opts ...grpc.CallOption) (*Ptz, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Ptz)
-	err := c.cc.Invoke(ctx, PtzApi_UpdatePtz_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PtzApi_UpdatePtz_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -69,8 +73,9 @@ func (c *ptzApiClient) UpdatePtz(ctx context.Context, in *UpdatePtzRequest, opts
 }
 
 func (c *ptzApiClient) Stop(ctx context.Context, in *StopPtzRequest, opts ...grpc.CallOption) (*Ptz, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Ptz)
-	err := c.cc.Invoke(ctx, PtzApi_Stop_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PtzApi_Stop_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -78,20 +83,22 @@ func (c *ptzApiClient) Stop(ctx context.Context, in *StopPtzRequest, opts ...grp
 }
 
 func (c *ptzApiClient) CreatePreset(ctx context.Context, in *CreatePtzPresetRequest, opts ...grpc.CallOption) (*PtzPreset, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PtzPreset)
-	err := c.cc.Invoke(ctx, PtzApi_CreatePreset_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PtzApi_CreatePreset_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *ptzApiClient) PullPtz(ctx context.Context, in *PullPtzRequest, opts ...grpc.CallOption) (PtzApi_PullPtzClient, error) {
-	stream, err := c.cc.NewStream(ctx, &PtzApi_ServiceDesc.Streams[0], PtzApi_PullPtz_FullMethodName, opts...)
+func (c *ptzApiClient) PullPtz(ctx context.Context, in *PullPtzRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullPtzResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &PtzApi_ServiceDesc.Streams[0], PtzApi_PullPtz_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &ptzApiPullPtzClient{stream}
+	x := &grpc.GenericClientStream[PullPtzRequest, PullPtzResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -101,26 +108,14 @@ func (c *ptzApiClient) PullPtz(ctx context.Context, in *PullPtzRequest, opts ...
 	return x, nil
 }
 
-type PtzApi_PullPtzClient interface {
-	Recv() (*PullPtzResponse, error)
-	grpc.ClientStream
-}
-
-type ptzApiPullPtzClient struct {
-	grpc.ClientStream
-}
-
-func (x *ptzApiPullPtzClient) Recv() (*PullPtzResponse, error) {
-	m := new(PullPtzResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type PtzApi_PullPtzClient = grpc.ServerStreamingClient[PullPtzResponse]
 
 // PtzApiServer is the server API for PtzApi service.
 // All implementations must embed UnimplementedPtzApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Allows interaction with a pan-tilt-zoom device, typically a camera.
 type PtzApiServer interface {
 	// Set the target state for the device
 	GetPtz(context.Context, *GetPtzRequest) (*Ptz, error)
@@ -131,13 +126,16 @@ type PtzApiServer interface {
 	// Create a preset for the ptz position. If no preset ptz position is specified then use the current position
 	CreatePreset(context.Context, *CreatePtzPresetRequest) (*PtzPreset, error)
 	// Get notified of changes to the OnOffState of a device
-	PullPtz(*PullPtzRequest, PtzApi_PullPtzServer) error
+	PullPtz(*PullPtzRequest, grpc.ServerStreamingServer[PullPtzResponse]) error
 	mustEmbedUnimplementedPtzApiServer()
 }
 
-// UnimplementedPtzApiServer must be embedded to have forward compatible implementations.
-type UnimplementedPtzApiServer struct {
-}
+// UnimplementedPtzApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPtzApiServer struct{}
 
 func (UnimplementedPtzApiServer) GetPtz(context.Context, *GetPtzRequest) (*Ptz, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetPtz not implemented")
@@ -151,10 +149,11 @@ func (UnimplementedPtzApiServer) Stop(context.Context, *StopPtzRequest) (*Ptz, e
 func (UnimplementedPtzApiServer) CreatePreset(context.Context, *CreatePtzPresetRequest) (*PtzPreset, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreatePreset not implemented")
 }
-func (UnimplementedPtzApiServer) PullPtz(*PullPtzRequest, PtzApi_PullPtzServer) error {
+func (UnimplementedPtzApiServer) PullPtz(*PullPtzRequest, grpc.ServerStreamingServer[PullPtzResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullPtz not implemented")
 }
 func (UnimplementedPtzApiServer) mustEmbedUnimplementedPtzApiServer() {}
+func (UnimplementedPtzApiServer) testEmbeddedByValue()                {}
 
 // UnsafePtzApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PtzApiServer will
@@ -164,6 +163,13 @@ type UnsafePtzApiServer interface {
 }
 
 func RegisterPtzApiServer(s grpc.ServiceRegistrar, srv PtzApiServer) {
+	// If the following call pancis, it indicates UnimplementedPtzApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&PtzApi_ServiceDesc, srv)
 }
 
@@ -244,21 +250,11 @@ func _PtzApi_PullPtz_Handler(srv interface{}, stream grpc.ServerStream) error {
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(PtzApiServer).PullPtz(m, &ptzApiPullPtzServer{stream})
+	return srv.(PtzApiServer).PullPtz(m, &grpc.GenericServerStream[PullPtzRequest, PullPtzResponse]{ServerStream: stream})
 }
 
-type PtzApi_PullPtzServer interface {
-	Send(*PullPtzResponse) error
-	grpc.ServerStream
-}
-
-type ptzApiPullPtzServer struct {
-	grpc.ServerStream
-}
-
-func (x *ptzApiPullPtzServer) Send(m *PullPtzResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type PtzApi_PullPtzServer = grpc.ServerStreamingServer[PullPtzResponse]
 
 // PtzApi_ServiceDesc is the grpc.ServiceDesc for PtzApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -301,6 +297,8 @@ const (
 // PtzInfoClient is the client API for PtzInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type PtzInfoClient interface {
 	// Get information about how a named device implements Ptz features
 	DescribePtz(ctx context.Context, in *DescribePtzRequest, opts ...grpc.CallOption) (*PtzSupport, error)
@@ -315,8 +313,9 @@ func NewPtzInfoClient(cc grpc.ClientConnInterface) PtzInfoClient {
 }
 
 func (c *ptzInfoClient) DescribePtz(ctx context.Context, in *DescribePtzRequest, opts ...grpc.CallOption) (*PtzSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PtzSupport)
-	err := c.cc.Invoke(ctx, PtzInfo_DescribePtz_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PtzInfo_DescribePtz_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -325,21 +324,27 @@ func (c *ptzInfoClient) DescribePtz(ctx context.Context, in *DescribePtzRequest,
 
 // PtzInfoServer is the server API for PtzInfo service.
 // All implementations must embed UnimplementedPtzInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type PtzInfoServer interface {
 	// Get information about how a named device implements Ptz features
 	DescribePtz(context.Context, *DescribePtzRequest) (*PtzSupport, error)
 	mustEmbedUnimplementedPtzInfoServer()
 }
 
-// UnimplementedPtzInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedPtzInfoServer struct {
-}
+// UnimplementedPtzInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPtzInfoServer struct{}
 
 func (UnimplementedPtzInfoServer) DescribePtz(context.Context, *DescribePtzRequest) (*PtzSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribePtz not implemented")
 }
 func (UnimplementedPtzInfoServer) mustEmbedUnimplementedPtzInfoServer() {}
+func (UnimplementedPtzInfoServer) testEmbeddedByValue()                 {}
 
 // UnsafePtzInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PtzInfoServer will
@@ -349,6 +354,13 @@ type UnsafePtzInfoServer interface {
 }
 
 func RegisterPtzInfoServer(s grpc.ServiceRegistrar, srv PtzInfoServer) {
+	// If the following call pancis, it indicates UnimplementedPtzInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&PtzInfo_ServiceDesc, srv)
 }
 
