@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/publication.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	PublicationApi_CreatePublication_FullMethodName      = "/smartcore.traits.PublicationApi/CreatePublication"
@@ -32,6 +32,11 @@ const (
 // PublicationApiClient is the client API for PublicationApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// The PublicationApi describes the capabilities of a device to manage and publish content.
+// This could be used by a central device to publish configuration updates for other devices to receive.
+// The primary resource of this trait is a collection of Publications.
+// Publication comprises of some binary data and metadata surrounding the versioning and publication of that data.
 type PublicationApiClient interface {
 	// Create a new publication.
 	// The returned Publication will include an id unique for this new publication within the named device.
@@ -51,12 +56,12 @@ type PublicationApiClient interface {
 	DeletePublication(ctx context.Context, in *DeletePublicationRequest, opts ...grpc.CallOption) (*Publication, error)
 	// Subscribe to changes for a single publication.
 	// The stream is closed if the publication is deleted.
-	PullPublication(ctx context.Context, in *PullPublicationRequest, opts ...grpc.CallOption) (PublicationApi_PullPublicationClient, error)
+	PullPublication(ctx context.Context, in *PullPublicationRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullPublicationResponse], error)
 	// List all publications available on the device.
 	// The version returned for each publication will be the latest.
 	ListPublications(ctx context.Context, in *ListPublicationsRequest, opts ...grpc.CallOption) (*ListPublicationsResponse, error)
 	// Subscribe to changes in publications on the device.
-	PullPublications(ctx context.Context, in *PullPublicationsRequest, opts ...grpc.CallOption) (PublicationApi_PullPublicationsClient, error)
+	PullPublications(ctx context.Context, in *PullPublicationsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullPublicationsResponse], error)
 	// Acknowledge the receipt of a publication.
 	// This API should only be called when the audience for a publication has received that publication.
 	// Typically the audience will call this method itself after a get or pull publication method returns a new version.
@@ -72,8 +77,9 @@ func NewPublicationApiClient(cc grpc.ClientConnInterface) PublicationApiClient {
 }
 
 func (c *publicationApiClient) CreatePublication(ctx context.Context, in *CreatePublicationRequest, opts ...grpc.CallOption) (*Publication, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Publication)
-	err := c.cc.Invoke(ctx, PublicationApi_CreatePublication_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PublicationApi_CreatePublication_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -81,8 +87,9 @@ func (c *publicationApiClient) CreatePublication(ctx context.Context, in *Create
 }
 
 func (c *publicationApiClient) GetPublication(ctx context.Context, in *GetPublicationRequest, opts ...grpc.CallOption) (*Publication, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Publication)
-	err := c.cc.Invoke(ctx, PublicationApi_GetPublication_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PublicationApi_GetPublication_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -90,8 +97,9 @@ func (c *publicationApiClient) GetPublication(ctx context.Context, in *GetPublic
 }
 
 func (c *publicationApiClient) UpdatePublication(ctx context.Context, in *UpdatePublicationRequest, opts ...grpc.CallOption) (*Publication, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Publication)
-	err := c.cc.Invoke(ctx, PublicationApi_UpdatePublication_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PublicationApi_UpdatePublication_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -99,20 +107,22 @@ func (c *publicationApiClient) UpdatePublication(ctx context.Context, in *Update
 }
 
 func (c *publicationApiClient) DeletePublication(ctx context.Context, in *DeletePublicationRequest, opts ...grpc.CallOption) (*Publication, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Publication)
-	err := c.cc.Invoke(ctx, PublicationApi_DeletePublication_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PublicationApi_DeletePublication_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *publicationApiClient) PullPublication(ctx context.Context, in *PullPublicationRequest, opts ...grpc.CallOption) (PublicationApi_PullPublicationClient, error) {
-	stream, err := c.cc.NewStream(ctx, &PublicationApi_ServiceDesc.Streams[0], PublicationApi_PullPublication_FullMethodName, opts...)
+func (c *publicationApiClient) PullPublication(ctx context.Context, in *PullPublicationRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullPublicationResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &PublicationApi_ServiceDesc.Streams[0], PublicationApi_PullPublication_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &publicationApiPullPublicationClient{stream}
+	x := &grpc.GenericClientStream[PullPublicationRequest, PullPublicationResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -122,38 +132,26 @@ func (c *publicationApiClient) PullPublication(ctx context.Context, in *PullPubl
 	return x, nil
 }
 
-type PublicationApi_PullPublicationClient interface {
-	Recv() (*PullPublicationResponse, error)
-	grpc.ClientStream
-}
-
-type publicationApiPullPublicationClient struct {
-	grpc.ClientStream
-}
-
-func (x *publicationApiPullPublicationClient) Recv() (*PullPublicationResponse, error) {
-	m := new(PullPublicationResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type PublicationApi_PullPublicationClient = grpc.ServerStreamingClient[PullPublicationResponse]
 
 func (c *publicationApiClient) ListPublications(ctx context.Context, in *ListPublicationsRequest, opts ...grpc.CallOption) (*ListPublicationsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListPublicationsResponse)
-	err := c.cc.Invoke(ctx, PublicationApi_ListPublications_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PublicationApi_ListPublications_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *publicationApiClient) PullPublications(ctx context.Context, in *PullPublicationsRequest, opts ...grpc.CallOption) (PublicationApi_PullPublicationsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &PublicationApi_ServiceDesc.Streams[1], PublicationApi_PullPublications_FullMethodName, opts...)
+func (c *publicationApiClient) PullPublications(ctx context.Context, in *PullPublicationsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullPublicationsResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &PublicationApi_ServiceDesc.Streams[1], PublicationApi_PullPublications_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &publicationApiPullPublicationsClient{stream}
+	x := &grpc.GenericClientStream[PullPublicationsRequest, PullPublicationsResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -163,26 +161,13 @@ func (c *publicationApiClient) PullPublications(ctx context.Context, in *PullPub
 	return x, nil
 }
 
-type PublicationApi_PullPublicationsClient interface {
-	Recv() (*PullPublicationsResponse, error)
-	grpc.ClientStream
-}
-
-type publicationApiPullPublicationsClient struct {
-	grpc.ClientStream
-}
-
-func (x *publicationApiPullPublicationsClient) Recv() (*PullPublicationsResponse, error) {
-	m := new(PullPublicationsResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type PublicationApi_PullPublicationsClient = grpc.ServerStreamingClient[PullPublicationsResponse]
 
 func (c *publicationApiClient) AcknowledgePublication(ctx context.Context, in *AcknowledgePublicationRequest, opts ...grpc.CallOption) (*Publication, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Publication)
-	err := c.cc.Invoke(ctx, PublicationApi_AcknowledgePublication_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, PublicationApi_AcknowledgePublication_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -191,7 +176,12 @@ func (c *publicationApiClient) AcknowledgePublication(ctx context.Context, in *A
 
 // PublicationApiServer is the server API for PublicationApi service.
 // All implementations must embed UnimplementedPublicationApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// The PublicationApi describes the capabilities of a device to manage and publish content.
+// This could be used by a central device to publish configuration updates for other devices to receive.
+// The primary resource of this trait is a collection of Publications.
+// Publication comprises of some binary data and metadata surrounding the versioning and publication of that data.
 type PublicationApiServer interface {
 	// Create a new publication.
 	// The returned Publication will include an id unique for this new publication within the named device.
@@ -211,12 +201,12 @@ type PublicationApiServer interface {
 	DeletePublication(context.Context, *DeletePublicationRequest) (*Publication, error)
 	// Subscribe to changes for a single publication.
 	// The stream is closed if the publication is deleted.
-	PullPublication(*PullPublicationRequest, PublicationApi_PullPublicationServer) error
+	PullPublication(*PullPublicationRequest, grpc.ServerStreamingServer[PullPublicationResponse]) error
 	// List all publications available on the device.
 	// The version returned for each publication will be the latest.
 	ListPublications(context.Context, *ListPublicationsRequest) (*ListPublicationsResponse, error)
 	// Subscribe to changes in publications on the device.
-	PullPublications(*PullPublicationsRequest, PublicationApi_PullPublicationsServer) error
+	PullPublications(*PullPublicationsRequest, grpc.ServerStreamingServer[PullPublicationsResponse]) error
 	// Acknowledge the receipt of a publication.
 	// This API should only be called when the audience for a publication has received that publication.
 	// Typically the audience will call this method itself after a get or pull publication method returns a new version.
@@ -224,9 +214,12 @@ type PublicationApiServer interface {
 	mustEmbedUnimplementedPublicationApiServer()
 }
 
-// UnimplementedPublicationApiServer must be embedded to have forward compatible implementations.
-type UnimplementedPublicationApiServer struct {
-}
+// UnimplementedPublicationApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPublicationApiServer struct{}
 
 func (UnimplementedPublicationApiServer) CreatePublication(context.Context, *CreatePublicationRequest) (*Publication, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreatePublication not implemented")
@@ -240,19 +233,20 @@ func (UnimplementedPublicationApiServer) UpdatePublication(context.Context, *Upd
 func (UnimplementedPublicationApiServer) DeletePublication(context.Context, *DeletePublicationRequest) (*Publication, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DeletePublication not implemented")
 }
-func (UnimplementedPublicationApiServer) PullPublication(*PullPublicationRequest, PublicationApi_PullPublicationServer) error {
+func (UnimplementedPublicationApiServer) PullPublication(*PullPublicationRequest, grpc.ServerStreamingServer[PullPublicationResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullPublication not implemented")
 }
 func (UnimplementedPublicationApiServer) ListPublications(context.Context, *ListPublicationsRequest) (*ListPublicationsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListPublications not implemented")
 }
-func (UnimplementedPublicationApiServer) PullPublications(*PullPublicationsRequest, PublicationApi_PullPublicationsServer) error {
+func (UnimplementedPublicationApiServer) PullPublications(*PullPublicationsRequest, grpc.ServerStreamingServer[PullPublicationsResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullPublications not implemented")
 }
 func (UnimplementedPublicationApiServer) AcknowledgePublication(context.Context, *AcknowledgePublicationRequest) (*Publication, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method AcknowledgePublication not implemented")
 }
 func (UnimplementedPublicationApiServer) mustEmbedUnimplementedPublicationApiServer() {}
+func (UnimplementedPublicationApiServer) testEmbeddedByValue()                        {}
 
 // UnsafePublicationApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PublicationApiServer will
@@ -262,6 +256,13 @@ type UnsafePublicationApiServer interface {
 }
 
 func RegisterPublicationApiServer(s grpc.ServiceRegistrar, srv PublicationApiServer) {
+	// If the following call pancis, it indicates UnimplementedPublicationApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&PublicationApi_ServiceDesc, srv)
 }
 
@@ -342,21 +343,11 @@ func _PublicationApi_PullPublication_Handler(srv interface{}, stream grpc.Server
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(PublicationApiServer).PullPublication(m, &publicationApiPullPublicationServer{stream})
+	return srv.(PublicationApiServer).PullPublication(m, &grpc.GenericServerStream[PullPublicationRequest, PullPublicationResponse]{ServerStream: stream})
 }
 
-type PublicationApi_PullPublicationServer interface {
-	Send(*PullPublicationResponse) error
-	grpc.ServerStream
-}
-
-type publicationApiPullPublicationServer struct {
-	grpc.ServerStream
-}
-
-func (x *publicationApiPullPublicationServer) Send(m *PullPublicationResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type PublicationApi_PullPublicationServer = grpc.ServerStreamingServer[PullPublicationResponse]
 
 func _PublicationApi_ListPublications_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ListPublicationsRequest)
@@ -381,21 +372,11 @@ func _PublicationApi_PullPublications_Handler(srv interface{}, stream grpc.Serve
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(PublicationApiServer).PullPublications(m, &publicationApiPullPublicationsServer{stream})
+	return srv.(PublicationApiServer).PullPublications(m, &grpc.GenericServerStream[PullPublicationsRequest, PullPublicationsResponse]{ServerStream: stream})
 }
 
-type PublicationApi_PullPublicationsServer interface {
-	Send(*PullPublicationsResponse) error
-	grpc.ServerStream
-}
-
-type publicationApiPullPublicationsServer struct {
-	grpc.ServerStream
-}
-
-func (x *publicationApiPullPublicationsServer) Send(m *PullPublicationsResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type PublicationApi_PullPublicationsServer = grpc.ServerStreamingServer[PullPublicationsResponse]
 
 func _PublicationApi_AcknowledgePublication_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(AcknowledgePublicationRequest)
