@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/speaker.proto
 
 package traits
@@ -16,8 +16,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	SpeakerApi_GetVolume_FullMethodName    = "/smartcore.traits.SpeakerApi/GetVolume"
@@ -28,12 +28,17 @@ const (
 // SpeakerApiClient is the client API for SpeakerApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Trait for devices that support some kind of audio output. The most basic is a simple on/off speaker, it's optional to
+// support volume adjustment on a scale.
+//
+// For the sakes of this trait "volume" means the combination of a gain and a mute state.
 type SpeakerApiClient interface {
 	// Get the current state of the volume for the device
 	GetVolume(ctx context.Context, in *GetSpeakerVolumeRequest, opts ...grpc.CallOption) (*types.AudioLevel, error)
 	// update the volume state for the device
 	UpdateVolume(ctx context.Context, in *UpdateSpeakerVolumeRequest, opts ...grpc.CallOption) (*types.AudioLevel, error)
-	PullVolume(ctx context.Context, in *PullSpeakerVolumeRequest, opts ...grpc.CallOption) (SpeakerApi_PullVolumeClient, error)
+	PullVolume(ctx context.Context, in *PullSpeakerVolumeRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullSpeakerVolumeResponse], error)
 }
 
 type speakerApiClient struct {
@@ -45,8 +50,9 @@ func NewSpeakerApiClient(cc grpc.ClientConnInterface) SpeakerApiClient {
 }
 
 func (c *speakerApiClient) GetVolume(ctx context.Context, in *GetSpeakerVolumeRequest, opts ...grpc.CallOption) (*types.AudioLevel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(types.AudioLevel)
-	err := c.cc.Invoke(ctx, SpeakerApi_GetVolume_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SpeakerApi_GetVolume_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,20 +60,22 @@ func (c *speakerApiClient) GetVolume(ctx context.Context, in *GetSpeakerVolumeRe
 }
 
 func (c *speakerApiClient) UpdateVolume(ctx context.Context, in *UpdateSpeakerVolumeRequest, opts ...grpc.CallOption) (*types.AudioLevel, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(types.AudioLevel)
-	err := c.cc.Invoke(ctx, SpeakerApi_UpdateVolume_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SpeakerApi_UpdateVolume_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *speakerApiClient) PullVolume(ctx context.Context, in *PullSpeakerVolumeRequest, opts ...grpc.CallOption) (SpeakerApi_PullVolumeClient, error) {
-	stream, err := c.cc.NewStream(ctx, &SpeakerApi_ServiceDesc.Streams[0], SpeakerApi_PullVolume_FullMethodName, opts...)
+func (c *speakerApiClient) PullVolume(ctx context.Context, in *PullSpeakerVolumeRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullSpeakerVolumeResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &SpeakerApi_ServiceDesc.Streams[0], SpeakerApi_PullVolume_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &speakerApiPullVolumeClient{stream}
+	x := &grpc.GenericClientStream[PullSpeakerVolumeRequest, PullSpeakerVolumeResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -77,38 +85,32 @@ func (c *speakerApiClient) PullVolume(ctx context.Context, in *PullSpeakerVolume
 	return x, nil
 }
 
-type SpeakerApi_PullVolumeClient interface {
-	Recv() (*PullSpeakerVolumeResponse, error)
-	grpc.ClientStream
-}
-
-type speakerApiPullVolumeClient struct {
-	grpc.ClientStream
-}
-
-func (x *speakerApiPullVolumeClient) Recv() (*PullSpeakerVolumeResponse, error) {
-	m := new(PullSpeakerVolumeResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type SpeakerApi_PullVolumeClient = grpc.ServerStreamingClient[PullSpeakerVolumeResponse]
 
 // SpeakerApiServer is the server API for SpeakerApi service.
 // All implementations must embed UnimplementedSpeakerApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Trait for devices that support some kind of audio output. The most basic is a simple on/off speaker, it's optional to
+// support volume adjustment on a scale.
+//
+// For the sakes of this trait "volume" means the combination of a gain and a mute state.
 type SpeakerApiServer interface {
 	// Get the current state of the volume for the device
 	GetVolume(context.Context, *GetSpeakerVolumeRequest) (*types.AudioLevel, error)
 	// update the volume state for the device
 	UpdateVolume(context.Context, *UpdateSpeakerVolumeRequest) (*types.AudioLevel, error)
-	PullVolume(*PullSpeakerVolumeRequest, SpeakerApi_PullVolumeServer) error
+	PullVolume(*PullSpeakerVolumeRequest, grpc.ServerStreamingServer[PullSpeakerVolumeResponse]) error
 	mustEmbedUnimplementedSpeakerApiServer()
 }
 
-// UnimplementedSpeakerApiServer must be embedded to have forward compatible implementations.
-type UnimplementedSpeakerApiServer struct {
-}
+// UnimplementedSpeakerApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedSpeakerApiServer struct{}
 
 func (UnimplementedSpeakerApiServer) GetVolume(context.Context, *GetSpeakerVolumeRequest) (*types.AudioLevel, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetVolume not implemented")
@@ -116,10 +118,11 @@ func (UnimplementedSpeakerApiServer) GetVolume(context.Context, *GetSpeakerVolum
 func (UnimplementedSpeakerApiServer) UpdateVolume(context.Context, *UpdateSpeakerVolumeRequest) (*types.AudioLevel, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateVolume not implemented")
 }
-func (UnimplementedSpeakerApiServer) PullVolume(*PullSpeakerVolumeRequest, SpeakerApi_PullVolumeServer) error {
+func (UnimplementedSpeakerApiServer) PullVolume(*PullSpeakerVolumeRequest, grpc.ServerStreamingServer[PullSpeakerVolumeResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullVolume not implemented")
 }
 func (UnimplementedSpeakerApiServer) mustEmbedUnimplementedSpeakerApiServer() {}
+func (UnimplementedSpeakerApiServer) testEmbeddedByValue()                    {}
 
 // UnsafeSpeakerApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to SpeakerApiServer will
@@ -129,6 +132,13 @@ type UnsafeSpeakerApiServer interface {
 }
 
 func RegisterSpeakerApiServer(s grpc.ServiceRegistrar, srv SpeakerApiServer) {
+	// If the following call pancis, it indicates UnimplementedSpeakerApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&SpeakerApi_ServiceDesc, srv)
 }
 
@@ -173,21 +183,11 @@ func _SpeakerApi_PullVolume_Handler(srv interface{}, stream grpc.ServerStream) e
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(SpeakerApiServer).PullVolume(m, &speakerApiPullVolumeServer{stream})
+	return srv.(SpeakerApiServer).PullVolume(m, &grpc.GenericServerStream[PullSpeakerVolumeRequest, PullSpeakerVolumeResponse]{ServerStream: stream})
 }
 
-type SpeakerApi_PullVolumeServer interface {
-	Send(*PullSpeakerVolumeResponse) error
-	grpc.ServerStream
-}
-
-type speakerApiPullVolumeServer struct {
-	grpc.ServerStream
-}
-
-func (x *speakerApiPullVolumeServer) Send(m *PullSpeakerVolumeResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type SpeakerApi_PullVolumeServer = grpc.ServerStreamingServer[PullSpeakerVolumeResponse]
 
 // SpeakerApi_ServiceDesc is the grpc.ServiceDesc for SpeakerApi service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -222,6 +222,8 @@ const (
 // SpeakerInfoClient is the client API for SpeakerInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type SpeakerInfoClient interface {
 	// Get information about how a named device implements volume features
 	DescribeVolume(ctx context.Context, in *DescribeVolumeRequest, opts ...grpc.CallOption) (*VolumeSupport, error)
@@ -236,8 +238,9 @@ func NewSpeakerInfoClient(cc grpc.ClientConnInterface) SpeakerInfoClient {
 }
 
 func (c *speakerInfoClient) DescribeVolume(ctx context.Context, in *DescribeVolumeRequest, opts ...grpc.CallOption) (*VolumeSupport, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(VolumeSupport)
-	err := c.cc.Invoke(ctx, SpeakerInfo_DescribeVolume_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SpeakerInfo_DescribeVolume_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -246,21 +249,27 @@ func (c *speakerInfoClient) DescribeVolume(ctx context.Context, in *DescribeVolu
 
 // SpeakerInfoServer is the server API for SpeakerInfo service.
 // All implementations must embed UnimplementedSpeakerInfoServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Describes the capabilities of a specific named device with respect to this trait.
 type SpeakerInfoServer interface {
 	// Get information about how a named device implements volume features
 	DescribeVolume(context.Context, *DescribeVolumeRequest) (*VolumeSupport, error)
 	mustEmbedUnimplementedSpeakerInfoServer()
 }
 
-// UnimplementedSpeakerInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedSpeakerInfoServer struct {
-}
+// UnimplementedSpeakerInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedSpeakerInfoServer struct{}
 
 func (UnimplementedSpeakerInfoServer) DescribeVolume(context.Context, *DescribeVolumeRequest) (*VolumeSupport, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DescribeVolume not implemented")
 }
 func (UnimplementedSpeakerInfoServer) mustEmbedUnimplementedSpeakerInfoServer() {}
+func (UnimplementedSpeakerInfoServer) testEmbeddedByValue()                     {}
 
 // UnsafeSpeakerInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to SpeakerInfoServer will
@@ -270,6 +279,13 @@ type UnsafeSpeakerInfoServer interface {
 }
 
 func RegisterSpeakerInfoServer(s grpc.ServiceRegistrar, srv SpeakerInfoServer) {
+	// If the following call pancis, it indicates UnimplementedSpeakerInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&SpeakerInfo_ServiceDesc, srv)
 }
 
