@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: traits/vending.proto
 
 package traits
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	VendingApi_ListConsumables_FullMethodName = "/smartcore.traits.VendingApi/ListConsumables"
@@ -33,12 +33,23 @@ const (
 // VendingApiClient is the client API for VendingApi service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// VendingApi describes devices that hold a stock and can dispense their consumables on command.
+// Applies to devices like vending machines, water taps, coffee machines, and pet food dispensers.
+//
+// The VendingApi encodes two different resources, the consumables and the stock.
+// Consumables represent information about the products that are available from this device.
+// Stock is a measure of how many/much of a consumables is currently available.
+// When referring to the stock of all consumables, this trait uses the term inventory.
+//
+// Quantities, both available and dispensed, can be either discrete items like apples, or continuous like water.
+// Some consumables are both discrete and continuous, you can have 0.3L or 1 cup of water.
 type VendingApiClient interface {
 	// List the consumables available from this device.
 	// This has no information on stock, only descriptive information about the physical item.
 	ListConsumables(ctx context.Context, in *ListConsumablesRequest, opts ...grpc.CallOption) (*ListConsumablesResponse, error)
 	// Subscribe to change in consumables available from this device.
-	PullConsumables(ctx context.Context, in *PullConsumablesRequest, opts ...grpc.CallOption) (VendingApi_PullConsumablesClient, error)
+	PullConsumables(ctx context.Context, in *PullConsumablesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullConsumablesResponse], error)
 	// Get the current stock levels for a consumable.
 	// If the consumable is unknown will return NotFound.
 	// If the consumable is out of stock, returns a Stock with a zero remaining quantity.
@@ -50,12 +61,12 @@ type VendingApiClient interface {
 	// that it was refilled with 100 new tasty cookies.
 	UpdateStock(ctx context.Context, in *UpdateStockRequest, opts ...grpc.CallOption) (*Consumable_Stock, error)
 	// Subscribe to changes in the stock levels for a consumable.
-	PullStock(ctx context.Context, in *PullStockRequest, opts ...grpc.CallOption) (VendingApi_PullStockClient, error)
+	PullStock(ctx context.Context, in *PullStockRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullStockResponse], error)
 	// List the stock for each consumable in the device.
 	// The device should return zero-quantity stock for consumables that are out-of-stock.
 	ListInventory(ctx context.Context, in *ListInventoryRequest, opts ...grpc.CallOption) (*ListInventoryResponse, error)
 	// Subscribe to changes in consumable stock.
-	PullInventory(ctx context.Context, in *PullInventoryRequest, opts ...grpc.CallOption) (VendingApi_PullInventoryClient, error)
+	PullInventory(ctx context.Context, in *PullInventoryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullInventoryResponse], error)
 	// Trigger the dispensing of some quantity of consumables.
 	// Devices should return only stock items that are mentioned in the request.
 	Dispense(ctx context.Context, in *DispenseRequest, opts ...grpc.CallOption) (*Consumable_Stock, error)
@@ -73,20 +84,22 @@ func NewVendingApiClient(cc grpc.ClientConnInterface) VendingApiClient {
 }
 
 func (c *vendingApiClient) ListConsumables(ctx context.Context, in *ListConsumablesRequest, opts ...grpc.CallOption) (*ListConsumablesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListConsumablesResponse)
-	err := c.cc.Invoke(ctx, VendingApi_ListConsumables_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, VendingApi_ListConsumables_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *vendingApiClient) PullConsumables(ctx context.Context, in *PullConsumablesRequest, opts ...grpc.CallOption) (VendingApi_PullConsumablesClient, error) {
-	stream, err := c.cc.NewStream(ctx, &VendingApi_ServiceDesc.Streams[0], VendingApi_PullConsumables_FullMethodName, opts...)
+func (c *vendingApiClient) PullConsumables(ctx context.Context, in *PullConsumablesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullConsumablesResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &VendingApi_ServiceDesc.Streams[0], VendingApi_PullConsumables_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &vendingApiPullConsumablesClient{stream}
+	x := &grpc.GenericClientStream[PullConsumablesRequest, PullConsumablesResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -96,26 +109,13 @@ func (c *vendingApiClient) PullConsumables(ctx context.Context, in *PullConsumab
 	return x, nil
 }
 
-type VendingApi_PullConsumablesClient interface {
-	Recv() (*PullConsumablesResponse, error)
-	grpc.ClientStream
-}
-
-type vendingApiPullConsumablesClient struct {
-	grpc.ClientStream
-}
-
-func (x *vendingApiPullConsumablesClient) Recv() (*PullConsumablesResponse, error) {
-	m := new(PullConsumablesResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type VendingApi_PullConsumablesClient = grpc.ServerStreamingClient[PullConsumablesResponse]
 
 func (c *vendingApiClient) GetStock(ctx context.Context, in *GetStockRequest, opts ...grpc.CallOption) (*Consumable_Stock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Consumable_Stock)
-	err := c.cc.Invoke(ctx, VendingApi_GetStock_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, VendingApi_GetStock_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -123,20 +123,22 @@ func (c *vendingApiClient) GetStock(ctx context.Context, in *GetStockRequest, op
 }
 
 func (c *vendingApiClient) UpdateStock(ctx context.Context, in *UpdateStockRequest, opts ...grpc.CallOption) (*Consumable_Stock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Consumable_Stock)
-	err := c.cc.Invoke(ctx, VendingApi_UpdateStock_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, VendingApi_UpdateStock_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *vendingApiClient) PullStock(ctx context.Context, in *PullStockRequest, opts ...grpc.CallOption) (VendingApi_PullStockClient, error) {
-	stream, err := c.cc.NewStream(ctx, &VendingApi_ServiceDesc.Streams[1], VendingApi_PullStock_FullMethodName, opts...)
+func (c *vendingApiClient) PullStock(ctx context.Context, in *PullStockRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullStockResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &VendingApi_ServiceDesc.Streams[1], VendingApi_PullStock_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &vendingApiPullStockClient{stream}
+	x := &grpc.GenericClientStream[PullStockRequest, PullStockResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -146,38 +148,26 @@ func (c *vendingApiClient) PullStock(ctx context.Context, in *PullStockRequest, 
 	return x, nil
 }
 
-type VendingApi_PullStockClient interface {
-	Recv() (*PullStockResponse, error)
-	grpc.ClientStream
-}
-
-type vendingApiPullStockClient struct {
-	grpc.ClientStream
-}
-
-func (x *vendingApiPullStockClient) Recv() (*PullStockResponse, error) {
-	m := new(PullStockResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type VendingApi_PullStockClient = grpc.ServerStreamingClient[PullStockResponse]
 
 func (c *vendingApiClient) ListInventory(ctx context.Context, in *ListInventoryRequest, opts ...grpc.CallOption) (*ListInventoryResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListInventoryResponse)
-	err := c.cc.Invoke(ctx, VendingApi_ListInventory_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, VendingApi_ListInventory_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *vendingApiClient) PullInventory(ctx context.Context, in *PullInventoryRequest, opts ...grpc.CallOption) (VendingApi_PullInventoryClient, error) {
-	stream, err := c.cc.NewStream(ctx, &VendingApi_ServiceDesc.Streams[2], VendingApi_PullInventory_FullMethodName, opts...)
+func (c *vendingApiClient) PullInventory(ctx context.Context, in *PullInventoryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[PullInventoryResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &VendingApi_ServiceDesc.Streams[2], VendingApi_PullInventory_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &vendingApiPullInventoryClient{stream}
+	x := &grpc.GenericClientStream[PullInventoryRequest, PullInventoryResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -187,26 +177,13 @@ func (c *vendingApiClient) PullInventory(ctx context.Context, in *PullInventoryR
 	return x, nil
 }
 
-type VendingApi_PullInventoryClient interface {
-	Recv() (*PullInventoryResponse, error)
-	grpc.ClientStream
-}
-
-type vendingApiPullInventoryClient struct {
-	grpc.ClientStream
-}
-
-func (x *vendingApiPullInventoryClient) Recv() (*PullInventoryResponse, error) {
-	m := new(PullInventoryResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type VendingApi_PullInventoryClient = grpc.ServerStreamingClient[PullInventoryResponse]
 
 func (c *vendingApiClient) Dispense(ctx context.Context, in *DispenseRequest, opts ...grpc.CallOption) (*Consumable_Stock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Consumable_Stock)
-	err := c.cc.Invoke(ctx, VendingApi_Dispense_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, VendingApi_Dispense_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -214,8 +191,9 @@ func (c *vendingApiClient) Dispense(ctx context.Context, in *DispenseRequest, op
 }
 
 func (c *vendingApiClient) StopDispense(ctx context.Context, in *StopDispenseRequest, opts ...grpc.CallOption) (*Consumable_Stock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Consumable_Stock)
-	err := c.cc.Invoke(ctx, VendingApi_StopDispense_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, VendingApi_StopDispense_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -224,13 +202,24 @@ func (c *vendingApiClient) StopDispense(ctx context.Context, in *StopDispenseReq
 
 // VendingApiServer is the server API for VendingApi service.
 // All implementations must embed UnimplementedVendingApiServer
-// for forward compatibility
+// for forward compatibility.
+//
+// VendingApi describes devices that hold a stock and can dispense their consumables on command.
+// Applies to devices like vending machines, water taps, coffee machines, and pet food dispensers.
+//
+// The VendingApi encodes two different resources, the consumables and the stock.
+// Consumables represent information about the products that are available from this device.
+// Stock is a measure of how many/much of a consumables is currently available.
+// When referring to the stock of all consumables, this trait uses the term inventory.
+//
+// Quantities, both available and dispensed, can be either discrete items like apples, or continuous like water.
+// Some consumables are both discrete and continuous, you can have 0.3L or 1 cup of water.
 type VendingApiServer interface {
 	// List the consumables available from this device.
 	// This has no information on stock, only descriptive information about the physical item.
 	ListConsumables(context.Context, *ListConsumablesRequest) (*ListConsumablesResponse, error)
 	// Subscribe to change in consumables available from this device.
-	PullConsumables(*PullConsumablesRequest, VendingApi_PullConsumablesServer) error
+	PullConsumables(*PullConsumablesRequest, grpc.ServerStreamingServer[PullConsumablesResponse]) error
 	// Get the current stock levels for a consumable.
 	// If the consumable is unknown will return NotFound.
 	// If the consumable is out of stock, returns a Stock with a zero remaining quantity.
@@ -242,12 +231,12 @@ type VendingApiServer interface {
 	// that it was refilled with 100 new tasty cookies.
 	UpdateStock(context.Context, *UpdateStockRequest) (*Consumable_Stock, error)
 	// Subscribe to changes in the stock levels for a consumable.
-	PullStock(*PullStockRequest, VendingApi_PullStockServer) error
+	PullStock(*PullStockRequest, grpc.ServerStreamingServer[PullStockResponse]) error
 	// List the stock for each consumable in the device.
 	// The device should return zero-quantity stock for consumables that are out-of-stock.
 	ListInventory(context.Context, *ListInventoryRequest) (*ListInventoryResponse, error)
 	// Subscribe to changes in consumable stock.
-	PullInventory(*PullInventoryRequest, VendingApi_PullInventoryServer) error
+	PullInventory(*PullInventoryRequest, grpc.ServerStreamingServer[PullInventoryResponse]) error
 	// Trigger the dispensing of some quantity of consumables.
 	// Devices should return only stock items that are mentioned in the request.
 	Dispense(context.Context, *DispenseRequest) (*Consumable_Stock, error)
@@ -257,14 +246,17 @@ type VendingApiServer interface {
 	mustEmbedUnimplementedVendingApiServer()
 }
 
-// UnimplementedVendingApiServer must be embedded to have forward compatible implementations.
-type UnimplementedVendingApiServer struct {
-}
+// UnimplementedVendingApiServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedVendingApiServer struct{}
 
 func (UnimplementedVendingApiServer) ListConsumables(context.Context, *ListConsumablesRequest) (*ListConsumablesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListConsumables not implemented")
 }
-func (UnimplementedVendingApiServer) PullConsumables(*PullConsumablesRequest, VendingApi_PullConsumablesServer) error {
+func (UnimplementedVendingApiServer) PullConsumables(*PullConsumablesRequest, grpc.ServerStreamingServer[PullConsumablesResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullConsumables not implemented")
 }
 func (UnimplementedVendingApiServer) GetStock(context.Context, *GetStockRequest) (*Consumable_Stock, error) {
@@ -273,13 +265,13 @@ func (UnimplementedVendingApiServer) GetStock(context.Context, *GetStockRequest)
 func (UnimplementedVendingApiServer) UpdateStock(context.Context, *UpdateStockRequest) (*Consumable_Stock, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateStock not implemented")
 }
-func (UnimplementedVendingApiServer) PullStock(*PullStockRequest, VendingApi_PullStockServer) error {
+func (UnimplementedVendingApiServer) PullStock(*PullStockRequest, grpc.ServerStreamingServer[PullStockResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullStock not implemented")
 }
 func (UnimplementedVendingApiServer) ListInventory(context.Context, *ListInventoryRequest) (*ListInventoryResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListInventory not implemented")
 }
-func (UnimplementedVendingApiServer) PullInventory(*PullInventoryRequest, VendingApi_PullInventoryServer) error {
+func (UnimplementedVendingApiServer) PullInventory(*PullInventoryRequest, grpc.ServerStreamingServer[PullInventoryResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method PullInventory not implemented")
 }
 func (UnimplementedVendingApiServer) Dispense(context.Context, *DispenseRequest) (*Consumable_Stock, error) {
@@ -289,6 +281,7 @@ func (UnimplementedVendingApiServer) StopDispense(context.Context, *StopDispense
 	return nil, status.Errorf(codes.Unimplemented, "method StopDispense not implemented")
 }
 func (UnimplementedVendingApiServer) mustEmbedUnimplementedVendingApiServer() {}
+func (UnimplementedVendingApiServer) testEmbeddedByValue()                    {}
 
 // UnsafeVendingApiServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to VendingApiServer will
@@ -298,6 +291,13 @@ type UnsafeVendingApiServer interface {
 }
 
 func RegisterVendingApiServer(s grpc.ServiceRegistrar, srv VendingApiServer) {
+	// If the following call pancis, it indicates UnimplementedVendingApiServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&VendingApi_ServiceDesc, srv)
 }
 
@@ -324,21 +324,11 @@ func _VendingApi_PullConsumables_Handler(srv interface{}, stream grpc.ServerStre
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(VendingApiServer).PullConsumables(m, &vendingApiPullConsumablesServer{stream})
+	return srv.(VendingApiServer).PullConsumables(m, &grpc.GenericServerStream[PullConsumablesRequest, PullConsumablesResponse]{ServerStream: stream})
 }
 
-type VendingApi_PullConsumablesServer interface {
-	Send(*PullConsumablesResponse) error
-	grpc.ServerStream
-}
-
-type vendingApiPullConsumablesServer struct {
-	grpc.ServerStream
-}
-
-func (x *vendingApiPullConsumablesServer) Send(m *PullConsumablesResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type VendingApi_PullConsumablesServer = grpc.ServerStreamingServer[PullConsumablesResponse]
 
 func _VendingApi_GetStock_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(GetStockRequest)
@@ -381,21 +371,11 @@ func _VendingApi_PullStock_Handler(srv interface{}, stream grpc.ServerStream) er
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(VendingApiServer).PullStock(m, &vendingApiPullStockServer{stream})
+	return srv.(VendingApiServer).PullStock(m, &grpc.GenericServerStream[PullStockRequest, PullStockResponse]{ServerStream: stream})
 }
 
-type VendingApi_PullStockServer interface {
-	Send(*PullStockResponse) error
-	grpc.ServerStream
-}
-
-type vendingApiPullStockServer struct {
-	grpc.ServerStream
-}
-
-func (x *vendingApiPullStockServer) Send(m *PullStockResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type VendingApi_PullStockServer = grpc.ServerStreamingServer[PullStockResponse]
 
 func _VendingApi_ListInventory_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ListInventoryRequest)
@@ -420,21 +400,11 @@ func _VendingApi_PullInventory_Handler(srv interface{}, stream grpc.ServerStream
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(VendingApiServer).PullInventory(m, &vendingApiPullInventoryServer{stream})
+	return srv.(VendingApiServer).PullInventory(m, &grpc.GenericServerStream[PullInventoryRequest, PullInventoryResponse]{ServerStream: stream})
 }
 
-type VendingApi_PullInventoryServer interface {
-	Send(*PullInventoryResponse) error
-	grpc.ServerStream
-}
-
-type vendingApiPullInventoryServer struct {
-	grpc.ServerStream
-}
-
-func (x *vendingApiPullInventoryServer) Send(m *PullInventoryResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type VendingApi_PullInventoryServer = grpc.ServerStreamingServer[PullInventoryResponse]
 
 func _VendingApi_Dispense_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(DispenseRequest)
@@ -524,8 +494,6 @@ var VendingApi_ServiceDesc = grpc.ServiceDesc{
 	Metadata: "traits/vending.proto",
 }
 
-const ()
-
 // VendingInfoClient is the client API for VendingInfo service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
@@ -542,16 +510,20 @@ func NewVendingInfoClient(cc grpc.ClientConnInterface) VendingInfoClient {
 
 // VendingInfoServer is the server API for VendingInfo service.
 // All implementations must embed UnimplementedVendingInfoServer
-// for forward compatibility
+// for forward compatibility.
 type VendingInfoServer interface {
 	mustEmbedUnimplementedVendingInfoServer()
 }
 
-// UnimplementedVendingInfoServer must be embedded to have forward compatible implementations.
-type UnimplementedVendingInfoServer struct {
-}
+// UnimplementedVendingInfoServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedVendingInfoServer struct{}
 
 func (UnimplementedVendingInfoServer) mustEmbedUnimplementedVendingInfoServer() {}
+func (UnimplementedVendingInfoServer) testEmbeddedByValue()                     {}
 
 // UnsafeVendingInfoServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to VendingInfoServer will
@@ -561,6 +533,13 @@ type UnsafeVendingInfoServer interface {
 }
 
 func RegisterVendingInfoServer(s grpc.ServiceRegistrar, srv VendingInfoServer) {
+	// If the following call pancis, it indicates UnimplementedVendingInfoServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&VendingInfo_ServiceDesc, srv)
 }
 
